add tests for mssql, mysql, oracle + refactor some code
