+++ ddl_spark_converter/db_from_spark_datatype_conversion_info/postgresql.py
@@ -1,0 +1,3 @@
+datatypes = {
+    
+}
