+++ ddl_spark_converter/db_to_spark_datatype_conversion_info/postgresql.py
@@ -1,0 +1,1604 @@
+datatypes = {
+    "DECIMAL": "DecimalType(38, 18)",
+    "DECIMAL(0)": "DecimalType(0, 0)",
+    "DECIMAL(1)": "DecimalType(1, 0)",
+    "DECIMAL(2)": "DecimalType(2, 0)",
+    "DECIMAL(3)": "DecimalType(3, 0)",
+    "DECIMAL(4)": "DecimalType(4, 0)",
+    "DECIMAL(5)": "DecimalType(5, 0)",
+    "DECIMAL(6)": "DecimalType(6, 0)",
+    "DECIMAL(7)": "DecimalType(7, 0)",
+    "DECIMAL(8)": "DecimalType(8, 0)",
+    "DECIMAL(9)": "DecimalType(9, 0)",
+    "DECIMAL(10)": "DecimalType(10, 0)",
+    "DECIMAL(11)": "DecimalType(11, 0)",
+    "DECIMAL(12)": "DecimalType(12, 0)",
+    "DECIMAL(13)": "DecimalType(13, 0)",
+    "DECIMAL(14)": "DecimalType(14, 0)",
+    "DECIMAL(15)": "DecimalType(15, 0)",
+    "DECIMAL(16)": "DecimalType(16, 0)",
+    "DECIMAL(17)": "DecimalType(17, 0)",
+    "DECIMAL(18)": "DecimalType(18, 0)",
+    "DECIMAL(19)": "DecimalType(19, 0)",
+    "DECIMAL(20)": "DecimalType(20, 0)",
+    "DECIMAL(21)": "DecimalType(21, 0)",
+    "DECIMAL(22)": "DecimalType(22, 0)",
+    "DECIMAL(23)": "DecimalType(23, 0)",
+    "DECIMAL(24)": "DecimalType(24, 0)",
+    "DECIMAL(25)": "DecimalType(25, 0)",
+    "DECIMAL(26)": "DecimalType(26, 0)",
+    "DECIMAL(27)": "DecimalType(27, 0)",
+    "DECIMAL(28)": "DecimalType(28, 0)",
+    "DECIMAL(29)": "DecimalType(29, 0)",
+    "DECIMAL(30)": "DecimalType(30, 0)",
+    "DECIMAL(31)": "DecimalType(31, 0)",
+    "DECIMAL(32)": "DecimalType(32, 0)",
+    "DECIMAL(33)": "DecimalType(33, 0)",
+    "DECIMAL(34)": "DecimalType(34, 0)",
+    "DECIMAL(35)": "DecimalType(35, 0)",
+    "DECIMAL(36)": "DecimalType(36, 0)",
+    "DECIMAL(37)": "DecimalType(37, 0)",
+    "DECIMAL(38)": "DecimalType(38, 0)",
+    "DECIMAL(0, 0)": "DecimalType(0, 0)",
+    "DECIMAL(0, 1)": "DecimalType(0, 1)",
+    "DECIMAL(0, 2)": "DecimalType(0, 2)",
+    "DECIMAL(0, 3)": "DecimalType(0, 3)",
+    "DECIMAL(0, 4)": "DecimalType(0, 4)",
+    "DECIMAL(0, 5)": "DecimalType(0, 5)",
+    "DECIMAL(0, 6)": "DecimalType(0, 6)",
+    "DECIMAL(0, 7)": "DecimalType(0, 7)",
+    "DECIMAL(0, 8)": "DecimalType(0, 8)",
+    "DECIMAL(0, 9)": "DecimalType(0, 9)",
+    "DECIMAL(0, 10)": "DecimalType(0, 10)",
+    "DECIMAL(0, 11)": "DecimalType(0, 11)",
+    "DECIMAL(0, 12)": "DecimalType(0, 12)",
+    "DECIMAL(0, 13)": "DecimalType(0, 13)",
+    "DECIMAL(0, 14)": "DecimalType(0, 14)",
+    "DECIMAL(0, 15)": "DecimalType(0, 15)",
+    "DECIMAL(0, 16)": "DecimalType(0, 16)",
+    "DECIMAL(0, 17)": "DecimalType(0, 17)",
+    "DECIMAL(0, 18)": "DecimalType(0, 18)",
+    "DECIMAL(0, 19)": "DecimalType(0, 19)",
+    "DECIMAL(0, 20)": "DecimalType(0, 20)",
+    "DECIMAL(0, 21)": "DecimalType(0, 21)",
+    "DECIMAL(0, 22)": "DecimalType(0, 22)",
+    "DECIMAL(0, 23)": "DecimalType(0, 23)",
+    "DECIMAL(0, 24)": "DecimalType(0, 24)",
+    "DECIMAL(0, 25)": "DecimalType(0, 25)",
+    "DECIMAL(0, 26)": "DecimalType(0, 26)",
+    "DECIMAL(0, 27)": "DecimalType(0, 27)",
+    "DECIMAL(0, 28)": "DecimalType(0, 28)",
+    "DECIMAL(0, 29)": "DecimalType(0, 29)",
+    "DECIMAL(0, 30)": "DecimalType(0, 30)",
+    "DECIMAL(0, 31)": "DecimalType(0, 31)",
+    "DECIMAL(0, 32)": "DecimalType(0, 32)",
+    "DECIMAL(0, 33)": "DecimalType(0, 33)",
+    "DECIMAL(0, 34)": "DecimalType(0, 34)",
+    "DECIMAL(0, 35)": "DecimalType(0, 35)",
+    "DECIMAL(0, 36)": "DecimalType(0, 36)",
+    "DECIMAL(0, 37)": "DecimalType(0, 37)",
+    "DECIMAL(0, 38)": "DecimalType(0, 38)",
+    "DECIMAL(1, 0)": "DecimalType(1, 0)",
+    "DECIMAL(1, 1)": "DecimalType(1, 1)",
+    "DECIMAL(1, 2)": "DecimalType(1, 2)",
+    "DECIMAL(1, 3)": "DecimalType(1, 3)",
+    "DECIMAL(1, 4)": "DecimalType(1, 4)",
+    "DECIMAL(1, 5)": "DecimalType(1, 5)",
+    "DECIMAL(1, 6)": "DecimalType(1, 6)",
+    "DECIMAL(1, 7)": "DecimalType(1, 7)",
+    "DECIMAL(1, 8)": "DecimalType(1, 8)",
+    "DECIMAL(1, 9)": "DecimalType(1, 9)",
+    "DECIMAL(1, 10)": "DecimalType(1, 10)",
+    "DECIMAL(1, 11)": "DecimalType(1, 11)",
+    "DECIMAL(1, 12)": "DecimalType(1, 12)",
+    "DECIMAL(1, 13)": "DecimalType(1, 13)",
+    "DECIMAL(1, 14)": "DecimalType(1, 14)",
+    "DECIMAL(1, 15)": "DecimalType(1, 15)",
+    "DECIMAL(1, 16)": "DecimalType(1, 16)",
+    "DECIMAL(1, 17)": "DecimalType(1, 17)",
+    "DECIMAL(1, 18)": "DecimalType(1, 18)",
+    "DECIMAL(1, 19)": "DecimalType(1, 19)",
+    "DECIMAL(1, 20)": "DecimalType(1, 20)",
+    "DECIMAL(1, 21)": "DecimalType(1, 21)",
+    "DECIMAL(1, 22)": "DecimalType(1, 22)",
+    "DECIMAL(1, 23)": "DecimalType(1, 23)",
+    "DECIMAL(1, 24)": "DecimalType(1, 24)",
+    "DECIMAL(1, 25)": "DecimalType(1, 25)",
+    "DECIMAL(1, 26)": "DecimalType(1, 26)",
+    "DECIMAL(1, 27)": "DecimalType(1, 27)",
+    "DECIMAL(1, 28)": "DecimalType(1, 28)",
+    "DECIMAL(1, 29)": "DecimalType(1, 29)",
+    "DECIMAL(1, 30)": "DecimalType(1, 30)",
+    "DECIMAL(1, 31)": "DecimalType(1, 31)",
+    "DECIMAL(1, 32)": "DecimalType(1, 32)",
+    "DECIMAL(1, 33)": "DecimalType(1, 33)",
+    "DECIMAL(1, 34)": "DecimalType(1, 34)",
+    "DECIMAL(1, 35)": "DecimalType(1, 35)",
+    "DECIMAL(1, 36)": "DecimalType(1, 36)",
+    "DECIMAL(1, 37)": "DecimalType(1, 37)",
+    "DECIMAL(1, 38)": "DecimalType(1, 38)",
+    "DECIMAL(2, 0)": "DecimalType(2, 0)",
+    "DECIMAL(2, 1)": "DecimalType(2, 1)",
+    "DECIMAL(2, 2)": "DecimalType(2, 2)",
+    "DECIMAL(2, 3)": "DecimalType(2, 3)",
+    "DECIMAL(2, 4)": "DecimalType(2, 4)",
+    "DECIMAL(2, 5)": "DecimalType(2, 5)",
+    "DECIMAL(2, 6)": "DecimalType(2, 6)",
+    "DECIMAL(2, 7)": "DecimalType(2, 7)",
+    "DECIMAL(2, 8)": "DecimalType(2, 8)",
+    "DECIMAL(2, 9)": "DecimalType(2, 9)",
+    "DECIMAL(2, 10)": "DecimalType(2, 10)",
+    "DECIMAL(2, 11)": "DecimalType(2, 11)",
+    "DECIMAL(2, 12)": "DecimalType(2, 12)",
+    "DECIMAL(2, 13)": "DecimalType(2, 13)",
+    "DECIMAL(2, 14)": "DecimalType(2, 14)",
+    "DECIMAL(2, 15)": "DecimalType(2, 15)",
+    "DECIMAL(2, 16)": "DecimalType(2, 16)",
+    "DECIMAL(2, 17)": "DecimalType(2, 17)",
+    "DECIMAL(2, 18)": "DecimalType(2, 18)",
+    "DECIMAL(2, 19)": "DecimalType(2, 19)",
+    "DECIMAL(2, 20)": "DecimalType(2, 20)",
+    "DECIMAL(2, 21)": "DecimalType(2, 21)",
+    "DECIMAL(2, 22)": "DecimalType(2, 22)",
+    "DECIMAL(2, 23)": "DecimalType(2, 23)",
+    "DECIMAL(2, 24)": "DecimalType(2, 24)",
+    "DECIMAL(2, 25)": "DecimalType(2, 25)",
+    "DECIMAL(2, 26)": "DecimalType(2, 26)",
+    "DECIMAL(2, 27)": "DecimalType(2, 27)",
+    "DECIMAL(2, 28)": "DecimalType(2, 28)",
+    "DECIMAL(2, 29)": "DecimalType(2, 29)",
+    "DECIMAL(2, 30)": "DecimalType(2, 30)",
+    "DECIMAL(2, 31)": "DecimalType(2, 31)",
+    "DECIMAL(2, 32)": "DecimalType(2, 32)",
+    "DECIMAL(2, 33)": "DecimalType(2, 33)",
+    "DECIMAL(2, 34)": "DecimalType(2, 34)",
+    "DECIMAL(2, 35)": "DecimalType(2, 35)",
+    "DECIMAL(2, 36)": "DecimalType(2, 36)",
+    "DECIMAL(2, 37)": "DecimalType(2, 37)",
+    "DECIMAL(2, 38)": "DecimalType(2, 38)",
+    "DECIMAL(3, 0)": "DecimalType(3, 0)",
+    "DECIMAL(3, 1)": "DecimalType(3, 1)",
+    "DECIMAL(3, 2)": "DecimalType(3, 2)",
+    "DECIMAL(3, 3)": "DecimalType(3, 3)",
+    "DECIMAL(3, 4)": "DecimalType(3, 4)",
+    "DECIMAL(3, 5)": "DecimalType(3, 5)",
+    "DECIMAL(3, 6)": "DecimalType(3, 6)",
+    "DECIMAL(3, 7)": "DecimalType(3, 7)",
+    "DECIMAL(3, 8)": "DecimalType(3, 8)",
+    "DECIMAL(3, 9)": "DecimalType(3, 9)",
+    "DECIMAL(3, 10)": "DecimalType(3, 10)",
+    "DECIMAL(3, 11)": "DecimalType(3, 11)",
+    "DECIMAL(3, 12)": "DecimalType(3, 12)",
+    "DECIMAL(3, 13)": "DecimalType(3, 13)",
+    "DECIMAL(3, 14)": "DecimalType(3, 14)",
+    "DECIMAL(3, 15)": "DecimalType(3, 15)",
+    "DECIMAL(3, 16)": "DecimalType(3, 16)",
+    "DECIMAL(3, 17)": "DecimalType(3, 17)",
+    "DECIMAL(3, 18)": "DecimalType(3, 18)",
+    "DECIMAL(3, 19)": "DecimalType(3, 19)",
+    "DECIMAL(3, 20)": "DecimalType(3, 20)",
+    "DECIMAL(3, 21)": "DecimalType(3, 21)",
+    "DECIMAL(3, 22)": "DecimalType(3, 22)",
+    "DECIMAL(3, 23)": "DecimalType(3, 23)",
+    "DECIMAL(3, 24)": "DecimalType(3, 24)",
+    "DECIMAL(3, 25)": "DecimalType(3, 25)",
+    "DECIMAL(3, 26)": "DecimalType(3, 26)",
+    "DECIMAL(3, 27)": "DecimalType(3, 27)",
+    "DECIMAL(3, 28)": "DecimalType(3, 28)",
+    "DECIMAL(3, 29)": "DecimalType(3, 29)",
+    "DECIMAL(3, 30)": "DecimalType(3, 30)",
+    "DECIMAL(3, 31)": "DecimalType(3, 31)",
+    "DECIMAL(3, 32)": "DecimalType(3, 32)",
+    "DECIMAL(3, 33)": "DecimalType(3, 33)",
+    "DECIMAL(3, 34)": "DecimalType(3, 34)",
+    "DECIMAL(3, 35)": "DecimalType(3, 35)",
+    "DECIMAL(3, 36)": "DecimalType(3, 36)",
+    "DECIMAL(3, 37)": "DecimalType(3, 37)",
+    "DECIMAL(3, 38)": "DecimalType(3, 38)",
+    "DECIMAL(4, 0)": "DecimalType(4, 0)",
+    "DECIMAL(4, 1)": "DecimalType(4, 1)",
+    "DECIMAL(4, 2)": "DecimalType(4, 2)",
+    "DECIMAL(4, 3)": "DecimalType(4, 3)",
+    "DECIMAL(4, 4)": "DecimalType(4, 4)",
+    "DECIMAL(4, 5)": "DecimalType(4, 5)",
+    "DECIMAL(4, 6)": "DecimalType(4, 6)",
+    "DECIMAL(4, 7)": "DecimalType(4, 7)",
+    "DECIMAL(4, 8)": "DecimalType(4, 8)",
+    "DECIMAL(4, 9)": "DecimalType(4, 9)",
+    "DECIMAL(4, 10)": "DecimalType(4, 10)",
+    "DECIMAL(4, 11)": "DecimalType(4, 11)",
+    "DECIMAL(4, 12)": "DecimalType(4, 12)",
+    "DECIMAL(4, 13)": "DecimalType(4, 13)",
+    "DECIMAL(4, 14)": "DecimalType(4, 14)",
+    "DECIMAL(4, 15)": "DecimalType(4, 15)",
+    "DECIMAL(4, 16)": "DecimalType(4, 16)",
+    "DECIMAL(4, 17)": "DecimalType(4, 17)",
+    "DECIMAL(4, 18)": "DecimalType(4, 18)",
+    "DECIMAL(4, 19)": "DecimalType(4, 19)",
+    "DECIMAL(4, 20)": "DecimalType(4, 20)",
+    "DECIMAL(4, 21)": "DecimalType(4, 21)",
+    "DECIMAL(4, 22)": "DecimalType(4, 22)",
+    "DECIMAL(4, 23)": "DecimalType(4, 23)",
+    "DECIMAL(4, 24)": "DecimalType(4, 24)",
+    "DECIMAL(4, 25)": "DecimalType(4, 25)",
+    "DECIMAL(4, 26)": "DecimalType(4, 26)",
+    "DECIMAL(4, 27)": "DecimalType(4, 27)",
+    "DECIMAL(4, 28)": "DecimalType(4, 28)",
+    "DECIMAL(4, 29)": "DecimalType(4, 29)",
+    "DECIMAL(4, 30)": "DecimalType(4, 30)",
+    "DECIMAL(4, 31)": "DecimalType(4, 31)",
+    "DECIMAL(4, 32)": "DecimalType(4, 32)",
+    "DECIMAL(4, 33)": "DecimalType(4, 33)",
+    "DECIMAL(4, 34)": "DecimalType(4, 34)",
+    "DECIMAL(4, 35)": "DecimalType(4, 35)",
+    "DECIMAL(4, 36)": "DecimalType(4, 36)",
+    "DECIMAL(4, 37)": "DecimalType(4, 37)",
+    "DECIMAL(4, 38)": "DecimalType(4, 38)",
+    "DECIMAL(5, 0)": "DecimalType(5, 0)",
+    "DECIMAL(5, 1)": "DecimalType(5, 1)",
+    "DECIMAL(5, 2)": "DecimalType(5, 2)",
+    "DECIMAL(5, 3)": "DecimalType(5, 3)",
+    "DECIMAL(5, 4)": "DecimalType(5, 4)",
+    "DECIMAL(5, 5)": "DecimalType(5, 5)",
+    "DECIMAL(5, 6)": "DecimalType(5, 6)",
+    "DECIMAL(5, 7)": "DecimalType(5, 7)",
+    "DECIMAL(5, 8)": "DecimalType(5, 8)",
+    "DECIMAL(5, 9)": "DecimalType(5, 9)",
+    "DECIMAL(5, 10)": "DecimalType(5, 10)",
+    "DECIMAL(5, 11)": "DecimalType(5, 11)",
+    "DECIMAL(5, 12)": "DecimalType(5, 12)",
+    "DECIMAL(5, 13)": "DecimalType(5, 13)",
+    "DECIMAL(5, 14)": "DecimalType(5, 14)",
+    "DECIMAL(5, 15)": "DecimalType(5, 15)",
+    "DECIMAL(5, 16)": "DecimalType(5, 16)",
+    "DECIMAL(5, 17)": "DecimalType(5, 17)",
+    "DECIMAL(5, 18)": "DecimalType(5, 18)",
+    "DECIMAL(5, 19)": "DecimalType(5, 19)",
+    "DECIMAL(5, 20)": "DecimalType(5, 20)",
+    "DECIMAL(5, 21)": "DecimalType(5, 21)",
+    "DECIMAL(5, 22)": "DecimalType(5, 22)",
+    "DECIMAL(5, 23)": "DecimalType(5, 23)",
+    "DECIMAL(5, 24)": "DecimalType(5, 24)",
+    "DECIMAL(5, 25)": "DecimalType(5, 25)",
+    "DECIMAL(5, 26)": "DecimalType(5, 26)",
+    "DECIMAL(5, 27)": "DecimalType(5, 27)",
+    "DECIMAL(5, 28)": "DecimalType(5, 28)",
+    "DECIMAL(5, 29)": "DecimalType(5, 29)",
+    "DECIMAL(5, 30)": "DecimalType(5, 30)",
+    "DECIMAL(5, 31)": "DecimalType(5, 31)",
+    "DECIMAL(5, 32)": "DecimalType(5, 32)",
+    "DECIMAL(5, 33)": "DecimalType(5, 33)",
+    "DECIMAL(5, 34)": "DecimalType(5, 34)",
+    "DECIMAL(5, 35)": "DecimalType(5, 35)",
+    "DECIMAL(5, 36)": "DecimalType(5, 36)",
+    "DECIMAL(5, 37)": "DecimalType(5, 37)",
+    "DECIMAL(5, 38)": "DecimalType(5, 38)",
+    "DECIMAL(6, 0)": "DecimalType(6, 0)",
+    "DECIMAL(6, 1)": "DecimalType(6, 1)",
+    "DECIMAL(6, 2)": "DecimalType(6, 2)",
+    "DECIMAL(6, 3)": "DecimalType(6, 3)",
+    "DECIMAL(6, 4)": "DecimalType(6, 4)",
+    "DECIMAL(6, 5)": "DecimalType(6, 5)",
+    "DECIMAL(6, 6)": "DecimalType(6, 6)",
+    "DECIMAL(6, 7)": "DecimalType(6, 7)",
+    "DECIMAL(6, 8)": "DecimalType(6, 8)",
+    "DECIMAL(6, 9)": "DecimalType(6, 9)",
+    "DECIMAL(6, 10)": "DecimalType(6, 10)",
+    "DECIMAL(6, 11)": "DecimalType(6, 11)",
+    "DECIMAL(6, 12)": "DecimalType(6, 12)",
+    "DECIMAL(6, 13)": "DecimalType(6, 13)",
+    "DECIMAL(6, 14)": "DecimalType(6, 14)",
+    "DECIMAL(6, 15)": "DecimalType(6, 15)",
+    "DECIMAL(6, 16)": "DecimalType(6, 16)",
+    "DECIMAL(6, 17)": "DecimalType(6, 17)",
+    "DECIMAL(6, 18)": "DecimalType(6, 18)",
+    "DECIMAL(6, 19)": "DecimalType(6, 19)",
+    "DECIMAL(6, 20)": "DecimalType(6, 20)",
+    "DECIMAL(6, 21)": "DecimalType(6, 21)",
+    "DECIMAL(6, 22)": "DecimalType(6, 22)",
+    "DECIMAL(6, 23)": "DecimalType(6, 23)",
+    "DECIMAL(6, 24)": "DecimalType(6, 24)",
+    "DECIMAL(6, 25)": "DecimalType(6, 25)",
+    "DECIMAL(6, 26)": "DecimalType(6, 26)",
+    "DECIMAL(6, 27)": "DecimalType(6, 27)",
+    "DECIMAL(6, 28)": "DecimalType(6, 28)",
+    "DECIMAL(6, 29)": "DecimalType(6, 29)",
+    "DECIMAL(6, 30)": "DecimalType(6, 30)",
+    "DECIMAL(6, 31)": "DecimalType(6, 31)",
+    "DECIMAL(6, 32)": "DecimalType(6, 32)",
+    "DECIMAL(6, 33)": "DecimalType(6, 33)",
+    "DECIMAL(6, 34)": "DecimalType(6, 34)",
+    "DECIMAL(6, 35)": "DecimalType(6, 35)",
+    "DECIMAL(6, 36)": "DecimalType(6, 36)",
+    "DECIMAL(6, 37)": "DecimalType(6, 37)",
+    "DECIMAL(6, 38)": "DecimalType(6, 38)",
+    "DECIMAL(7, 0)": "DecimalType(7, 0)",
+    "DECIMAL(7, 1)": "DecimalType(7, 1)",
+    "DECIMAL(7, 2)": "DecimalType(7, 2)",
+    "DECIMAL(7, 3)": "DecimalType(7, 3)",
+    "DECIMAL(7, 4)": "DecimalType(7, 4)",
+    "DECIMAL(7, 5)": "DecimalType(7, 5)",
+    "DECIMAL(7, 6)": "DecimalType(7, 6)",
+    "DECIMAL(7, 7)": "DecimalType(7, 7)",
+    "DECIMAL(7, 8)": "DecimalType(7, 8)",
+    "DECIMAL(7, 9)": "DecimalType(7, 9)",
+    "DECIMAL(7, 10)": "DecimalType(7, 10)",
+    "DECIMAL(7, 11)": "DecimalType(7, 11)",
+    "DECIMAL(7, 12)": "DecimalType(7, 12)",
+    "DECIMAL(7, 13)": "DecimalType(7, 13)",
+    "DECIMAL(7, 14)": "DecimalType(7, 14)",
+    "DECIMAL(7, 15)": "DecimalType(7, 15)",
+    "DECIMAL(7, 16)": "DecimalType(7, 16)",
+    "DECIMAL(7, 17)": "DecimalType(7, 17)",
+    "DECIMAL(7, 18)": "DecimalType(7, 18)",
+    "DECIMAL(7, 19)": "DecimalType(7, 19)",
+    "DECIMAL(7, 20)": "DecimalType(7, 20)",
+    "DECIMAL(7, 21)": "DecimalType(7, 21)",
+    "DECIMAL(7, 22)": "DecimalType(7, 22)",
+    "DECIMAL(7, 23)": "DecimalType(7, 23)",
+    "DECIMAL(7, 24)": "DecimalType(7, 24)",
+    "DECIMAL(7, 25)": "DecimalType(7, 25)",
+    "DECIMAL(7, 26)": "DecimalType(7, 26)",
+    "DECIMAL(7, 27)": "DecimalType(7, 27)",
+    "DECIMAL(7, 28)": "DecimalType(7, 28)",
+    "DECIMAL(7, 29)": "DecimalType(7, 29)",
+    "DECIMAL(7, 30)": "DecimalType(7, 30)",
+    "DECIMAL(7, 31)": "DecimalType(7, 31)",
+    "DECIMAL(7, 32)": "DecimalType(7, 32)",
+    "DECIMAL(7, 33)": "DecimalType(7, 33)",
+    "DECIMAL(7, 34)": "DecimalType(7, 34)",
+    "DECIMAL(7, 35)": "DecimalType(7, 35)",
+    "DECIMAL(7, 36)": "DecimalType(7, 36)",
+    "DECIMAL(7, 37)": "DecimalType(7, 37)",
+    "DECIMAL(7, 38)": "DecimalType(7, 38)",
+    "DECIMAL(8, 0)": "DecimalType(8, 0)",
+    "DECIMAL(8, 1)": "DecimalType(8, 1)",
+    "DECIMAL(8, 2)": "DecimalType(8, 2)",
+    "DECIMAL(8, 3)": "DecimalType(8, 3)",
+    "DECIMAL(8, 4)": "DecimalType(8, 4)",
+    "DECIMAL(8, 5)": "DecimalType(8, 5)",
+    "DECIMAL(8, 6)": "DecimalType(8, 6)",
+    "DECIMAL(8, 7)": "DecimalType(8, 7)",
+    "DECIMAL(8, 8)": "DecimalType(8, 8)",
+    "DECIMAL(8, 9)": "DecimalType(8, 9)",
+    "DECIMAL(8, 10)": "DecimalType(8, 10)",
+    "DECIMAL(8, 11)": "DecimalType(8, 11)",
+    "DECIMAL(8, 12)": "DecimalType(8, 12)",
+    "DECIMAL(8, 13)": "DecimalType(8, 13)",
+    "DECIMAL(8, 14)": "DecimalType(8, 14)",
+    "DECIMAL(8, 15)": "DecimalType(8, 15)",
+    "DECIMAL(8, 16)": "DecimalType(8, 16)",
+    "DECIMAL(8, 17)": "DecimalType(8, 17)",
+    "DECIMAL(8, 18)": "DecimalType(8, 18)",
+    "DECIMAL(8, 19)": "DecimalType(8, 19)",
+    "DECIMAL(8, 20)": "DecimalType(8, 20)",
+    "DECIMAL(8, 21)": "DecimalType(8, 21)",
+    "DECIMAL(8, 22)": "DecimalType(8, 22)",
+    "DECIMAL(8, 23)": "DecimalType(8, 23)",
+    "DECIMAL(8, 24)": "DecimalType(8, 24)",
+    "DECIMAL(8, 25)": "DecimalType(8, 25)",
+    "DECIMAL(8, 26)": "DecimalType(8, 26)",
+    "DECIMAL(8, 27)": "DecimalType(8, 27)",
+    "DECIMAL(8, 28)": "DecimalType(8, 28)",
+    "DECIMAL(8, 29)": "DecimalType(8, 29)",
+    "DECIMAL(8, 30)": "DecimalType(8, 30)",
+    "DECIMAL(8, 31)": "DecimalType(8, 31)",
+    "DECIMAL(8, 32)": "DecimalType(8, 32)",
+    "DECIMAL(8, 33)": "DecimalType(8, 33)",
+    "DECIMAL(8, 34)": "DecimalType(8, 34)",
+    "DECIMAL(8, 35)": "DecimalType(8, 35)",
+    "DECIMAL(8, 36)": "DecimalType(8, 36)",
+    "DECIMAL(8, 37)": "DecimalType(8, 37)",
+    "DECIMAL(8, 38)": "DecimalType(8, 38)",
+    "DECIMAL(9, 0)": "DecimalType(9, 0)",
+    "DECIMAL(9, 1)": "DecimalType(9, 1)",
+    "DECIMAL(9, 2)": "DecimalType(9, 2)",
+    "DECIMAL(9, 3)": "DecimalType(9, 3)",
+    "DECIMAL(9, 4)": "DecimalType(9, 4)",
+    "DECIMAL(9, 5)": "DecimalType(9, 5)",
+    "DECIMAL(9, 6)": "DecimalType(9, 6)",
+    "DECIMAL(9, 7)": "DecimalType(9, 7)",
+    "DECIMAL(9, 8)": "DecimalType(9, 8)",
+    "DECIMAL(9, 9)": "DecimalType(9, 9)",
+    "DECIMAL(9, 10)": "DecimalType(9, 10)",
+    "DECIMAL(9, 11)": "DecimalType(9, 11)",
+    "DECIMAL(9, 12)": "DecimalType(9, 12)",
+    "DECIMAL(9, 13)": "DecimalType(9, 13)",
+    "DECIMAL(9, 14)": "DecimalType(9, 14)",
+    "DECIMAL(9, 15)": "DecimalType(9, 15)",
+    "DECIMAL(9, 16)": "DecimalType(9, 16)",
+    "DECIMAL(9, 17)": "DecimalType(9, 17)",
+    "DECIMAL(9, 18)": "DecimalType(9, 18)",
+    "DECIMAL(9, 19)": "DecimalType(9, 19)",
+    "DECIMAL(9, 20)": "DecimalType(9, 20)",
+    "DECIMAL(9, 21)": "DecimalType(9, 21)",
+    "DECIMAL(9, 22)": "DecimalType(9, 22)",
+    "DECIMAL(9, 23)": "DecimalType(9, 23)",
+    "DECIMAL(9, 24)": "DecimalType(9, 24)",
+    "DECIMAL(9, 25)": "DecimalType(9, 25)",
+    "DECIMAL(9, 26)": "DecimalType(9, 26)",
+    "DECIMAL(9, 27)": "DecimalType(9, 27)",
+    "DECIMAL(9, 28)": "DecimalType(9, 28)",
+    "DECIMAL(9, 29)": "DecimalType(9, 29)",
+    "DECIMAL(9, 30)": "DecimalType(9, 30)",
+    "DECIMAL(9, 31)": "DecimalType(9, 31)",
+    "DECIMAL(9, 32)": "DecimalType(9, 32)",
+    "DECIMAL(9, 33)": "DecimalType(9, 33)",
+    "DECIMAL(9, 34)": "DecimalType(9, 34)",
+    "DECIMAL(9, 35)": "DecimalType(9, 35)",
+    "DECIMAL(9, 36)": "DecimalType(9, 36)",
+    "DECIMAL(9, 37)": "DecimalType(9, 37)",
+    "DECIMAL(9, 38)": "DecimalType(9, 38)",
+    "DECIMAL(10, 0)": "DecimalType(10, 0)",
+    "DECIMAL(10, 1)": "DecimalType(10, 1)",
+    "DECIMAL(10, 2)": "DecimalType(10, 2)",
+    "DECIMAL(10, 3)": "DecimalType(10, 3)",
+    "DECIMAL(10, 4)": "DecimalType(10, 4)",
+    "DECIMAL(10, 5)": "DecimalType(10, 5)",
+    "DECIMAL(10, 6)": "DecimalType(10, 6)",
+    "DECIMAL(10, 7)": "DecimalType(10, 7)",
+    "DECIMAL(10, 8)": "DecimalType(10, 8)",
+    "DECIMAL(10, 9)": "DecimalType(10, 9)",
+    "DECIMAL(10, 10)": "DecimalType(10, 10)",
+    "DECIMAL(10, 11)": "DecimalType(10, 11)",
+    "DECIMAL(10, 12)": "DecimalType(10, 12)",
+    "DECIMAL(10, 13)": "DecimalType(10, 13)",
+    "DECIMAL(10, 14)": "DecimalType(10, 14)",
+    "DECIMAL(10, 15)": "DecimalType(10, 15)",
+    "DECIMAL(10, 16)": "DecimalType(10, 16)",
+    "DECIMAL(10, 17)": "DecimalType(10, 17)",
+    "DECIMAL(10, 18)": "DecimalType(10, 18)",
+    "DECIMAL(10, 19)": "DecimalType(10, 19)",
+    "DECIMAL(10, 20)": "DecimalType(10, 20)",
+    "DECIMAL(10, 21)": "DecimalType(10, 21)",
+    "DECIMAL(10, 22)": "DecimalType(10, 22)",
+    "DECIMAL(10, 23)": "DecimalType(10, 23)",
+    "DECIMAL(10, 24)": "DecimalType(10, 24)",
+    "DECIMAL(10, 25)": "DecimalType(10, 25)",
+    "DECIMAL(10, 26)": "DecimalType(10, 26)",
+    "DECIMAL(10, 27)": "DecimalType(10, 27)",
+    "DECIMAL(10, 28)": "DecimalType(10, 28)",
+    "DECIMAL(10, 29)": "DecimalType(10, 29)",
+    "DECIMAL(10, 30)": "DecimalType(10, 30)",
+    "DECIMAL(10, 31)": "DecimalType(10, 31)",
+    "DECIMAL(10, 32)": "DecimalType(10, 32)",
+    "DECIMAL(10, 33)": "DecimalType(10, 33)",
+    "DECIMAL(10, 34)": "DecimalType(10, 34)",
+    "DECIMAL(10, 35)": "DecimalType(10, 35)",
+    "DECIMAL(10, 36)": "DecimalType(10, 36)",
+    "DECIMAL(10, 37)": "DecimalType(10, 37)",
+    "DECIMAL(10, 38)": "DecimalType(10, 38)",
+    "DECIMAL(11, 0)": "DecimalType(11, 0)",
+    "DECIMAL(11, 1)": "DecimalType(11, 1)",
+    "DECIMAL(11, 2)": "DecimalType(11, 2)",
+    "DECIMAL(11, 3)": "DecimalType(11, 3)",
+    "DECIMAL(11, 4)": "DecimalType(11, 4)",
+    "DECIMAL(11, 5)": "DecimalType(11, 5)",
+    "DECIMAL(11, 6)": "DecimalType(11, 6)",
+    "DECIMAL(11, 7)": "DecimalType(11, 7)",
+    "DECIMAL(11, 8)": "DecimalType(11, 8)",
+    "DECIMAL(11, 9)": "DecimalType(11, 9)",
+    "DECIMAL(11, 10)": "DecimalType(11, 10)",
+    "DECIMAL(11, 11)": "DecimalType(11, 11)",
+    "DECIMAL(11, 12)": "DecimalType(11, 12)",
+    "DECIMAL(11, 13)": "DecimalType(11, 13)",
+    "DECIMAL(11, 14)": "DecimalType(11, 14)",
+    "DECIMAL(11, 15)": "DecimalType(11, 15)",
+    "DECIMAL(11, 16)": "DecimalType(11, 16)",
+    "DECIMAL(11, 17)": "DecimalType(11, 17)",
+    "DECIMAL(11, 18)": "DecimalType(11, 18)",
+    "DECIMAL(11, 19)": "DecimalType(11, 19)",
+    "DECIMAL(11, 20)": "DecimalType(11, 20)",
+    "DECIMAL(11, 21)": "DecimalType(11, 21)",
+    "DECIMAL(11, 22)": "DecimalType(11, 22)",
+    "DECIMAL(11, 23)": "DecimalType(11, 23)",
+    "DECIMAL(11, 24)": "DecimalType(11, 24)",
+    "DECIMAL(11, 25)": "DecimalType(11, 25)",
+    "DECIMAL(11, 26)": "DecimalType(11, 26)",
+    "DECIMAL(11, 27)": "DecimalType(11, 27)",
+    "DECIMAL(11, 28)": "DecimalType(11, 28)",
+    "DECIMAL(11, 29)": "DecimalType(11, 29)",
+    "DECIMAL(11, 30)": "DecimalType(11, 30)",
+    "DECIMAL(11, 31)": "DecimalType(11, 31)",
+    "DECIMAL(11, 32)": "DecimalType(11, 32)",
+    "DECIMAL(11, 33)": "DecimalType(11, 33)",
+    "DECIMAL(11, 34)": "DecimalType(11, 34)",
+    "DECIMAL(11, 35)": "DecimalType(11, 35)",
+    "DECIMAL(11, 36)": "DecimalType(11, 36)",
+    "DECIMAL(11, 37)": "DecimalType(11, 37)",
+    "DECIMAL(11, 38)": "DecimalType(11, 38)",
+    "DECIMAL(12, 0)": "DecimalType(12, 0)",
+    "DECIMAL(12, 1)": "DecimalType(12, 1)",
+    "DECIMAL(12, 2)": "DecimalType(12, 2)",
+    "DECIMAL(12, 3)": "DecimalType(12, 3)",
+    "DECIMAL(12, 4)": "DecimalType(12, 4)",
+    "DECIMAL(12, 5)": "DecimalType(12, 5)",
+    "DECIMAL(12, 6)": "DecimalType(12, 6)",
+    "DECIMAL(12, 7)": "DecimalType(12, 7)",
+    "DECIMAL(12, 8)": "DecimalType(12, 8)",
+    "DECIMAL(12, 9)": "DecimalType(12, 9)",
+    "DECIMAL(12, 10)": "DecimalType(12, 10)",
+    "DECIMAL(12, 11)": "DecimalType(12, 11)",
+    "DECIMAL(12, 12)": "DecimalType(12, 12)",
+    "DECIMAL(12, 13)": "DecimalType(12, 13)",
+    "DECIMAL(12, 14)": "DecimalType(12, 14)",
+    "DECIMAL(12, 15)": "DecimalType(12, 15)",
+    "DECIMAL(12, 16)": "DecimalType(12, 16)",
+    "DECIMAL(12, 17)": "DecimalType(12, 17)",
+    "DECIMAL(12, 18)": "DecimalType(12, 18)",
+    "DECIMAL(12, 19)": "DecimalType(12, 19)",
+    "DECIMAL(12, 20)": "DecimalType(12, 20)",
+    "DECIMAL(12, 21)": "DecimalType(12, 21)",
+    "DECIMAL(12, 22)": "DecimalType(12, 22)",
+    "DECIMAL(12, 23)": "DecimalType(12, 23)",
+    "DECIMAL(12, 24)": "DecimalType(12, 24)",
+    "DECIMAL(12, 25)": "DecimalType(12, 25)",
+    "DECIMAL(12, 26)": "DecimalType(12, 26)",
+    "DECIMAL(12, 27)": "DecimalType(12, 27)",
+    "DECIMAL(12, 28)": "DecimalType(12, 28)",
+    "DECIMAL(12, 29)": "DecimalType(12, 29)",
+    "DECIMAL(12, 30)": "DecimalType(12, 30)",
+    "DECIMAL(12, 31)": "DecimalType(12, 31)",
+    "DECIMAL(12, 32)": "DecimalType(12, 32)",
+    "DECIMAL(12, 33)": "DecimalType(12, 33)",
+    "DECIMAL(12, 34)": "DecimalType(12, 34)",
+    "DECIMAL(12, 35)": "DecimalType(12, 35)",
+    "DECIMAL(12, 36)": "DecimalType(12, 36)",
+    "DECIMAL(12, 37)": "DecimalType(12, 37)",
+    "DECIMAL(12, 38)": "DecimalType(12, 38)",
+    "DECIMAL(13, 0)": "DecimalType(13, 0)",
+    "DECIMAL(13, 1)": "DecimalType(13, 1)",
+    "DECIMAL(13, 2)": "DecimalType(13, 2)",
+    "DECIMAL(13, 3)": "DecimalType(13, 3)",
+    "DECIMAL(13, 4)": "DecimalType(13, 4)",
+    "DECIMAL(13, 5)": "DecimalType(13, 5)",
+    "DECIMAL(13, 6)": "DecimalType(13, 6)",
+    "DECIMAL(13, 7)": "DecimalType(13, 7)",
+    "DECIMAL(13, 8)": "DecimalType(13, 8)",
+    "DECIMAL(13, 9)": "DecimalType(13, 9)",
+    "DECIMAL(13, 10)": "DecimalType(13, 10)",
+    "DECIMAL(13, 11)": "DecimalType(13, 11)",
+    "DECIMAL(13, 12)": "DecimalType(13, 12)",
+    "DECIMAL(13, 13)": "DecimalType(13, 13)",
+    "DECIMAL(13, 14)": "DecimalType(13, 14)",
+    "DECIMAL(13, 15)": "DecimalType(13, 15)",
+    "DECIMAL(13, 16)": "DecimalType(13, 16)",
+    "DECIMAL(13, 17)": "DecimalType(13, 17)",
+    "DECIMAL(13, 18)": "DecimalType(13, 18)",
+    "DECIMAL(13, 19)": "DecimalType(13, 19)",
+    "DECIMAL(13, 20)": "DecimalType(13, 20)",
+    "DECIMAL(13, 21)": "DecimalType(13, 21)",
+    "DECIMAL(13, 22)": "DecimalType(13, 22)",
+    "DECIMAL(13, 23)": "DecimalType(13, 23)",
+    "DECIMAL(13, 24)": "DecimalType(13, 24)",
+    "DECIMAL(13, 25)": "DecimalType(13, 25)",
+    "DECIMAL(13, 26)": "DecimalType(13, 26)",
+    "DECIMAL(13, 27)": "DecimalType(13, 27)",
+    "DECIMAL(13, 28)": "DecimalType(13, 28)",
+    "DECIMAL(13, 29)": "DecimalType(13, 29)",
+    "DECIMAL(13, 30)": "DecimalType(13, 30)",
+    "DECIMAL(13, 31)": "DecimalType(13, 31)",
+    "DECIMAL(13, 32)": "DecimalType(13, 32)",
+    "DECIMAL(13, 33)": "DecimalType(13, 33)",
+    "DECIMAL(13, 34)": "DecimalType(13, 34)",
+    "DECIMAL(13, 35)": "DecimalType(13, 35)",
+    "DECIMAL(13, 36)": "DecimalType(13, 36)",
+    "DECIMAL(13, 37)": "DecimalType(13, 37)",
+    "DECIMAL(13, 38)": "DecimalType(13, 38)",
+    "DECIMAL(14, 0)": "DecimalType(14, 0)",
+    "DECIMAL(14, 1)": "DecimalType(14, 1)",
+    "DECIMAL(14, 2)": "DecimalType(14, 2)",
+    "DECIMAL(14, 3)": "DecimalType(14, 3)",
+    "DECIMAL(14, 4)": "DecimalType(14, 4)",
+    "DECIMAL(14, 5)": "DecimalType(14, 5)",
+    "DECIMAL(14, 6)": "DecimalType(14, 6)",
+    "DECIMAL(14, 7)": "DecimalType(14, 7)",
+    "DECIMAL(14, 8)": "DecimalType(14, 8)",
+    "DECIMAL(14, 9)": "DecimalType(14, 9)",
+    "DECIMAL(14, 10)": "DecimalType(14, 10)",
+    "DECIMAL(14, 11)": "DecimalType(14, 11)",
+    "DECIMAL(14, 12)": "DecimalType(14, 12)",
+    "DECIMAL(14, 13)": "DecimalType(14, 13)",
+    "DECIMAL(14, 14)": "DecimalType(14, 14)",
+    "DECIMAL(14, 15)": "DecimalType(14, 15)",
+    "DECIMAL(14, 16)": "DecimalType(14, 16)",
+    "DECIMAL(14, 17)": "DecimalType(14, 17)",
+    "DECIMAL(14, 18)": "DecimalType(14, 18)",
+    "DECIMAL(14, 19)": "DecimalType(14, 19)",
+    "DECIMAL(14, 20)": "DecimalType(14, 20)",
+    "DECIMAL(14, 21)": "DecimalType(14, 21)",
+    "DECIMAL(14, 22)": "DecimalType(14, 22)",
+    "DECIMAL(14, 23)": "DecimalType(14, 23)",
+    "DECIMAL(14, 24)": "DecimalType(14, 24)",
+    "DECIMAL(14, 25)": "DecimalType(14, 25)",
+    "DECIMAL(14, 26)": "DecimalType(14, 26)",
+    "DECIMAL(14, 27)": "DecimalType(14, 27)",
+    "DECIMAL(14, 28)": "DecimalType(14, 28)",
+    "DECIMAL(14, 29)": "DecimalType(14, 29)",
+    "DECIMAL(14, 30)": "DecimalType(14, 30)",
+    "DECIMAL(14, 31)": "DecimalType(14, 31)",
+    "DECIMAL(14, 32)": "DecimalType(14, 32)",
+    "DECIMAL(14, 33)": "DecimalType(14, 33)",
+    "DECIMAL(14, 34)": "DecimalType(14, 34)",
+    "DECIMAL(14, 35)": "DecimalType(14, 35)",
+    "DECIMAL(14, 36)": "DecimalType(14, 36)",
+    "DECIMAL(14, 37)": "DecimalType(14, 37)",
+    "DECIMAL(14, 38)": "DecimalType(14, 38)",
+    "DECIMAL(15, 0)": "DecimalType(15, 0)",
+    "DECIMAL(15, 1)": "DecimalType(15, 1)",
+    "DECIMAL(15, 2)": "DecimalType(15, 2)",
+    "DECIMAL(15, 3)": "DecimalType(15, 3)",
+    "DECIMAL(15, 4)": "DecimalType(15, 4)",
+    "DECIMAL(15, 5)": "DecimalType(15, 5)",
+    "DECIMAL(15, 6)": "DecimalType(15, 6)",
+    "DECIMAL(15, 7)": "DecimalType(15, 7)",
+    "DECIMAL(15, 8)": "DecimalType(15, 8)",
+    "DECIMAL(15, 9)": "DecimalType(15, 9)",
+    "DECIMAL(15, 10)": "DecimalType(15, 10)",
+    "DECIMAL(15, 11)": "DecimalType(15, 11)",
+    "DECIMAL(15, 12)": "DecimalType(15, 12)",
+    "DECIMAL(15, 13)": "DecimalType(15, 13)",
+    "DECIMAL(15, 14)": "DecimalType(15, 14)",
+    "DECIMAL(15, 15)": "DecimalType(15, 15)",
+    "DECIMAL(15, 16)": "DecimalType(15, 16)",
+    "DECIMAL(15, 17)": "DecimalType(15, 17)",
+    "DECIMAL(15, 18)": "DecimalType(15, 18)",
+    "DECIMAL(15, 19)": "DecimalType(15, 19)",
+    "DECIMAL(15, 20)": "DecimalType(15, 20)",
+    "DECIMAL(15, 21)": "DecimalType(15, 21)",
+    "DECIMAL(15, 22)": "DecimalType(15, 22)",
+    "DECIMAL(15, 23)": "DecimalType(15, 23)",
+    "DECIMAL(15, 24)": "DecimalType(15, 24)",
+    "DECIMAL(15, 25)": "DecimalType(15, 25)",
+    "DECIMAL(15, 26)": "DecimalType(15, 26)",
+    "DECIMAL(15, 27)": "DecimalType(15, 27)",
+    "DECIMAL(15, 28)": "DecimalType(15, 28)",
+    "DECIMAL(15, 29)": "DecimalType(15, 29)",
+    "DECIMAL(15, 30)": "DecimalType(15, 30)",
+    "DECIMAL(15, 31)": "DecimalType(15, 31)",
+    "DECIMAL(15, 32)": "DecimalType(15, 32)",
+    "DECIMAL(15, 33)": "DecimalType(15, 33)",
+    "DECIMAL(15, 34)": "DecimalType(15, 34)",
+    "DECIMAL(15, 35)": "DecimalType(15, 35)",
+    "DECIMAL(15, 36)": "DecimalType(15, 36)",
+    "DECIMAL(15, 37)": "DecimalType(15, 37)",
+    "DECIMAL(15, 38)": "DecimalType(15, 38)",
+    "DECIMAL(16, 0)": "DecimalType(16, 0)",
+    "DECIMAL(16, 1)": "DecimalType(16, 1)",
+    "DECIMAL(16, 2)": "DecimalType(16, 2)",
+    "DECIMAL(16, 3)": "DecimalType(16, 3)",
+    "DECIMAL(16, 4)": "DecimalType(16, 4)",
+    "DECIMAL(16, 5)": "DecimalType(16, 5)",
+    "DECIMAL(16, 6)": "DecimalType(16, 6)",
+    "DECIMAL(16, 7)": "DecimalType(16, 7)",
+    "DECIMAL(16, 8)": "DecimalType(16, 8)",
+    "DECIMAL(16, 9)": "DecimalType(16, 9)",
+    "DECIMAL(16, 10)": "DecimalType(16, 10)",
+    "DECIMAL(16, 11)": "DecimalType(16, 11)",
+    "DECIMAL(16, 12)": "DecimalType(16, 12)",
+    "DECIMAL(16, 13)": "DecimalType(16, 13)",
+    "DECIMAL(16, 14)": "DecimalType(16, 14)",
+    "DECIMAL(16, 15)": "DecimalType(16, 15)",
+    "DECIMAL(16, 16)": "DecimalType(16, 16)",
+    "DECIMAL(16, 17)": "DecimalType(16, 17)",
+    "DECIMAL(16, 18)": "DecimalType(16, 18)",
+    "DECIMAL(16, 19)": "DecimalType(16, 19)",
+    "DECIMAL(16, 20)": "DecimalType(16, 20)",
+    "DECIMAL(16, 21)": "DecimalType(16, 21)",
+    "DECIMAL(16, 22)": "DecimalType(16, 22)",
+    "DECIMAL(16, 23)": "DecimalType(16, 23)",
+    "DECIMAL(16, 24)": "DecimalType(16, 24)",
+    "DECIMAL(16, 25)": "DecimalType(16, 25)",
+    "DECIMAL(16, 26)": "DecimalType(16, 26)",
+    "DECIMAL(16, 27)": "DecimalType(16, 27)",
+    "DECIMAL(16, 28)": "DecimalType(16, 28)",
+    "DECIMAL(16, 29)": "DecimalType(16, 29)",
+    "DECIMAL(16, 30)": "DecimalType(16, 30)",
+    "DECIMAL(16, 31)": "DecimalType(16, 31)",
+    "DECIMAL(16, 32)": "DecimalType(16, 32)",
+    "DECIMAL(16, 33)": "DecimalType(16, 33)",
+    "DECIMAL(16, 34)": "DecimalType(16, 34)",
+    "DECIMAL(16, 35)": "DecimalType(16, 35)",
+    "DECIMAL(16, 36)": "DecimalType(16, 36)",
+    "DECIMAL(16, 37)": "DecimalType(16, 37)",
+    "DECIMAL(16, 38)": "DecimalType(16, 38)",
+    "DECIMAL(17, 0)": "DecimalType(17, 0)",
+    "DECIMAL(17, 1)": "DecimalType(17, 1)",
+    "DECIMAL(17, 2)": "DecimalType(17, 2)",
+    "DECIMAL(17, 3)": "DecimalType(17, 3)",
+    "DECIMAL(17, 4)": "DecimalType(17, 4)",
+    "DECIMAL(17, 5)": "DecimalType(17, 5)",
+    "DECIMAL(17, 6)": "DecimalType(17, 6)",
+    "DECIMAL(17, 7)": "DecimalType(17, 7)",
+    "DECIMAL(17, 8)": "DecimalType(17, 8)",
+    "DECIMAL(17, 9)": "DecimalType(17, 9)",
+    "DECIMAL(17, 10)": "DecimalType(17, 10)",
+    "DECIMAL(17, 11)": "DecimalType(17, 11)",
+    "DECIMAL(17, 12)": "DecimalType(17, 12)",
+    "DECIMAL(17, 13)": "DecimalType(17, 13)",
+    "DECIMAL(17, 14)": "DecimalType(17, 14)",
+    "DECIMAL(17, 15)": "DecimalType(17, 15)",
+    "DECIMAL(17, 16)": "DecimalType(17, 16)",
+    "DECIMAL(17, 17)": "DecimalType(17, 17)",
+    "DECIMAL(17, 18)": "DecimalType(17, 18)",
+    "DECIMAL(17, 19)": "DecimalType(17, 19)",
+    "DECIMAL(17, 20)": "DecimalType(17, 20)",
+    "DECIMAL(17, 21)": "DecimalType(17, 21)",
+    "DECIMAL(17, 22)": "DecimalType(17, 22)",
+    "DECIMAL(17, 23)": "DecimalType(17, 23)",
+    "DECIMAL(17, 24)": "DecimalType(17, 24)",
+    "DECIMAL(17, 25)": "DecimalType(17, 25)",
+    "DECIMAL(17, 26)": "DecimalType(17, 26)",
+    "DECIMAL(17, 27)": "DecimalType(17, 27)",
+    "DECIMAL(17, 28)": "DecimalType(17, 28)",
+    "DECIMAL(17, 29)": "DecimalType(17, 29)",
+    "DECIMAL(17, 30)": "DecimalType(17, 30)",
+    "DECIMAL(17, 31)": "DecimalType(17, 31)",
+    "DECIMAL(17, 32)": "DecimalType(17, 32)",
+    "DECIMAL(17, 33)": "DecimalType(17, 33)",
+    "DECIMAL(17, 34)": "DecimalType(17, 34)",
+    "DECIMAL(17, 35)": "DecimalType(17, 35)",
+    "DECIMAL(17, 36)": "DecimalType(17, 36)",
+    "DECIMAL(17, 37)": "DecimalType(17, 37)",
+    "DECIMAL(17, 38)": "DecimalType(17, 38)",
+    "DECIMAL(18, 0)": "DecimalType(18, 0)",
+    "DECIMAL(18, 1)": "DecimalType(18, 1)",
+    "DECIMAL(18, 2)": "DecimalType(18, 2)",
+    "DECIMAL(18, 3)": "DecimalType(18, 3)",
+    "DECIMAL(18, 4)": "DecimalType(18, 4)",
+    "DECIMAL(18, 5)": "DecimalType(18, 5)",
+    "DECIMAL(18, 6)": "DecimalType(18, 6)",
+    "DECIMAL(18, 7)": "DecimalType(18, 7)",
+    "DECIMAL(18, 8)": "DecimalType(18, 8)",
+    "DECIMAL(18, 9)": "DecimalType(18, 9)",
+    "DECIMAL(18, 10)": "DecimalType(18, 10)",
+    "DECIMAL(18, 11)": "DecimalType(18, 11)",
+    "DECIMAL(18, 12)": "DecimalType(18, 12)",
+    "DECIMAL(18, 13)": "DecimalType(18, 13)",
+    "DECIMAL(18, 14)": "DecimalType(18, 14)",
+    "DECIMAL(18, 15)": "DecimalType(18, 15)",
+    "DECIMAL(18, 16)": "DecimalType(18, 16)",
+    "DECIMAL(18, 17)": "DecimalType(18, 17)",
+    "DECIMAL(18, 18)": "DecimalType(18, 18)",
+    "DECIMAL(18, 19)": "DecimalType(18, 19)",
+    "DECIMAL(18, 20)": "DecimalType(18, 20)",
+    "DECIMAL(18, 21)": "DecimalType(18, 21)",
+    "DECIMAL(18, 22)": "DecimalType(18, 22)",
+    "DECIMAL(18, 23)": "DecimalType(18, 23)",
+    "DECIMAL(18, 24)": "DecimalType(18, 24)",
+    "DECIMAL(18, 25)": "DecimalType(18, 25)",
+    "DECIMAL(18, 26)": "DecimalType(18, 26)",
+    "DECIMAL(18, 27)": "DecimalType(18, 27)",
+    "DECIMAL(18, 28)": "DecimalType(18, 28)",
+    "DECIMAL(18, 29)": "DecimalType(18, 29)",
+    "DECIMAL(18, 30)": "DecimalType(18, 30)",
+    "DECIMAL(18, 31)": "DecimalType(18, 31)",
+    "DECIMAL(18, 32)": "DecimalType(18, 32)",
+    "DECIMAL(18, 33)": "DecimalType(18, 33)",
+    "DECIMAL(18, 34)": "DecimalType(18, 34)",
+    "DECIMAL(18, 35)": "DecimalType(18, 35)",
+    "DECIMAL(18, 36)": "DecimalType(18, 36)",
+    "DECIMAL(18, 37)": "DecimalType(18, 37)",
+    "DECIMAL(18, 38)": "DecimalType(18, 38)",
+    "DECIMAL(19, 0)": "DecimalType(19, 0)",
+    "DECIMAL(19, 1)": "DecimalType(19, 1)",
+    "DECIMAL(19, 2)": "DecimalType(19, 2)",
+    "DECIMAL(19, 3)": "DecimalType(19, 3)",
+    "DECIMAL(19, 4)": "DecimalType(19, 4)",
+    "DECIMAL(19, 5)": "DecimalType(19, 5)",
+    "DECIMAL(19, 6)": "DecimalType(19, 6)",
+    "DECIMAL(19, 7)": "DecimalType(19, 7)",
+    "DECIMAL(19, 8)": "DecimalType(19, 8)",
+    "DECIMAL(19, 9)": "DecimalType(19, 9)",
+    "DECIMAL(19, 10)": "DecimalType(19, 10)",
+    "DECIMAL(19, 11)": "DecimalType(19, 11)",
+    "DECIMAL(19, 12)": "DecimalType(19, 12)",
+    "DECIMAL(19, 13)": "DecimalType(19, 13)",
+    "DECIMAL(19, 14)": "DecimalType(19, 14)",
+    "DECIMAL(19, 15)": "DecimalType(19, 15)",
+    "DECIMAL(19, 16)": "DecimalType(19, 16)",
+    "DECIMAL(19, 17)": "DecimalType(19, 17)",
+    "DECIMAL(19, 18)": "DecimalType(19, 18)",
+    "DECIMAL(19, 19)": "DecimalType(19, 19)",
+    "DECIMAL(19, 20)": "DecimalType(19, 20)",
+    "DECIMAL(19, 21)": "DecimalType(19, 21)",
+    "DECIMAL(19, 22)": "DecimalType(19, 22)",
+    "DECIMAL(19, 23)": "DecimalType(19, 23)",
+    "DECIMAL(19, 24)": "DecimalType(19, 24)",
+    "DECIMAL(19, 25)": "DecimalType(19, 25)",
+    "DECIMAL(19, 26)": "DecimalType(19, 26)",
+    "DECIMAL(19, 27)": "DecimalType(19, 27)",
+    "DECIMAL(19, 28)": "DecimalType(19, 28)",
+    "DECIMAL(19, 29)": "DecimalType(19, 29)",
+    "DECIMAL(19, 30)": "DecimalType(19, 30)",
+    "DECIMAL(19, 31)": "DecimalType(19, 31)",
+    "DECIMAL(19, 32)": "DecimalType(19, 32)",
+    "DECIMAL(19, 33)": "DecimalType(19, 33)",
+    "DECIMAL(19, 34)": "DecimalType(19, 34)",
+    "DECIMAL(19, 35)": "DecimalType(19, 35)",
+    "DECIMAL(19, 36)": "DecimalType(19, 36)",
+    "DECIMAL(19, 37)": "DecimalType(19, 37)",
+    "DECIMAL(19, 38)": "DecimalType(19, 38)",
+    "DECIMAL(20, 0)": "DecimalType(20, 0)",
+    "DECIMAL(20, 1)": "DecimalType(20, 1)",
+    "DECIMAL(20, 2)": "DecimalType(20, 2)",
+    "DECIMAL(20, 3)": "DecimalType(20, 3)",
+    "DECIMAL(20, 4)": "DecimalType(20, 4)",
+    "DECIMAL(20, 5)": "DecimalType(20, 5)",
+    "DECIMAL(20, 6)": "DecimalType(20, 6)",
+    "DECIMAL(20, 7)": "DecimalType(20, 7)",
+    "DECIMAL(20, 8)": "DecimalType(20, 8)",
+    "DECIMAL(20, 9)": "DecimalType(20, 9)",
+    "DECIMAL(20, 10)": "DecimalType(20, 10)",
+    "DECIMAL(20, 11)": "DecimalType(20, 11)",
+    "DECIMAL(20, 12)": "DecimalType(20, 12)",
+    "DECIMAL(20, 13)": "DecimalType(20, 13)",
+    "DECIMAL(20, 14)": "DecimalType(20, 14)",
+    "DECIMAL(20, 15)": "DecimalType(20, 15)",
+    "DECIMAL(20, 16)": "DecimalType(20, 16)",
+    "DECIMAL(20, 17)": "DecimalType(20, 17)",
+    "DECIMAL(20, 18)": "DecimalType(20, 18)",
+    "DECIMAL(20, 19)": "DecimalType(20, 19)",
+    "DECIMAL(20, 20)": "DecimalType(20, 20)",
+    "DECIMAL(20, 21)": "DecimalType(20, 21)",
+    "DECIMAL(20, 22)": "DecimalType(20, 22)",
+    "DECIMAL(20, 23)": "DecimalType(20, 23)",
+    "DECIMAL(20, 24)": "DecimalType(20, 24)",
+    "DECIMAL(20, 25)": "DecimalType(20, 25)",
+    "DECIMAL(20, 26)": "DecimalType(20, 26)",
+    "DECIMAL(20, 27)": "DecimalType(20, 27)",
+    "DECIMAL(20, 28)": "DecimalType(20, 28)",
+    "DECIMAL(20, 29)": "DecimalType(20, 29)",
+    "DECIMAL(20, 30)": "DecimalType(20, 30)",
+    "DECIMAL(20, 31)": "DecimalType(20, 31)",
+    "DECIMAL(20, 32)": "DecimalType(20, 32)",
+    "DECIMAL(20, 33)": "DecimalType(20, 33)",
+    "DECIMAL(20, 34)": "DecimalType(20, 34)",
+    "DECIMAL(20, 35)": "DecimalType(20, 35)",
+    "DECIMAL(20, 36)": "DecimalType(20, 36)",
+    "DECIMAL(20, 37)": "DecimalType(20, 37)",
+    "DECIMAL(20, 38)": "DecimalType(20, 38)",
+    "DECIMAL(21, 0)": "DecimalType(21, 0)",
+    "DECIMAL(21, 1)": "DecimalType(21, 1)",
+    "DECIMAL(21, 2)": "DecimalType(21, 2)",
+    "DECIMAL(21, 3)": "DecimalType(21, 3)",
+    "DECIMAL(21, 4)": "DecimalType(21, 4)",
+    "DECIMAL(21, 5)": "DecimalType(21, 5)",
+    "DECIMAL(21, 6)": "DecimalType(21, 6)",
+    "DECIMAL(21, 7)": "DecimalType(21, 7)",
+    "DECIMAL(21, 8)": "DecimalType(21, 8)",
+    "DECIMAL(21, 9)": "DecimalType(21, 9)",
+    "DECIMAL(21, 10)": "DecimalType(21, 10)",
+    "DECIMAL(21, 11)": "DecimalType(21, 11)",
+    "DECIMAL(21, 12)": "DecimalType(21, 12)",
+    "DECIMAL(21, 13)": "DecimalType(21, 13)",
+    "DECIMAL(21, 14)": "DecimalType(21, 14)",
+    "DECIMAL(21, 15)": "DecimalType(21, 15)",
+    "DECIMAL(21, 16)": "DecimalType(21, 16)",
+    "DECIMAL(21, 17)": "DecimalType(21, 17)",
+    "DECIMAL(21, 18)": "DecimalType(21, 18)",
+    "DECIMAL(21, 19)": "DecimalType(21, 19)",
+    "DECIMAL(21, 20)": "DecimalType(21, 20)",
+    "DECIMAL(21, 21)": "DecimalType(21, 21)",
+    "DECIMAL(21, 22)": "DecimalType(21, 22)",
+    "DECIMAL(21, 23)": "DecimalType(21, 23)",
+    "DECIMAL(21, 24)": "DecimalType(21, 24)",
+    "DECIMAL(21, 25)": "DecimalType(21, 25)",
+    "DECIMAL(21, 26)": "DecimalType(21, 26)",
+    "DECIMAL(21, 27)": "DecimalType(21, 27)",
+    "DECIMAL(21, 28)": "DecimalType(21, 28)",
+    "DECIMAL(21, 29)": "DecimalType(21, 29)",
+    "DECIMAL(21, 30)": "DecimalType(21, 30)",
+    "DECIMAL(21, 31)": "DecimalType(21, 31)",
+    "DECIMAL(21, 32)": "DecimalType(21, 32)",
+    "DECIMAL(21, 33)": "DecimalType(21, 33)",
+    "DECIMAL(21, 34)": "DecimalType(21, 34)",
+    "DECIMAL(21, 35)": "DecimalType(21, 35)",
+    "DECIMAL(21, 36)": "DecimalType(21, 36)",
+    "DECIMAL(21, 37)": "DecimalType(21, 37)",
+    "DECIMAL(21, 38)": "DecimalType(21, 38)",
+    "DECIMAL(22, 0)": "DecimalType(22, 0)",
+    "DECIMAL(22, 1)": "DecimalType(22, 1)",
+    "DECIMAL(22, 2)": "DecimalType(22, 2)",
+    "DECIMAL(22, 3)": "DecimalType(22, 3)",
+    "DECIMAL(22, 4)": "DecimalType(22, 4)",
+    "DECIMAL(22, 5)": "DecimalType(22, 5)",
+    "DECIMAL(22, 6)": "DecimalType(22, 6)",
+    "DECIMAL(22, 7)": "DecimalType(22, 7)",
+    "DECIMAL(22, 8)": "DecimalType(22, 8)",
+    "DECIMAL(22, 9)": "DecimalType(22, 9)",
+    "DECIMAL(22, 10)": "DecimalType(22, 10)",
+    "DECIMAL(22, 11)": "DecimalType(22, 11)",
+    "DECIMAL(22, 12)": "DecimalType(22, 12)",
+    "DECIMAL(22, 13)": "DecimalType(22, 13)",
+    "DECIMAL(22, 14)": "DecimalType(22, 14)",
+    "DECIMAL(22, 15)": "DecimalType(22, 15)",
+    "DECIMAL(22, 16)": "DecimalType(22, 16)",
+    "DECIMAL(22, 17)": "DecimalType(22, 17)",
+    "DECIMAL(22, 18)": "DecimalType(22, 18)",
+    "DECIMAL(22, 19)": "DecimalType(22, 19)",
+    "DECIMAL(22, 20)": "DecimalType(22, 20)",
+    "DECIMAL(22, 21)": "DecimalType(22, 21)",
+    "DECIMAL(22, 22)": "DecimalType(22, 22)",
+    "DECIMAL(22, 23)": "DecimalType(22, 23)",
+    "DECIMAL(22, 24)": "DecimalType(22, 24)",
+    "DECIMAL(22, 25)": "DecimalType(22, 25)",
+    "DECIMAL(22, 26)": "DecimalType(22, 26)",
+    "DECIMAL(22, 27)": "DecimalType(22, 27)",
+    "DECIMAL(22, 28)": "DecimalType(22, 28)",
+    "DECIMAL(22, 29)": "DecimalType(22, 29)",
+    "DECIMAL(22, 30)": "DecimalType(22, 30)",
+    "DECIMAL(22, 31)": "DecimalType(22, 31)",
+    "DECIMAL(22, 32)": "DecimalType(22, 32)",
+    "DECIMAL(22, 33)": "DecimalType(22, 33)",
+    "DECIMAL(22, 34)": "DecimalType(22, 34)",
+    "DECIMAL(22, 35)": "DecimalType(22, 35)",
+    "DECIMAL(22, 36)": "DecimalType(22, 36)",
+    "DECIMAL(22, 37)": "DecimalType(22, 37)",
+    "DECIMAL(22, 38)": "DecimalType(22, 38)",
+    "DECIMAL(23, 0)": "DecimalType(23, 0)",
+    "DECIMAL(23, 1)": "DecimalType(23, 1)",
+    "DECIMAL(23, 2)": "DecimalType(23, 2)",
+    "DECIMAL(23, 3)": "DecimalType(23, 3)",
+    "DECIMAL(23, 4)": "DecimalType(23, 4)",
+    "DECIMAL(23, 5)": "DecimalType(23, 5)",
+    "DECIMAL(23, 6)": "DecimalType(23, 6)",
+    "DECIMAL(23, 7)": "DecimalType(23, 7)",
+    "DECIMAL(23, 8)": "DecimalType(23, 8)",
+    "DECIMAL(23, 9)": "DecimalType(23, 9)",
+    "DECIMAL(23, 10)": "DecimalType(23, 10)",
+    "DECIMAL(23, 11)": "DecimalType(23, 11)",
+    "DECIMAL(23, 12)": "DecimalType(23, 12)",
+    "DECIMAL(23, 13)": "DecimalType(23, 13)",
+    "DECIMAL(23, 14)": "DecimalType(23, 14)",
+    "DECIMAL(23, 15)": "DecimalType(23, 15)",
+    "DECIMAL(23, 16)": "DecimalType(23, 16)",
+    "DECIMAL(23, 17)": "DecimalType(23, 17)",
+    "DECIMAL(23, 18)": "DecimalType(23, 18)",
+    "DECIMAL(23, 19)": "DecimalType(23, 19)",
+    "DECIMAL(23, 20)": "DecimalType(23, 20)",
+    "DECIMAL(23, 21)": "DecimalType(23, 21)",
+    "DECIMAL(23, 22)": "DecimalType(23, 22)",
+    "DECIMAL(23, 23)": "DecimalType(23, 23)",
+    "DECIMAL(23, 24)": "DecimalType(23, 24)",
+    "DECIMAL(23, 25)": "DecimalType(23, 25)",
+    "DECIMAL(23, 26)": "DecimalType(23, 26)",
+    "DECIMAL(23, 27)": "DecimalType(23, 27)",
+    "DECIMAL(23, 28)": "DecimalType(23, 28)",
+    "DECIMAL(23, 29)": "DecimalType(23, 29)",
+    "DECIMAL(23, 30)": "DecimalType(23, 30)",
+    "DECIMAL(23, 31)": "DecimalType(23, 31)",
+    "DECIMAL(23, 32)": "DecimalType(23, 32)",
+    "DECIMAL(23, 33)": "DecimalType(23, 33)",
+    "DECIMAL(23, 34)": "DecimalType(23, 34)",
+    "DECIMAL(23, 35)": "DecimalType(23, 35)",
+    "DECIMAL(23, 36)": "DecimalType(23, 36)",
+    "DECIMAL(23, 37)": "DecimalType(23, 37)",
+    "DECIMAL(23, 38)": "DecimalType(23, 38)",
+    "DECIMAL(24, 0)": "DecimalType(24, 0)",
+    "DECIMAL(24, 1)": "DecimalType(24, 1)",
+    "DECIMAL(24, 2)": "DecimalType(24, 2)",
+    "DECIMAL(24, 3)": "DecimalType(24, 3)",
+    "DECIMAL(24, 4)": "DecimalType(24, 4)",
+    "DECIMAL(24, 5)": "DecimalType(24, 5)",
+    "DECIMAL(24, 6)": "DecimalType(24, 6)",
+    "DECIMAL(24, 7)": "DecimalType(24, 7)",
+    "DECIMAL(24, 8)": "DecimalType(24, 8)",
+    "DECIMAL(24, 9)": "DecimalType(24, 9)",
+    "DECIMAL(24, 10)": "DecimalType(24, 10)",
+    "DECIMAL(24, 11)": "DecimalType(24, 11)",
+    "DECIMAL(24, 12)": "DecimalType(24, 12)",
+    "DECIMAL(24, 13)": "DecimalType(24, 13)",
+    "DECIMAL(24, 14)": "DecimalType(24, 14)",
+    "DECIMAL(24, 15)": "DecimalType(24, 15)",
+    "DECIMAL(24, 16)": "DecimalType(24, 16)",
+    "DECIMAL(24, 17)": "DecimalType(24, 17)",
+    "DECIMAL(24, 18)": "DecimalType(24, 18)",
+    "DECIMAL(24, 19)": "DecimalType(24, 19)",
+    "DECIMAL(24, 20)": "DecimalType(24, 20)",
+    "DECIMAL(24, 21)": "DecimalType(24, 21)",
+    "DECIMAL(24, 22)": "DecimalType(24, 22)",
+    "DECIMAL(24, 23)": "DecimalType(24, 23)",
+    "DECIMAL(24, 24)": "DecimalType(24, 24)",
+    "DECIMAL(24, 25)": "DecimalType(24, 25)",
+    "DECIMAL(24, 26)": "DecimalType(24, 26)",
+    "DECIMAL(24, 27)": "DecimalType(24, 27)",
+    "DECIMAL(24, 28)": "DecimalType(24, 28)",
+    "DECIMAL(24, 29)": "DecimalType(24, 29)",
+    "DECIMAL(24, 30)": "DecimalType(24, 30)",
+    "DECIMAL(24, 31)": "DecimalType(24, 31)",
+    "DECIMAL(24, 32)": "DecimalType(24, 32)",
+    "DECIMAL(24, 33)": "DecimalType(24, 33)",
+    "DECIMAL(24, 34)": "DecimalType(24, 34)",
+    "DECIMAL(24, 35)": "DecimalType(24, 35)",
+    "DECIMAL(24, 36)": "DecimalType(24, 36)",
+    "DECIMAL(24, 37)": "DecimalType(24, 37)",
+    "DECIMAL(24, 38)": "DecimalType(24, 38)",
+    "DECIMAL(25, 0)": "DecimalType(25, 0)",
+    "DECIMAL(25, 1)": "DecimalType(25, 1)",
+    "DECIMAL(25, 2)": "DecimalType(25, 2)",
+    "DECIMAL(25, 3)": "DecimalType(25, 3)",
+    "DECIMAL(25, 4)": "DecimalType(25, 4)",
+    "DECIMAL(25, 5)": "DecimalType(25, 5)",
+    "DECIMAL(25, 6)": "DecimalType(25, 6)",
+    "DECIMAL(25, 7)": "DecimalType(25, 7)",
+    "DECIMAL(25, 8)": "DecimalType(25, 8)",
+    "DECIMAL(25, 9)": "DecimalType(25, 9)",
+    "DECIMAL(25, 10)": "DecimalType(25, 10)",
+    "DECIMAL(25, 11)": "DecimalType(25, 11)",
+    "DECIMAL(25, 12)": "DecimalType(25, 12)",
+    "DECIMAL(25, 13)": "DecimalType(25, 13)",
+    "DECIMAL(25, 14)": "DecimalType(25, 14)",
+    "DECIMAL(25, 15)": "DecimalType(25, 15)",
+    "DECIMAL(25, 16)": "DecimalType(25, 16)",
+    "DECIMAL(25, 17)": "DecimalType(25, 17)",
+    "DECIMAL(25, 18)": "DecimalType(25, 18)",
+    "DECIMAL(25, 19)": "DecimalType(25, 19)",
+    "DECIMAL(25, 20)": "DecimalType(25, 20)",
+    "DECIMAL(25, 21)": "DecimalType(25, 21)",
+    "DECIMAL(25, 22)": "DecimalType(25, 22)",
+    "DECIMAL(25, 23)": "DecimalType(25, 23)",
+    "DECIMAL(25, 24)": "DecimalType(25, 24)",
+    "DECIMAL(25, 25)": "DecimalType(25, 25)",
+    "DECIMAL(25, 26)": "DecimalType(25, 26)",
+    "DECIMAL(25, 27)": "DecimalType(25, 27)",
+    "DECIMAL(25, 28)": "DecimalType(25, 28)",
+    "DECIMAL(25, 29)": "DecimalType(25, 29)",
+    "DECIMAL(25, 30)": "DecimalType(25, 30)",
+    "DECIMAL(25, 31)": "DecimalType(25, 31)",
+    "DECIMAL(25, 32)": "DecimalType(25, 32)",
+    "DECIMAL(25, 33)": "DecimalType(25, 33)",
+    "DECIMAL(25, 34)": "DecimalType(25, 34)",
+    "DECIMAL(25, 35)": "DecimalType(25, 35)",
+    "DECIMAL(25, 36)": "DecimalType(25, 36)",
+    "DECIMAL(25, 37)": "DecimalType(25, 37)",
+    "DECIMAL(25, 38)": "DecimalType(25, 38)",
+    "DECIMAL(26, 0)": "DecimalType(26, 0)",
+    "DECIMAL(26, 1)": "DecimalType(26, 1)",
+    "DECIMAL(26, 2)": "DecimalType(26, 2)",
+    "DECIMAL(26, 3)": "DecimalType(26, 3)",
+    "DECIMAL(26, 4)": "DecimalType(26, 4)",
+    "DECIMAL(26, 5)": "DecimalType(26, 5)",
+    "DECIMAL(26, 6)": "DecimalType(26, 6)",
+    "DECIMAL(26, 7)": "DecimalType(26, 7)",
+    "DECIMAL(26, 8)": "DecimalType(26, 8)",
+    "DECIMAL(26, 9)": "DecimalType(26, 9)",
+    "DECIMAL(26, 10)": "DecimalType(26, 10)",
+    "DECIMAL(26, 11)": "DecimalType(26, 11)",
+    "DECIMAL(26, 12)": "DecimalType(26, 12)",
+    "DECIMAL(26, 13)": "DecimalType(26, 13)",
+    "DECIMAL(26, 14)": "DecimalType(26, 14)",
+    "DECIMAL(26, 15)": "DecimalType(26, 15)",
+    "DECIMAL(26, 16)": "DecimalType(26, 16)",
+    "DECIMAL(26, 17)": "DecimalType(26, 17)",
+    "DECIMAL(26, 18)": "DecimalType(26, 18)",
+    "DECIMAL(26, 19)": "DecimalType(26, 19)",
+    "DECIMAL(26, 20)": "DecimalType(26, 20)",
+    "DECIMAL(26, 21)": "DecimalType(26, 21)",
+    "DECIMAL(26, 22)": "DecimalType(26, 22)",
+    "DECIMAL(26, 23)": "DecimalType(26, 23)",
+    "DECIMAL(26, 24)": "DecimalType(26, 24)",
+    "DECIMAL(26, 25)": "DecimalType(26, 25)",
+    "DECIMAL(26, 26)": "DecimalType(26, 26)",
+    "DECIMAL(26, 27)": "DecimalType(26, 27)",
+    "DECIMAL(26, 28)": "DecimalType(26, 28)",
+    "DECIMAL(26, 29)": "DecimalType(26, 29)",
+    "DECIMAL(26, 30)": "DecimalType(26, 30)",
+    "DECIMAL(26, 31)": "DecimalType(26, 31)",
+    "DECIMAL(26, 32)": "DecimalType(26, 32)",
+    "DECIMAL(26, 33)": "DecimalType(26, 33)",
+    "DECIMAL(26, 34)": "DecimalType(26, 34)",
+    "DECIMAL(26, 35)": "DecimalType(26, 35)",
+    "DECIMAL(26, 36)": "DecimalType(26, 36)",
+    "DECIMAL(26, 37)": "DecimalType(26, 37)",
+    "DECIMAL(26, 38)": "DecimalType(26, 38)",
+    "DECIMAL(27, 0)": "DecimalType(27, 0)",
+    "DECIMAL(27, 1)": "DecimalType(27, 1)",
+    "DECIMAL(27, 2)": "DecimalType(27, 2)",
+    "DECIMAL(27, 3)": "DecimalType(27, 3)",
+    "DECIMAL(27, 4)": "DecimalType(27, 4)",
+    "DECIMAL(27, 5)": "DecimalType(27, 5)",
+    "DECIMAL(27, 6)": "DecimalType(27, 6)",
+    "DECIMAL(27, 7)": "DecimalType(27, 7)",
+    "DECIMAL(27, 8)": "DecimalType(27, 8)",
+    "DECIMAL(27, 9)": "DecimalType(27, 9)",
+    "DECIMAL(27, 10)": "DecimalType(27, 10)",
+    "DECIMAL(27, 11)": "DecimalType(27, 11)",
+    "DECIMAL(27, 12)": "DecimalType(27, 12)",
+    "DECIMAL(27, 13)": "DecimalType(27, 13)",
+    "DECIMAL(27, 14)": "DecimalType(27, 14)",
+    "DECIMAL(27, 15)": "DecimalType(27, 15)",
+    "DECIMAL(27, 16)": "DecimalType(27, 16)",
+    "DECIMAL(27, 17)": "DecimalType(27, 17)",
+    "DECIMAL(27, 18)": "DecimalType(27, 18)",
+    "DECIMAL(27, 19)": "DecimalType(27, 19)",
+    "DECIMAL(27, 20)": "DecimalType(27, 20)",
+    "DECIMAL(27, 21)": "DecimalType(27, 21)",
+    "DECIMAL(27, 22)": "DecimalType(27, 22)",
+    "DECIMAL(27, 23)": "DecimalType(27, 23)",
+    "DECIMAL(27, 24)": "DecimalType(27, 24)",
+    "DECIMAL(27, 25)": "DecimalType(27, 25)",
+    "DECIMAL(27, 26)": "DecimalType(27, 26)",
+    "DECIMAL(27, 27)": "DecimalType(27, 27)",
+    "DECIMAL(27, 28)": "DecimalType(27, 28)",
+    "DECIMAL(27, 29)": "DecimalType(27, 29)",
+    "DECIMAL(27, 30)": "DecimalType(27, 30)",
+    "DECIMAL(27, 31)": "DecimalType(27, 31)",
+    "DECIMAL(27, 32)": "DecimalType(27, 32)",
+    "DECIMAL(27, 33)": "DecimalType(27, 33)",
+    "DECIMAL(27, 34)": "DecimalType(27, 34)",
+    "DECIMAL(27, 35)": "DecimalType(27, 35)",
+    "DECIMAL(27, 36)": "DecimalType(27, 36)",
+    "DECIMAL(27, 37)": "DecimalType(27, 37)",
+    "DECIMAL(27, 38)": "DecimalType(27, 38)",
+    "DECIMAL(28, 0)": "DecimalType(28, 0)",
+    "DECIMAL(28, 1)": "DecimalType(28, 1)",
+    "DECIMAL(28, 2)": "DecimalType(28, 2)",
+    "DECIMAL(28, 3)": "DecimalType(28, 3)",
+    "DECIMAL(28, 4)": "DecimalType(28, 4)",
+    "DECIMAL(28, 5)": "DecimalType(28, 5)",
+    "DECIMAL(28, 6)": "DecimalType(28, 6)",
+    "DECIMAL(28, 7)": "DecimalType(28, 7)",
+    "DECIMAL(28, 8)": "DecimalType(28, 8)",
+    "DECIMAL(28, 9)": "DecimalType(28, 9)",
+    "DECIMAL(28, 10)": "DecimalType(28, 10)",
+    "DECIMAL(28, 11)": "DecimalType(28, 11)",
+    "DECIMAL(28, 12)": "DecimalType(28, 12)",
+    "DECIMAL(28, 13)": "DecimalType(28, 13)",
+    "DECIMAL(28, 14)": "DecimalType(28, 14)",
+    "DECIMAL(28, 15)": "DecimalType(28, 15)",
+    "DECIMAL(28, 16)": "DecimalType(28, 16)",
+    "DECIMAL(28, 17)": "DecimalType(28, 17)",
+    "DECIMAL(28, 18)": "DecimalType(28, 18)",
+    "DECIMAL(28, 19)": "DecimalType(28, 19)",
+    "DECIMAL(28, 20)": "DecimalType(28, 20)",
+    "DECIMAL(28, 21)": "DecimalType(28, 21)",
+    "DECIMAL(28, 22)": "DecimalType(28, 22)",
+    "DECIMAL(28, 23)": "DecimalType(28, 23)",
+    "DECIMAL(28, 24)": "DecimalType(28, 24)",
+    "DECIMAL(28, 25)": "DecimalType(28, 25)",
+    "DECIMAL(28, 26)": "DecimalType(28, 26)",
+    "DECIMAL(28, 27)": "DecimalType(28, 27)",
+    "DECIMAL(28, 28)": "DecimalType(28, 28)",
+    "DECIMAL(28, 29)": "DecimalType(28, 29)",
+    "DECIMAL(28, 30)": "DecimalType(28, 30)",
+    "DECIMAL(28, 31)": "DecimalType(28, 31)",
+    "DECIMAL(28, 32)": "DecimalType(28, 32)",
+    "DECIMAL(28, 33)": "DecimalType(28, 33)",
+    "DECIMAL(28, 34)": "DecimalType(28, 34)",
+    "DECIMAL(28, 35)": "DecimalType(28, 35)",
+    "DECIMAL(28, 36)": "DecimalType(28, 36)",
+    "DECIMAL(28, 37)": "DecimalType(28, 37)",
+    "DECIMAL(28, 38)": "DecimalType(28, 38)",
+    "DECIMAL(29, 0)": "DecimalType(29, 0)",
+    "DECIMAL(29, 1)": "DecimalType(29, 1)",
+    "DECIMAL(29, 2)": "DecimalType(29, 2)",
+    "DECIMAL(29, 3)": "DecimalType(29, 3)",
+    "DECIMAL(29, 4)": "DecimalType(29, 4)",
+    "DECIMAL(29, 5)": "DecimalType(29, 5)",
+    "DECIMAL(29, 6)": "DecimalType(29, 6)",
+    "DECIMAL(29, 7)": "DecimalType(29, 7)",
+    "DECIMAL(29, 8)": "DecimalType(29, 8)",
+    "DECIMAL(29, 9)": "DecimalType(29, 9)",
+    "DECIMAL(29, 10)": "DecimalType(29, 10)",
+    "DECIMAL(29, 11)": "DecimalType(29, 11)",
+    "DECIMAL(29, 12)": "DecimalType(29, 12)",
+    "DECIMAL(29, 13)": "DecimalType(29, 13)",
+    "DECIMAL(29, 14)": "DecimalType(29, 14)",
+    "DECIMAL(29, 15)": "DecimalType(29, 15)",
+    "DECIMAL(29, 16)": "DecimalType(29, 16)",
+    "DECIMAL(29, 17)": "DecimalType(29, 17)",
+    "DECIMAL(29, 18)": "DecimalType(29, 18)",
+    "DECIMAL(29, 19)": "DecimalType(29, 19)",
+    "DECIMAL(29, 20)": "DecimalType(29, 20)",
+    "DECIMAL(29, 21)": "DecimalType(29, 21)",
+    "DECIMAL(29, 22)": "DecimalType(29, 22)",
+    "DECIMAL(29, 23)": "DecimalType(29, 23)",
+    "DECIMAL(29, 24)": "DecimalType(29, 24)",
+    "DECIMAL(29, 25)": "DecimalType(29, 25)",
+    "DECIMAL(29, 26)": "DecimalType(29, 26)",
+    "DECIMAL(29, 27)": "DecimalType(29, 27)",
+    "DECIMAL(29, 28)": "DecimalType(29, 28)",
+    "DECIMAL(29, 29)": "DecimalType(29, 29)",
+    "DECIMAL(29, 30)": "DecimalType(29, 30)",
+    "DECIMAL(29, 31)": "DecimalType(29, 31)",
+    "DECIMAL(29, 32)": "DecimalType(29, 32)",
+    "DECIMAL(29, 33)": "DecimalType(29, 33)",
+    "DECIMAL(29, 34)": "DecimalType(29, 34)",
+    "DECIMAL(29, 35)": "DecimalType(29, 35)",
+    "DECIMAL(29, 36)": "DecimalType(29, 36)",
+    "DECIMAL(29, 37)": "DecimalType(29, 37)",
+    "DECIMAL(29, 38)": "DecimalType(29, 38)",
+    "DECIMAL(30, 0)": "DecimalType(30, 0)",
+    "DECIMAL(30, 1)": "DecimalType(30, 1)",
+    "DECIMAL(30, 2)": "DecimalType(30, 2)",
+    "DECIMAL(30, 3)": "DecimalType(30, 3)",
+    "DECIMAL(30, 4)": "DecimalType(30, 4)",
+    "DECIMAL(30, 5)": "DecimalType(30, 5)",
+    "DECIMAL(30, 6)": "DecimalType(30, 6)",
+    "DECIMAL(30, 7)": "DecimalType(30, 7)",
+    "DECIMAL(30, 8)": "DecimalType(30, 8)",
+    "DECIMAL(30, 9)": "DecimalType(30, 9)",
+    "DECIMAL(30, 10)": "DecimalType(30, 10)",
+    "DECIMAL(30, 11)": "DecimalType(30, 11)",
+    "DECIMAL(30, 12)": "DecimalType(30, 12)",
+    "DECIMAL(30, 13)": "DecimalType(30, 13)",
+    "DECIMAL(30, 14)": "DecimalType(30, 14)",
+    "DECIMAL(30, 15)": "DecimalType(30, 15)",
+    "DECIMAL(30, 16)": "DecimalType(30, 16)",
+    "DECIMAL(30, 17)": "DecimalType(30, 17)",
+    "DECIMAL(30, 18)": "DecimalType(30, 18)",
+    "DECIMAL(30, 19)": "DecimalType(30, 19)",
+    "DECIMAL(30, 20)": "DecimalType(30, 20)",
+    "DECIMAL(30, 21)": "DecimalType(30, 21)",
+    "DECIMAL(30, 22)": "DecimalType(30, 22)",
+    "DECIMAL(30, 23)": "DecimalType(30, 23)",
+    "DECIMAL(30, 24)": "DecimalType(30, 24)",
+    "DECIMAL(30, 25)": "DecimalType(30, 25)",
+    "DECIMAL(30, 26)": "DecimalType(30, 26)",
+    "DECIMAL(30, 27)": "DecimalType(30, 27)",
+    "DECIMAL(30, 28)": "DecimalType(30, 28)",
+    "DECIMAL(30, 29)": "DecimalType(30, 29)",
+    "DECIMAL(30, 30)": "DecimalType(30, 30)",
+    "DECIMAL(30, 31)": "DecimalType(30, 31)",
+    "DECIMAL(30, 32)": "DecimalType(30, 32)",
+    "DECIMAL(30, 33)": "DecimalType(30, 33)",
+    "DECIMAL(30, 34)": "DecimalType(30, 34)",
+    "DECIMAL(30, 35)": "DecimalType(30, 35)",
+    "DECIMAL(30, 36)": "DecimalType(30, 36)",
+    "DECIMAL(30, 37)": "DecimalType(30, 37)",
+    "DECIMAL(30, 38)": "DecimalType(30, 38)",
+    "DECIMAL(31, 0)": "DecimalType(31, 0)",
+    "DECIMAL(31, 1)": "DecimalType(31, 1)",
+    "DECIMAL(31, 2)": "DecimalType(31, 2)",
+    "DECIMAL(31, 3)": "DecimalType(31, 3)",
+    "DECIMAL(31, 4)": "DecimalType(31, 4)",
+    "DECIMAL(31, 5)": "DecimalType(31, 5)",
+    "DECIMAL(31, 6)": "DecimalType(31, 6)",
+    "DECIMAL(31, 7)": "DecimalType(31, 7)",
+    "DECIMAL(31, 8)": "DecimalType(31, 8)",
+    "DECIMAL(31, 9)": "DecimalType(31, 9)",
+    "DECIMAL(31, 10)": "DecimalType(31, 10)",
+    "DECIMAL(31, 11)": "DecimalType(31, 11)",
+    "DECIMAL(31, 12)": "DecimalType(31, 12)",
+    "DECIMAL(31, 13)": "DecimalType(31, 13)",
+    "DECIMAL(31, 14)": "DecimalType(31, 14)",
+    "DECIMAL(31, 15)": "DecimalType(31, 15)",
+    "DECIMAL(31, 16)": "DecimalType(31, 16)",
+    "DECIMAL(31, 17)": "DecimalType(31, 17)",
+    "DECIMAL(31, 18)": "DecimalType(31, 18)",
+    "DECIMAL(31, 19)": "DecimalType(31, 19)",
+    "DECIMAL(31, 20)": "DecimalType(31, 20)",
+    "DECIMAL(31, 21)": "DecimalType(31, 21)",
+    "DECIMAL(31, 22)": "DecimalType(31, 22)",
+    "DECIMAL(31, 23)": "DecimalType(31, 23)",
+    "DECIMAL(31, 24)": "DecimalType(31, 24)",
+    "DECIMAL(31, 25)": "DecimalType(31, 25)",
+    "DECIMAL(31, 26)": "DecimalType(31, 26)",
+    "DECIMAL(31, 27)": "DecimalType(31, 27)",
+    "DECIMAL(31, 28)": "DecimalType(31, 28)",
+    "DECIMAL(31, 29)": "DecimalType(31, 29)",
+    "DECIMAL(31, 30)": "DecimalType(31, 30)",
+    "DECIMAL(31, 31)": "DecimalType(31, 31)",
+    "DECIMAL(31, 32)": "DecimalType(31, 32)",
+    "DECIMAL(31, 33)": "DecimalType(31, 33)",
+    "DECIMAL(31, 34)": "DecimalType(31, 34)",
+    "DECIMAL(31, 35)": "DecimalType(31, 35)",
+    "DECIMAL(31, 36)": "DecimalType(31, 36)",
+    "DECIMAL(31, 37)": "DecimalType(31, 37)",
+    "DECIMAL(31, 38)": "DecimalType(31, 38)",
+    "DECIMAL(32, 0)": "DecimalType(32, 0)",
+    "DECIMAL(32, 1)": "DecimalType(32, 1)",
+    "DECIMAL(32, 2)": "DecimalType(32, 2)",
+    "DECIMAL(32, 3)": "DecimalType(32, 3)",
+    "DECIMAL(32, 4)": "DecimalType(32, 4)",
+    "DECIMAL(32, 5)": "DecimalType(32, 5)",
+    "DECIMAL(32, 6)": "DecimalType(32, 6)",
+    "DECIMAL(32, 7)": "DecimalType(32, 7)",
+    "DECIMAL(32, 8)": "DecimalType(32, 8)",
+    "DECIMAL(32, 9)": "DecimalType(32, 9)",
+    "DECIMAL(32, 10)": "DecimalType(32, 10)",
+    "DECIMAL(32, 11)": "DecimalType(32, 11)",
+    "DECIMAL(32, 12)": "DecimalType(32, 12)",
+    "DECIMAL(32, 13)": "DecimalType(32, 13)",
+    "DECIMAL(32, 14)": "DecimalType(32, 14)",
+    "DECIMAL(32, 15)": "DecimalType(32, 15)",
+    "DECIMAL(32, 16)": "DecimalType(32, 16)",
+    "DECIMAL(32, 17)": "DecimalType(32, 17)",
+    "DECIMAL(32, 18)": "DecimalType(32, 18)",
+    "DECIMAL(32, 19)": "DecimalType(32, 19)",
+    "DECIMAL(32, 20)": "DecimalType(32, 20)",
+    "DECIMAL(32, 21)": "DecimalType(32, 21)",
+    "DECIMAL(32, 22)": "DecimalType(32, 22)",
+    "DECIMAL(32, 23)": "DecimalType(32, 23)",
+    "DECIMAL(32, 24)": "DecimalType(32, 24)",
+    "DECIMAL(32, 25)": "DecimalType(32, 25)",
+    "DECIMAL(32, 26)": "DecimalType(32, 26)",
+    "DECIMAL(32, 27)": "DecimalType(32, 27)",
+    "DECIMAL(32, 28)": "DecimalType(32, 28)",
+    "DECIMAL(32, 29)": "DecimalType(32, 29)",
+    "DECIMAL(32, 30)": "DecimalType(32, 30)",
+    "DECIMAL(32, 31)": "DecimalType(32, 31)",
+    "DECIMAL(32, 32)": "DecimalType(32, 32)",
+    "DECIMAL(32, 33)": "DecimalType(32, 33)",
+    "DECIMAL(32, 34)": "DecimalType(32, 34)",
+    "DECIMAL(32, 35)": "DecimalType(32, 35)",
+    "DECIMAL(32, 36)": "DecimalType(32, 36)",
+    "DECIMAL(32, 37)": "DecimalType(32, 37)",
+    "DECIMAL(32, 38)": "DecimalType(32, 38)",
+    "DECIMAL(33, 0)": "DecimalType(33, 0)",
+    "DECIMAL(33, 1)": "DecimalType(33, 1)",
+    "DECIMAL(33, 2)": "DecimalType(33, 2)",
+    "DECIMAL(33, 3)": "DecimalType(33, 3)",
+    "DECIMAL(33, 4)": "DecimalType(33, 4)",
+    "DECIMAL(33, 5)": "DecimalType(33, 5)",
+    "DECIMAL(33, 6)": "DecimalType(33, 6)",
+    "DECIMAL(33, 7)": "DecimalType(33, 7)",
+    "DECIMAL(33, 8)": "DecimalType(33, 8)",
+    "DECIMAL(33, 9)": "DecimalType(33, 9)",
+    "DECIMAL(33, 10)": "DecimalType(33, 10)",
+    "DECIMAL(33, 11)": "DecimalType(33, 11)",
+    "DECIMAL(33, 12)": "DecimalType(33, 12)",
+    "DECIMAL(33, 13)": "DecimalType(33, 13)",
+    "DECIMAL(33, 14)": "DecimalType(33, 14)",
+    "DECIMAL(33, 15)": "DecimalType(33, 15)",
+    "DECIMAL(33, 16)": "DecimalType(33, 16)",
+    "DECIMAL(33, 17)": "DecimalType(33, 17)",
+    "DECIMAL(33, 18)": "DecimalType(33, 18)",
+    "DECIMAL(33, 19)": "DecimalType(33, 19)",
+    "DECIMAL(33, 20)": "DecimalType(33, 20)",
+    "DECIMAL(33, 21)": "DecimalType(33, 21)",
+    "DECIMAL(33, 22)": "DecimalType(33, 22)",
+    "DECIMAL(33, 23)": "DecimalType(33, 23)",
+    "DECIMAL(33, 24)": "DecimalType(33, 24)",
+    "DECIMAL(33, 25)": "DecimalType(33, 25)",
+    "DECIMAL(33, 26)": "DecimalType(33, 26)",
+    "DECIMAL(33, 27)": "DecimalType(33, 27)",
+    "DECIMAL(33, 28)": "DecimalType(33, 28)",
+    "DECIMAL(33, 29)": "DecimalType(33, 29)",
+    "DECIMAL(33, 30)": "DecimalType(33, 30)",
+    "DECIMAL(33, 31)": "DecimalType(33, 31)",
+    "DECIMAL(33, 32)": "DecimalType(33, 32)",
+    "DECIMAL(33, 33)": "DecimalType(33, 33)",
+    "DECIMAL(33, 34)": "DecimalType(33, 34)",
+    "DECIMAL(33, 35)": "DecimalType(33, 35)",
+    "DECIMAL(33, 36)": "DecimalType(33, 36)",
+    "DECIMAL(33, 37)": "DecimalType(33, 37)",
+    "DECIMAL(33, 38)": "DecimalType(33, 38)",
+    "DECIMAL(34, 0)": "DecimalType(34, 0)",
+    "DECIMAL(34, 1)": "DecimalType(34, 1)",
+    "DECIMAL(34, 2)": "DecimalType(34, 2)",
+    "DECIMAL(34, 3)": "DecimalType(34, 3)",
+    "DECIMAL(34, 4)": "DecimalType(34, 4)",
+    "DECIMAL(34, 5)": "DecimalType(34, 5)",
+    "DECIMAL(34, 6)": "DecimalType(34, 6)",
+    "DECIMAL(34, 7)": "DecimalType(34, 7)",
+    "DECIMAL(34, 8)": "DecimalType(34, 8)",
+    "DECIMAL(34, 9)": "DecimalType(34, 9)",
+    "DECIMAL(34, 10)": "DecimalType(34, 10)",
+    "DECIMAL(34, 11)": "DecimalType(34, 11)",
+    "DECIMAL(34, 12)": "DecimalType(34, 12)",
+    "DECIMAL(34, 13)": "DecimalType(34, 13)",
+    "DECIMAL(34, 14)": "DecimalType(34, 14)",
+    "DECIMAL(34, 15)": "DecimalType(34, 15)",
+    "DECIMAL(34, 16)": "DecimalType(34, 16)",
+    "DECIMAL(34, 17)": "DecimalType(34, 17)",
+    "DECIMAL(34, 18)": "DecimalType(34, 18)",
+    "DECIMAL(34, 19)": "DecimalType(34, 19)",
+    "DECIMAL(34, 20)": "DecimalType(34, 20)",
+    "DECIMAL(34, 21)": "DecimalType(34, 21)",
+    "DECIMAL(34, 22)": "DecimalType(34, 22)",
+    "DECIMAL(34, 23)": "DecimalType(34, 23)",
+    "DECIMAL(34, 24)": "DecimalType(34, 24)",
+    "DECIMAL(34, 25)": "DecimalType(34, 25)",
+    "DECIMAL(34, 26)": "DecimalType(34, 26)",
+    "DECIMAL(34, 27)": "DecimalType(34, 27)",
+    "DECIMAL(34, 28)": "DecimalType(34, 28)",
+    "DECIMAL(34, 29)": "DecimalType(34, 29)",
+    "DECIMAL(34, 30)": "DecimalType(34, 30)",
+    "DECIMAL(34, 31)": "DecimalType(34, 31)",
+    "DECIMAL(34, 32)": "DecimalType(34, 32)",
+    "DECIMAL(34, 33)": "DecimalType(34, 33)",
+    "DECIMAL(34, 34)": "DecimalType(34, 34)",
+    "DECIMAL(34, 35)": "DecimalType(34, 35)",
+    "DECIMAL(34, 36)": "DecimalType(34, 36)",
+    "DECIMAL(34, 37)": "DecimalType(34, 37)",
+    "DECIMAL(34, 38)": "DecimalType(34, 38)",
+    "DECIMAL(35, 0)": "DecimalType(35, 0)",
+    "DECIMAL(35, 1)": "DecimalType(35, 1)",
+    "DECIMAL(35, 2)": "DecimalType(35, 2)",
+    "DECIMAL(35, 3)": "DecimalType(35, 3)",
+    "DECIMAL(35, 4)": "DecimalType(35, 4)",
+    "DECIMAL(35, 5)": "DecimalType(35, 5)",
+    "DECIMAL(35, 6)": "DecimalType(35, 6)",
+    "DECIMAL(35, 7)": "DecimalType(35, 7)",
+    "DECIMAL(35, 8)": "DecimalType(35, 8)",
+    "DECIMAL(35, 9)": "DecimalType(35, 9)",
+    "DECIMAL(35, 10)": "DecimalType(35, 10)",
+    "DECIMAL(35, 11)": "DecimalType(35, 11)",
+    "DECIMAL(35, 12)": "DecimalType(35, 12)",
+    "DECIMAL(35, 13)": "DecimalType(35, 13)",
+    "DECIMAL(35, 14)": "DecimalType(35, 14)",
+    "DECIMAL(35, 15)": "DecimalType(35, 15)",
+    "DECIMAL(35, 16)": "DecimalType(35, 16)",
+    "DECIMAL(35, 17)": "DecimalType(35, 17)",
+    "DECIMAL(35, 18)": "DecimalType(35, 18)",
+    "DECIMAL(35, 19)": "DecimalType(35, 19)",
+    "DECIMAL(35, 20)": "DecimalType(35, 20)",
+    "DECIMAL(35, 21)": "DecimalType(35, 21)",
+    "DECIMAL(35, 22)": "DecimalType(35, 22)",
+    "DECIMAL(35, 23)": "DecimalType(35, 23)",
+    "DECIMAL(35, 24)": "DecimalType(35, 24)",
+    "DECIMAL(35, 25)": "DecimalType(35, 25)",
+    "DECIMAL(35, 26)": "DecimalType(35, 26)",
+    "DECIMAL(35, 27)": "DecimalType(35, 27)",
+    "DECIMAL(35, 28)": "DecimalType(35, 28)",
+    "DECIMAL(35, 29)": "DecimalType(35, 29)",
+    "DECIMAL(35, 30)": "DecimalType(35, 30)",
+    "DECIMAL(35, 31)": "DecimalType(35, 31)",
+    "DECIMAL(35, 32)": "DecimalType(35, 32)",
+    "DECIMAL(35, 33)": "DecimalType(35, 33)",
+    "DECIMAL(35, 34)": "DecimalType(35, 34)",
+    "DECIMAL(35, 35)": "DecimalType(35, 35)",
+    "DECIMAL(35, 36)": "DecimalType(35, 36)",
+    "DECIMAL(35, 37)": "DecimalType(35, 37)",
+    "DECIMAL(35, 38)": "DecimalType(35, 38)",
+    "DECIMAL(36, 0)": "DecimalType(36, 0)",
+    "DECIMAL(36, 1)": "DecimalType(36, 1)",
+    "DECIMAL(36, 2)": "DecimalType(36, 2)",
+    "DECIMAL(36, 3)": "DecimalType(36, 3)",
+    "DECIMAL(36, 4)": "DecimalType(36, 4)",
+    "DECIMAL(36, 5)": "DecimalType(36, 5)",
+    "DECIMAL(36, 6)": "DecimalType(36, 6)",
+    "DECIMAL(36, 7)": "DecimalType(36, 7)",
+    "DECIMAL(36, 8)": "DecimalType(36, 8)",
+    "DECIMAL(36, 9)": "DecimalType(36, 9)",
+    "DECIMAL(36, 10)": "DecimalType(36, 10)",
+    "DECIMAL(36, 11)": "DecimalType(36, 11)",
+    "DECIMAL(36, 12)": "DecimalType(36, 12)",
+    "DECIMAL(36, 13)": "DecimalType(36, 13)",
+    "DECIMAL(36, 14)": "DecimalType(36, 14)",
+    "DECIMAL(36, 15)": "DecimalType(36, 15)",
+    "DECIMAL(36, 16)": "DecimalType(36, 16)",
+    "DECIMAL(36, 17)": "DecimalType(36, 17)",
+    "DECIMAL(36, 18)": "DecimalType(36, 18)",
+    "DECIMAL(36, 19)": "DecimalType(36, 19)",
+    "DECIMAL(36, 20)": "DecimalType(36, 20)",
+    "DECIMAL(36, 21)": "DecimalType(36, 21)",
+    "DECIMAL(36, 22)": "DecimalType(36, 22)",
+    "DECIMAL(36, 23)": "DecimalType(36, 23)",
+    "DECIMAL(36, 24)": "DecimalType(36, 24)",
+    "DECIMAL(36, 25)": "DecimalType(36, 25)",
+    "DECIMAL(36, 26)": "DecimalType(36, 26)",
+    "DECIMAL(36, 27)": "DecimalType(36, 27)",
+    "DECIMAL(36, 28)": "DecimalType(36, 28)",
+    "DECIMAL(36, 29)": "DecimalType(36, 29)",
+    "DECIMAL(36, 30)": "DecimalType(36, 30)",
+    "DECIMAL(36, 31)": "DecimalType(36, 31)",
+    "DECIMAL(36, 32)": "DecimalType(36, 32)",
+    "DECIMAL(36, 33)": "DecimalType(36, 33)",
+    "DECIMAL(36, 34)": "DecimalType(36, 34)",
+    "DECIMAL(36, 35)": "DecimalType(36, 35)",
+    "DECIMAL(36, 36)": "DecimalType(36, 36)",
+    "DECIMAL(36, 37)": "DecimalType(36, 37)",
+    "DECIMAL(36, 38)": "DecimalType(36, 38)",
+    "DECIMAL(37, 0)": "DecimalType(37, 0)",
+    "DECIMAL(37, 1)": "DecimalType(37, 1)",
+    "DECIMAL(37, 2)": "DecimalType(37, 2)",
+    "DECIMAL(37, 3)": "DecimalType(37, 3)",
+    "DECIMAL(37, 4)": "DecimalType(37, 4)",
+    "DECIMAL(37, 5)": "DecimalType(37, 5)",
+    "DECIMAL(37, 6)": "DecimalType(37, 6)",
+    "DECIMAL(37, 7)": "DecimalType(37, 7)",
+    "DECIMAL(37, 8)": "DecimalType(37, 8)",
+    "DECIMAL(37, 9)": "DecimalType(37, 9)",
+    "DECIMAL(37, 10)": "DecimalType(37, 10)",
+    "DECIMAL(37, 11)": "DecimalType(37, 11)",
+    "DECIMAL(37, 12)": "DecimalType(37, 12)",
+    "DECIMAL(37, 13)": "DecimalType(37, 13)",
+    "DECIMAL(37, 14)": "DecimalType(37, 14)",
+    "DECIMAL(37, 15)": "DecimalType(37, 15)",
+    "DECIMAL(37, 16)": "DecimalType(37, 16)",
+    "DECIMAL(37, 17)": "DecimalType(37, 17)",
+    "DECIMAL(37, 18)": "DecimalType(37, 18)",
+    "DECIMAL(37, 19)": "DecimalType(37, 19)",
+    "DECIMAL(37, 20)": "DecimalType(37, 20)",
+    "DECIMAL(37, 21)": "DecimalType(37, 21)",
+    "DECIMAL(37, 22)": "DecimalType(37, 22)",
+    "DECIMAL(37, 23)": "DecimalType(37, 23)",
+    "DECIMAL(37, 24)": "DecimalType(37, 24)",
+    "DECIMAL(37, 25)": "DecimalType(37, 25)",
+    "DECIMAL(37, 26)": "DecimalType(37, 26)",
+    "DECIMAL(37, 27)": "DecimalType(37, 27)",
+    "DECIMAL(37, 28)": "DecimalType(37, 28)",
+    "DECIMAL(37, 29)": "DecimalType(37, 29)",
+    "DECIMAL(37, 30)": "DecimalType(37, 30)",
+    "DECIMAL(37, 31)": "DecimalType(37, 31)",
+    "DECIMAL(37, 32)": "DecimalType(37, 32)",
+    "DECIMAL(37, 33)": "DecimalType(37, 33)",
+    "DECIMAL(37, 34)": "DecimalType(37, 34)",
+    "DECIMAL(37, 35)": "DecimalType(37, 35)",
+    "DECIMAL(37, 36)": "DecimalType(37, 36)",
+    "DECIMAL(37, 37)": "DecimalType(37, 37)",
+    "DECIMAL(37, 38)": "DecimalType(37, 38)",
+    "DECIMAL(38, 0)": "DecimalType(38, 0)",
+    "DECIMAL(38, 1)": "DecimalType(38, 1)",
+    "DECIMAL(38, 2)": "DecimalType(38, 2)",
+    "DECIMAL(38, 3)": "DecimalType(38, 3)",
+    "DECIMAL(38, 4)": "DecimalType(38, 4)",
+    "DECIMAL(38, 5)": "DecimalType(38, 5)",
+    "DECIMAL(38, 6)": "DecimalType(38, 6)",
+    "DECIMAL(38, 7)": "DecimalType(38, 7)",
+    "DECIMAL(38, 8)": "DecimalType(38, 8)",
+    "DECIMAL(38, 9)": "DecimalType(38, 9)",
+    "DECIMAL(38, 10)": "DecimalType(38, 10)",
+    "DECIMAL(38, 11)": "DecimalType(38, 11)",
+    "DECIMAL(38, 12)": "DecimalType(38, 12)",
+    "DECIMAL(38, 13)": "DecimalType(38, 13)",
+    "DECIMAL(38, 14)": "DecimalType(38, 14)",
+    "DECIMAL(38, 15)": "DecimalType(38, 15)",
+    "DECIMAL(38, 16)": "DecimalType(38, 16)",
+    "DECIMAL(38, 17)": "DecimalType(38, 17)",
+    "DECIMAL(38, 18)": "DecimalType(38, 18)",
+    "DECIMAL(38, 19)": "DecimalType(38, 19)",
+    "DECIMAL(38, 20)": "DecimalType(38, 20)",
+    "DECIMAL(38, 21)": "DecimalType(38, 21)",
+    "DECIMAL(38, 22)": "DecimalType(38, 22)",
+    "DECIMAL(38, 23)": "DecimalType(38, 23)",
+    "DECIMAL(38, 24)": "DecimalType(38, 24)",
+    "DECIMAL(38, 25)": "DecimalType(38, 25)",
+    "DECIMAL(38, 26)": "DecimalType(38, 26)",
+    "DECIMAL(38, 27)": "DecimalType(38, 27)",
+    "DECIMAL(38, 28)": "DecimalType(38, 28)",
+    "DECIMAL(38, 29)": "DecimalType(38, 29)",
+    "DECIMAL(38, 30)": "DecimalType(38, 30)",
+    "DECIMAL(38, 31)": "DecimalType(38, 31)",
+    "DECIMAL(38, 32)": "DecimalType(38, 32)",
+    "DECIMAL(38, 33)": "DecimalType(38, 33)",
+    "DECIMAL(38, 34)": "DecimalType(38, 34)",
+    "DECIMAL(38, 35)": "DecimalType(38, 35)",
+    "DECIMAL(38, 36)": "DecimalType(38, 36)",
+    "DECIMAL(38, 37)": "DecimalType(38, 37)",
+    "DECIMAL(38, 38)": "DecimalType(38, 38)",
+    "REAL": "FloatType()",
+    "DOUBLE PRECISION": "DoubleType()",
+    "SMALLINT": "ShortType()",
+    "INTEGER": "IntegerType()",
+    "BIGINT": "LongType()",
+    "MONEY": "StringType()",
+    "INT4RANGE": "StringType()",
+    "INT8RANGE": "StringType()",
+    "NUMRANGE": "StringType()",
+    "INT2VECTOR": "StringType()",
+    "DATE": "DateType()",
+    "TIME": "TimestampType()",
+    "TIME WITH TIME ZONE": "TimestampType()",
+    "TIME(0) WITH TIME ZONE": "TimestampType()",
+    "TIME(1) WITH TIME ZONE": "TimestampType()",
+    "TIME(2) WITH TIME ZONE": "TimestampType()",
+    "TIME(3) WITH TIME ZONE": "TimestampType()",
+    "TIME(4) WITH TIME ZONE": "TimestampType()",
+    "TIME(5) WITH TIME ZONE": "TimestampType()",
+    "TIME(6) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP": "TimestampType()",
+    "TIMESTAMP WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(0) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(1) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(2) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(3) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(4) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(5) WITH TIME ZONE": "TimestampType()",
+    "TIMESTAMP(6) WITH TIME ZONE": "TimestampType()",
+    "INTERVAL": "StringType()",
+    "DATERANGE": "StringType()",
+    "TSRANGE": "StringType()",
+    "TSTZRANGE": "StringType()",
+    "CHARACTER": "StringType()",
+    "CHARACTER VARYING": "StringType()",
+    "TEXT": "StringType()",
+    "JSON": "StringType()",
+    "JSONB": "StringType()",
+    "XML": "StringType()",
+    "TSVECTOR": "StringType()",
+    "TSQUERY": "StringType()",
+}
