--- ddlsparkconverter/db_from_spark_datatype_conversion_info/oracle.py
@@ -1,1529 +1,0 @@
-datatypes = {
-    "DecimalType(0, 0)": "NUMBER(0, 0)",
-    "DecimalType(0, 1)": "NUMBER(0, 1)",
-    "DecimalType(0, 2)": "NUMBER(0, 2)",
-    "DecimalType(0, 3)": "NUMBER(0, 3)",
-    "DecimalType(0, 4)": "NUMBER(0, 4)",
-    "DecimalType(0, 5)": "NUMBER(0, 5)",
-    "DecimalType(0, 6)": "NUMBER(0, 6)",
-    "DecimalType(0, 7)": "NUMBER(0, 7)",
-    "DecimalType(0, 8)": "NUMBER(0, 8)",
-    "DecimalType(0, 9)": "NUMBER(0, 9)",
-    "DecimalType(0, 10)": "NUMBER(0, 10)",
-    "DecimalType(0, 11)": "NUMBER(0, 11)",
-    "DecimalType(0, 12)": "NUMBER(0, 12)",
-    "DecimalType(0, 13)": "NUMBER(0, 13)",
-    "DecimalType(0, 14)": "NUMBER(0, 14)",
-    "DecimalType(0, 15)": "NUMBER(0, 15)",
-    "DecimalType(0, 16)": "NUMBER(0, 16)",
-    "DecimalType(0, 17)": "NUMBER(0, 17)",
-    "DecimalType(0, 18)": "NUMBER(0, 18)",
-    "DecimalType(0, 19)": "NUMBER(0, 19)",
-    "DecimalType(0, 20)": "NUMBER(0, 20)",
-    "DecimalType(0, 21)": "NUMBER(0, 21)",
-    "DecimalType(0, 22)": "NUMBER(0, 22)",
-    "DecimalType(0, 23)": "NUMBER(0, 23)",
-    "DecimalType(0, 24)": "NUMBER(0, 24)",
-    "DecimalType(0, 25)": "NUMBER(0, 25)",
-    "DecimalType(0, 26)": "NUMBER(0, 26)",
-    "DecimalType(0, 27)": "NUMBER(0, 27)",
-    "DecimalType(0, 28)": "NUMBER(0, 28)",
-    "DecimalType(0, 29)": "NUMBER(0, 29)",
-    "DecimalType(0, 30)": "NUMBER(0, 30)",
-    "DecimalType(0, 31)": "NUMBER(0, 31)",
-    "DecimalType(0, 32)": "NUMBER(0, 32)",
-    "DecimalType(0, 33)": "NUMBER(0, 33)",
-    "DecimalType(0, 34)": "NUMBER(0, 34)",
-    "DecimalType(0, 35)": "NUMBER(0, 35)",
-    "DecimalType(0, 36)": "NUMBER(0, 36)",
-    "DecimalType(0, 37)": "NUMBER(0, 37)",
-    "DecimalType(0, 38)": "NUMBER(0, 38)",
-    "DecimalType(1, 0)": "NUMBER(1, 0)",
-    "DecimalType(1, 1)": "NUMBER(1, 1)",
-    "DecimalType(1, 2)": "NUMBER(1, 2)",
-    "DecimalType(1, 3)": "NUMBER(1, 3)",
-    "DecimalType(1, 4)": "NUMBER(1, 4)",
-    "DecimalType(1, 5)": "NUMBER(1, 5)",
-    "DecimalType(1, 6)": "NUMBER(1, 6)",
-    "DecimalType(1, 7)": "NUMBER(1, 7)",
-    "DecimalType(1, 8)": "NUMBER(1, 8)",
-    "DecimalType(1, 9)": "NUMBER(1, 9)",
-    "DecimalType(1, 10)": "NUMBER(1, 10)",
-    "DecimalType(1, 11)": "NUMBER(1, 11)",
-    "DecimalType(1, 12)": "NUMBER(1, 12)",
-    "DecimalType(1, 13)": "NUMBER(1, 13)",
-    "DecimalType(1, 14)": "NUMBER(1, 14)",
-    "DecimalType(1, 15)": "NUMBER(1, 15)",
-    "DecimalType(1, 16)": "NUMBER(1, 16)",
-    "DecimalType(1, 17)": "NUMBER(1, 17)",
-    "DecimalType(1, 18)": "NUMBER(1, 18)",
-    "DecimalType(1, 19)": "NUMBER(1, 19)",
-    "DecimalType(1, 20)": "NUMBER(1, 20)",
-    "DecimalType(1, 21)": "NUMBER(1, 21)",
-    "DecimalType(1, 22)": "NUMBER(1, 22)",
-    "DecimalType(1, 23)": "NUMBER(1, 23)",
-    "DecimalType(1, 24)": "NUMBER(1, 24)",
-    "DecimalType(1, 25)": "NUMBER(1, 25)",
-    "DecimalType(1, 26)": "NUMBER(1, 26)",
-    "DecimalType(1, 27)": "NUMBER(1, 27)",
-    "DecimalType(1, 28)": "NUMBER(1, 28)",
-    "DecimalType(1, 29)": "NUMBER(1, 29)",
-    "DecimalType(1, 30)": "NUMBER(1, 30)",
-    "DecimalType(1, 31)": "NUMBER(1, 31)",
-    "DecimalType(1, 32)": "NUMBER(1, 32)",
-    "DecimalType(1, 33)": "NUMBER(1, 33)",
-    "DecimalType(1, 34)": "NUMBER(1, 34)",
-    "DecimalType(1, 35)": "NUMBER(1, 35)",
-    "DecimalType(1, 36)": "NUMBER(1, 36)",
-    "DecimalType(1, 37)": "NUMBER(1, 37)",
-    "DecimalType(1, 38)": "NUMBER(1, 38)",
-    "DecimalType(2, 0)": "NUMBER(2, 0)",
-    "DecimalType(2, 1)": "NUMBER(2, 1)",
-    "DecimalType(2, 2)": "NUMBER(2, 2)",
-    "DecimalType(2, 3)": "NUMBER(2, 3)",
-    "DecimalType(2, 4)": "NUMBER(2, 4)",
-    "DecimalType(2, 5)": "NUMBER(2, 5)",
-    "DecimalType(2, 6)": "NUMBER(2, 6)",
-    "DecimalType(2, 7)": "NUMBER(2, 7)",
-    "DecimalType(2, 8)": "NUMBER(2, 8)",
-    "DecimalType(2, 9)": "NUMBER(2, 9)",
-    "DecimalType(2, 10)": "NUMBER(2, 10)",
-    "DecimalType(2, 11)": "NUMBER(2, 11)",
-    "DecimalType(2, 12)": "NUMBER(2, 12)",
-    "DecimalType(2, 13)": "NUMBER(2, 13)",
-    "DecimalType(2, 14)": "NUMBER(2, 14)",
-    "DecimalType(2, 15)": "NUMBER(2, 15)",
-    "DecimalType(2, 16)": "NUMBER(2, 16)",
-    "DecimalType(2, 17)": "NUMBER(2, 17)",
-    "DecimalType(2, 18)": "NUMBER(2, 18)",
-    "DecimalType(2, 19)": "NUMBER(2, 19)",
-    "DecimalType(2, 20)": "NUMBER(2, 20)",
-    "DecimalType(2, 21)": "NUMBER(2, 21)",
-    "DecimalType(2, 22)": "NUMBER(2, 22)",
-    "DecimalType(2, 23)": "NUMBER(2, 23)",
-    "DecimalType(2, 24)": "NUMBER(2, 24)",
-    "DecimalType(2, 25)": "NUMBER(2, 25)",
-    "DecimalType(2, 26)": "NUMBER(2, 26)",
-    "DecimalType(2, 27)": "NUMBER(2, 27)",
-    "DecimalType(2, 28)": "NUMBER(2, 28)",
-    "DecimalType(2, 29)": "NUMBER(2, 29)",
-    "DecimalType(2, 30)": "NUMBER(2, 30)",
-    "DecimalType(2, 31)": "NUMBER(2, 31)",
-    "DecimalType(2, 32)": "NUMBER(2, 32)",
-    "DecimalType(2, 33)": "NUMBER(2, 33)",
-    "DecimalType(2, 34)": "NUMBER(2, 34)",
-    "DecimalType(2, 35)": "NUMBER(2, 35)",
-    "DecimalType(2, 36)": "NUMBER(2, 36)",
-    "DecimalType(2, 37)": "NUMBER(2, 37)",
-    "DecimalType(2, 38)": "NUMBER(2, 38)",
-    "DecimalType(3, 0)": "NUMBER(3, 0)",
-    "DecimalType(3, 1)": "NUMBER(3, 1)",
-    "DecimalType(3, 2)": "NUMBER(3, 2)",
-    "DecimalType(3, 3)": "NUMBER(3, 3)",
-    "DecimalType(3, 4)": "NUMBER(3, 4)",
-    "DecimalType(3, 5)": "NUMBER(3, 5)",
-    "DecimalType(3, 6)": "NUMBER(3, 6)",
-    "DecimalType(3, 7)": "NUMBER(3, 7)",
-    "DecimalType(3, 8)": "NUMBER(3, 8)",
-    "DecimalType(3, 9)": "NUMBER(3, 9)",
-    "DecimalType(3, 10)": "NUMBER(3, 10)",
-    "DecimalType(3, 11)": "NUMBER(3, 11)",
-    "DecimalType(3, 12)": "NUMBER(3, 12)",
-    "DecimalType(3, 13)": "NUMBER(3, 13)",
-    "DecimalType(3, 14)": "NUMBER(3, 14)",
-    "DecimalType(3, 15)": "NUMBER(3, 15)",
-    "DecimalType(3, 16)": "NUMBER(3, 16)",
-    "DecimalType(3, 17)": "NUMBER(3, 17)",
-    "DecimalType(3, 18)": "NUMBER(3, 18)",
-    "DecimalType(3, 19)": "NUMBER(3, 19)",
-    "DecimalType(3, 20)": "NUMBER(3, 20)",
-    "DecimalType(3, 21)": "NUMBER(3, 21)",
-    "DecimalType(3, 22)": "NUMBER(3, 22)",
-    "DecimalType(3, 23)": "NUMBER(3, 23)",
-    "DecimalType(3, 24)": "NUMBER(3, 24)",
-    "DecimalType(3, 25)": "NUMBER(3, 25)",
-    "DecimalType(3, 26)": "NUMBER(3, 26)",
-    "DecimalType(3, 27)": "NUMBER(3, 27)",
-    "DecimalType(3, 28)": "NUMBER(3, 28)",
-    "DecimalType(3, 29)": "NUMBER(3, 29)",
-    "DecimalType(3, 30)": "NUMBER(3, 30)",
-    "DecimalType(3, 31)": "NUMBER(3, 31)",
-    "DecimalType(3, 32)": "NUMBER(3, 32)",
-    "DecimalType(3, 33)": "NUMBER(3, 33)",
-    "DecimalType(3, 34)": "NUMBER(3, 34)",
-    "DecimalType(3, 35)": "NUMBER(3, 35)",
-    "DecimalType(3, 36)": "NUMBER(3, 36)",
-    "DecimalType(3, 37)": "NUMBER(3, 37)",
-    "DecimalType(3, 38)": "NUMBER(3, 38)",
-    "DecimalType(4, 0)": "NUMBER(4, 0)",
-    "DecimalType(4, 1)": "NUMBER(4, 1)",
-    "DecimalType(4, 2)": "NUMBER(4, 2)",
-    "DecimalType(4, 3)": "NUMBER(4, 3)",
-    "DecimalType(4, 4)": "NUMBER(4, 4)",
-    "DecimalType(4, 5)": "NUMBER(4, 5)",
-    "DecimalType(4, 6)": "NUMBER(4, 6)",
-    "DecimalType(4, 7)": "NUMBER(4, 7)",
-    "DecimalType(4, 8)": "NUMBER(4, 8)",
-    "DecimalType(4, 9)": "NUMBER(4, 9)",
-    "DecimalType(4, 10)": "NUMBER(4, 10)",
-    "DecimalType(4, 11)": "NUMBER(4, 11)",
-    "DecimalType(4, 12)": "NUMBER(4, 12)",
-    "DecimalType(4, 13)": "NUMBER(4, 13)",
-    "DecimalType(4, 14)": "NUMBER(4, 14)",
-    "DecimalType(4, 15)": "NUMBER(4, 15)",
-    "DecimalType(4, 16)": "NUMBER(4, 16)",
-    "DecimalType(4, 17)": "NUMBER(4, 17)",
-    "DecimalType(4, 18)": "NUMBER(4, 18)",
-    "DecimalType(4, 19)": "NUMBER(4, 19)",
-    "DecimalType(4, 20)": "NUMBER(4, 20)",
-    "DecimalType(4, 21)": "NUMBER(4, 21)",
-    "DecimalType(4, 22)": "NUMBER(4, 22)",
-    "DecimalType(4, 23)": "NUMBER(4, 23)",
-    "DecimalType(4, 24)": "NUMBER(4, 24)",
-    "DecimalType(4, 25)": "NUMBER(4, 25)",
-    "DecimalType(4, 26)": "NUMBER(4, 26)",
-    "DecimalType(4, 27)": "NUMBER(4, 27)",
-    "DecimalType(4, 28)": "NUMBER(4, 28)",
-    "DecimalType(4, 29)": "NUMBER(4, 29)",
-    "DecimalType(4, 30)": "NUMBER(4, 30)",
-    "DecimalType(4, 31)": "NUMBER(4, 31)",
-    "DecimalType(4, 32)": "NUMBER(4, 32)",
-    "DecimalType(4, 33)": "NUMBER(4, 33)",
-    "DecimalType(4, 34)": "NUMBER(4, 34)",
-    "DecimalType(4, 35)": "NUMBER(4, 35)",
-    "DecimalType(4, 36)": "NUMBER(4, 36)",
-    "DecimalType(4, 37)": "NUMBER(4, 37)",
-    "DecimalType(4, 38)": "NUMBER(4, 38)",
-    "DecimalType(5, 0)": "NUMBER(5, 0)",
-    "DecimalType(5, 1)": "NUMBER(5, 1)",
-    "DecimalType(5, 2)": "NUMBER(5, 2)",
-    "DecimalType(5, 3)": "NUMBER(5, 3)",
-    "DecimalType(5, 4)": "NUMBER(5, 4)",
-    "DecimalType(5, 5)": "NUMBER(5, 5)",
-    "DecimalType(5, 6)": "NUMBER(5, 6)",
-    "DecimalType(5, 7)": "NUMBER(5, 7)",
-    "DecimalType(5, 8)": "NUMBER(5, 8)",
-    "DecimalType(5, 9)": "NUMBER(5, 9)",
-    "DecimalType(5, 10)": "NUMBER(5, 10)",
-    "DecimalType(5, 11)": "NUMBER(5, 11)",
-    "DecimalType(5, 12)": "NUMBER(5, 12)",
-    "DecimalType(5, 13)": "NUMBER(5, 13)",
-    "DecimalType(5, 14)": "NUMBER(5, 14)",
-    "DecimalType(5, 15)": "NUMBER(5, 15)",
-    "DecimalType(5, 16)": "NUMBER(5, 16)",
-    "DecimalType(5, 17)": "NUMBER(5, 17)",
-    "DecimalType(5, 18)": "NUMBER(5, 18)",
-    "DecimalType(5, 19)": "NUMBER(5, 19)",
-    "DecimalType(5, 20)": "NUMBER(5, 20)",
-    "DecimalType(5, 21)": "NUMBER(5, 21)",
-    "DecimalType(5, 22)": "NUMBER(5, 22)",
-    "DecimalType(5, 23)": "NUMBER(5, 23)",
-    "DecimalType(5, 24)": "NUMBER(5, 24)",
-    "DecimalType(5, 25)": "NUMBER(5, 25)",
-    "DecimalType(5, 26)": "NUMBER(5, 26)",
-    "DecimalType(5, 27)": "NUMBER(5, 27)",
-    "DecimalType(5, 28)": "NUMBER(5, 28)",
-    "DecimalType(5, 29)": "NUMBER(5, 29)",
-    "DecimalType(5, 30)": "NUMBER(5, 30)",
-    "DecimalType(5, 31)": "NUMBER(5, 31)",
-    "DecimalType(5, 32)": "NUMBER(5, 32)",
-    "DecimalType(5, 33)": "NUMBER(5, 33)",
-    "DecimalType(5, 34)": "NUMBER(5, 34)",
-    "DecimalType(5, 35)": "NUMBER(5, 35)",
-    "DecimalType(5, 36)": "NUMBER(5, 36)",
-    "DecimalType(5, 37)": "NUMBER(5, 37)",
-    "DecimalType(5, 38)": "NUMBER(5, 38)",
-    "DecimalType(6, 0)": "NUMBER(6, 0)",
-    "DecimalType(6, 1)": "NUMBER(6, 1)",
-    "DecimalType(6, 2)": "NUMBER(6, 2)",
-    "DecimalType(6, 3)": "NUMBER(6, 3)",
-    "DecimalType(6, 4)": "NUMBER(6, 4)",
-    "DecimalType(6, 5)": "NUMBER(6, 5)",
-    "DecimalType(6, 6)": "NUMBER(6, 6)",
-    "DecimalType(6, 7)": "NUMBER(6, 7)",
-    "DecimalType(6, 8)": "NUMBER(6, 8)",
-    "DecimalType(6, 9)": "NUMBER(6, 9)",
-    "DecimalType(6, 10)": "NUMBER(6, 10)",
-    "DecimalType(6, 11)": "NUMBER(6, 11)",
-    "DecimalType(6, 12)": "NUMBER(6, 12)",
-    "DecimalType(6, 13)": "NUMBER(6, 13)",
-    "DecimalType(6, 14)": "NUMBER(6, 14)",
-    "DecimalType(6, 15)": "NUMBER(6, 15)",
-    "DecimalType(6, 16)": "NUMBER(6, 16)",
-    "DecimalType(6, 17)": "NUMBER(6, 17)",
-    "DecimalType(6, 18)": "NUMBER(6, 18)",
-    "DecimalType(6, 19)": "NUMBER(6, 19)",
-    "DecimalType(6, 20)": "NUMBER(6, 20)",
-    "DecimalType(6, 21)": "NUMBER(6, 21)",
-    "DecimalType(6, 22)": "NUMBER(6, 22)",
-    "DecimalType(6, 23)": "NUMBER(6, 23)",
-    "DecimalType(6, 24)": "NUMBER(6, 24)",
-    "DecimalType(6, 25)": "NUMBER(6, 25)",
-    "DecimalType(6, 26)": "NUMBER(6, 26)",
-    "DecimalType(6, 27)": "NUMBER(6, 27)",
-    "DecimalType(6, 28)": "NUMBER(6, 28)",
-    "DecimalType(6, 29)": "NUMBER(6, 29)",
-    "DecimalType(6, 30)": "NUMBER(6, 30)",
-    "DecimalType(6, 31)": "NUMBER(6, 31)",
-    "DecimalType(6, 32)": "NUMBER(6, 32)",
-    "DecimalType(6, 33)": "NUMBER(6, 33)",
-    "DecimalType(6, 34)": "NUMBER(6, 34)",
-    "DecimalType(6, 35)": "NUMBER(6, 35)",
-    "DecimalType(6, 36)": "NUMBER(6, 36)",
-    "DecimalType(6, 37)": "NUMBER(6, 37)",
-    "DecimalType(6, 38)": "NUMBER(6, 38)",
-    "DecimalType(7, 0)": "NUMBER(7, 0)",
-    "DecimalType(7, 1)": "NUMBER(7, 1)",
-    "DecimalType(7, 2)": "NUMBER(7, 2)",
-    "DecimalType(7, 3)": "NUMBER(7, 3)",
-    "DecimalType(7, 4)": "NUMBER(7, 4)",
-    "DecimalType(7, 5)": "NUMBER(7, 5)",
-    "DecimalType(7, 6)": "NUMBER(7, 6)",
-    "DecimalType(7, 7)": "NUMBER(7, 7)",
-    "DecimalType(7, 8)": "NUMBER(7, 8)",
-    "DecimalType(7, 9)": "NUMBER(7, 9)",
-    "DecimalType(7, 10)": "NUMBER(7, 10)",
-    "DecimalType(7, 11)": "NUMBER(7, 11)",
-    "DecimalType(7, 12)": "NUMBER(7, 12)",
-    "DecimalType(7, 13)": "NUMBER(7, 13)",
-    "DecimalType(7, 14)": "NUMBER(7, 14)",
-    "DecimalType(7, 15)": "NUMBER(7, 15)",
-    "DecimalType(7, 16)": "NUMBER(7, 16)",
-    "DecimalType(7, 17)": "NUMBER(7, 17)",
-    "DecimalType(7, 18)": "NUMBER(7, 18)",
-    "DecimalType(7, 19)": "NUMBER(7, 19)",
-    "DecimalType(7, 20)": "NUMBER(7, 20)",
-    "DecimalType(7, 21)": "NUMBER(7, 21)",
-    "DecimalType(7, 22)": "NUMBER(7, 22)",
-    "DecimalType(7, 23)": "NUMBER(7, 23)",
-    "DecimalType(7, 24)": "NUMBER(7, 24)",
-    "DecimalType(7, 25)": "NUMBER(7, 25)",
-    "DecimalType(7, 26)": "NUMBER(7, 26)",
-    "DecimalType(7, 27)": "NUMBER(7, 27)",
-    "DecimalType(7, 28)": "NUMBER(7, 28)",
-    "DecimalType(7, 29)": "NUMBER(7, 29)",
-    "DecimalType(7, 30)": "NUMBER(7, 30)",
-    "DecimalType(7, 31)": "NUMBER(7, 31)",
-    "DecimalType(7, 32)": "NUMBER(7, 32)",
-    "DecimalType(7, 33)": "NUMBER(7, 33)",
-    "DecimalType(7, 34)": "NUMBER(7, 34)",
-    "DecimalType(7, 35)": "NUMBER(7, 35)",
-    "DecimalType(7, 36)": "NUMBER(7, 36)",
-    "DecimalType(7, 37)": "NUMBER(7, 37)",
-    "DecimalType(7, 38)": "NUMBER(7, 38)",
-    "DecimalType(8, 0)": "NUMBER(8, 0)",
-    "DecimalType(8, 1)": "NUMBER(8, 1)",
-    "DecimalType(8, 2)": "NUMBER(8, 2)",
-    "DecimalType(8, 3)": "NUMBER(8, 3)",
-    "DecimalType(8, 4)": "NUMBER(8, 4)",
-    "DecimalType(8, 5)": "NUMBER(8, 5)",
-    "DecimalType(8, 6)": "NUMBER(8, 6)",
-    "DecimalType(8, 7)": "NUMBER(8, 7)",
-    "DecimalType(8, 8)": "NUMBER(8, 8)",
-    "DecimalType(8, 9)": "NUMBER(8, 9)",
-    "DecimalType(8, 10)": "NUMBER(8, 10)",
-    "DecimalType(8, 11)": "NUMBER(8, 11)",
-    "DecimalType(8, 12)": "NUMBER(8, 12)",
-    "DecimalType(8, 13)": "NUMBER(8, 13)",
-    "DecimalType(8, 14)": "NUMBER(8, 14)",
-    "DecimalType(8, 15)": "NUMBER(8, 15)",
-    "DecimalType(8, 16)": "NUMBER(8, 16)",
-    "DecimalType(8, 17)": "NUMBER(8, 17)",
-    "DecimalType(8, 18)": "NUMBER(8, 18)",
-    "DecimalType(8, 19)": "NUMBER(8, 19)",
-    "DecimalType(8, 20)": "NUMBER(8, 20)",
-    "DecimalType(8, 21)": "NUMBER(8, 21)",
-    "DecimalType(8, 22)": "NUMBER(8, 22)",
-    "DecimalType(8, 23)": "NUMBER(8, 23)",
-    "DecimalType(8, 24)": "NUMBER(8, 24)",
-    "DecimalType(8, 25)": "NUMBER(8, 25)",
-    "DecimalType(8, 26)": "NUMBER(8, 26)",
-    "DecimalType(8, 27)": "NUMBER(8, 27)",
-    "DecimalType(8, 28)": "NUMBER(8, 28)",
-    "DecimalType(8, 29)": "NUMBER(8, 29)",
-    "DecimalType(8, 30)": "NUMBER(8, 30)",
-    "DecimalType(8, 31)": "NUMBER(8, 31)",
-    "DecimalType(8, 32)": "NUMBER(8, 32)",
-    "DecimalType(8, 33)": "NUMBER(8, 33)",
-    "DecimalType(8, 34)": "NUMBER(8, 34)",
-    "DecimalType(8, 35)": "NUMBER(8, 35)",
-    "DecimalType(8, 36)": "NUMBER(8, 36)",
-    "DecimalType(8, 37)": "NUMBER(8, 37)",
-    "DecimalType(8, 38)": "NUMBER(8, 38)",
-    "DecimalType(9, 0)": "NUMBER(9, 0)",
-    "DecimalType(9, 1)": "NUMBER(9, 1)",
-    "DecimalType(9, 2)": "NUMBER(9, 2)",
-    "DecimalType(9, 3)": "NUMBER(9, 3)",
-    "DecimalType(9, 4)": "NUMBER(9, 4)",
-    "DecimalType(9, 5)": "NUMBER(9, 5)",
-    "DecimalType(9, 6)": "NUMBER(9, 6)",
-    "DecimalType(9, 7)": "NUMBER(9, 7)",
-    "DecimalType(9, 8)": "NUMBER(9, 8)",
-    "DecimalType(9, 9)": "NUMBER(9, 9)",
-    "DecimalType(9, 10)": "NUMBER(9, 10)",
-    "DecimalType(9, 11)": "NUMBER(9, 11)",
-    "DecimalType(9, 12)": "NUMBER(9, 12)",
-    "DecimalType(9, 13)": "NUMBER(9, 13)",
-    "DecimalType(9, 14)": "NUMBER(9, 14)",
-    "DecimalType(9, 15)": "NUMBER(9, 15)",
-    "DecimalType(9, 16)": "NUMBER(9, 16)",
-    "DecimalType(9, 17)": "NUMBER(9, 17)",
-    "DecimalType(9, 18)": "NUMBER(9, 18)",
-    "DecimalType(9, 19)": "NUMBER(9, 19)",
-    "DecimalType(9, 20)": "NUMBER(9, 20)",
-    "DecimalType(9, 21)": "NUMBER(9, 21)",
-    "DecimalType(9, 22)": "NUMBER(9, 22)",
-    "DecimalType(9, 23)": "NUMBER(9, 23)",
-    "DecimalType(9, 24)": "NUMBER(9, 24)",
-    "DecimalType(9, 25)": "NUMBER(9, 25)",
-    "DecimalType(9, 26)": "NUMBER(9, 26)",
-    "DecimalType(9, 27)": "NUMBER(9, 27)",
-    "DecimalType(9, 28)": "NUMBER(9, 28)",
-    "DecimalType(9, 29)": "NUMBER(9, 29)",
-    "DecimalType(9, 30)": "NUMBER(9, 30)",
-    "DecimalType(9, 31)": "NUMBER(9, 31)",
-    "DecimalType(9, 32)": "NUMBER(9, 32)",
-    "DecimalType(9, 33)": "NUMBER(9, 33)",
-    "DecimalType(9, 34)": "NUMBER(9, 34)",
-    "DecimalType(9, 35)": "NUMBER(9, 35)",
-    "DecimalType(9, 36)": "NUMBER(9, 36)",
-    "DecimalType(9, 37)": "NUMBER(9, 37)",
-    "DecimalType(9, 38)": "NUMBER(9, 38)",
-    "DecimalType(10, 0)": "NUMBER(10, 0)",
-    "DecimalType(10, 1)": "NUMBER(10, 1)",
-    "DecimalType(10, 2)": "NUMBER(10, 2)",
-    "DecimalType(10, 3)": "NUMBER(10, 3)",
-    "DecimalType(10, 4)": "NUMBER(10, 4)",
-    "DecimalType(10, 5)": "NUMBER(10, 5)",
-    "DecimalType(10, 6)": "NUMBER(10, 6)",
-    "DecimalType(10, 7)": "NUMBER(10, 7)",
-    "DecimalType(10, 8)": "NUMBER(10, 8)",
-    "DecimalType(10, 9)": "NUMBER(10, 9)",
-    "DecimalType(10, 10)": "NUMBER(10, 10)",
-    "DecimalType(10, 11)": "NUMBER(10, 11)",
-    "DecimalType(10, 12)": "NUMBER(10, 12)",
-    "DecimalType(10, 13)": "NUMBER(10, 13)",
-    "DecimalType(10, 14)": "NUMBER(10, 14)",
-    "DecimalType(10, 15)": "NUMBER(10, 15)",
-    "DecimalType(10, 16)": "NUMBER(10, 16)",
-    "DecimalType(10, 17)": "NUMBER(10, 17)",
-    "DecimalType(10, 18)": "NUMBER(10, 18)",
-    "DecimalType(10, 19)": "NUMBER(10, 19)",
-    "DecimalType(10, 20)": "NUMBER(10, 20)",
-    "DecimalType(10, 21)": "NUMBER(10, 21)",
-    "DecimalType(10, 22)": "NUMBER(10, 22)",
-    "DecimalType(10, 23)": "NUMBER(10, 23)",
-    "DecimalType(10, 24)": "NUMBER(10, 24)",
-    "DecimalType(10, 25)": "NUMBER(10, 25)",
-    "DecimalType(10, 26)": "NUMBER(10, 26)",
-    "DecimalType(10, 27)": "NUMBER(10, 27)",
-    "DecimalType(10, 28)": "NUMBER(10, 28)",
-    "DecimalType(10, 29)": "NUMBER(10, 29)",
-    "DecimalType(10, 30)": "NUMBER(10, 30)",
-    "DecimalType(10, 31)": "NUMBER(10, 31)",
-    "DecimalType(10, 32)": "NUMBER(10, 32)",
-    "DecimalType(10, 33)": "NUMBER(10, 33)",
-    "DecimalType(10, 34)": "NUMBER(10, 34)",
-    "DecimalType(10, 35)": "NUMBER(10, 35)",
-    "DecimalType(10, 36)": "NUMBER(10, 36)",
-    "DecimalType(10, 37)": "NUMBER(10, 37)",
-    "DecimalType(10, 38)": "NUMBER(10, 38)",
-    "DecimalType(11, 0)": "NUMBER(11, 0)",
-    "DecimalType(11, 1)": "NUMBER(11, 1)",
-    "DecimalType(11, 2)": "NUMBER(11, 2)",
-    "DecimalType(11, 3)": "NUMBER(11, 3)",
-    "DecimalType(11, 4)": "NUMBER(11, 4)",
-    "DecimalType(11, 5)": "NUMBER(11, 5)",
-    "DecimalType(11, 6)": "NUMBER(11, 6)",
-    "DecimalType(11, 7)": "NUMBER(11, 7)",
-    "DecimalType(11, 8)": "NUMBER(11, 8)",
-    "DecimalType(11, 9)": "NUMBER(11, 9)",
-    "DecimalType(11, 10)": "NUMBER(11, 10)",
-    "DecimalType(11, 11)": "NUMBER(11, 11)",
-    "DecimalType(11, 12)": "NUMBER(11, 12)",
-    "DecimalType(11, 13)": "NUMBER(11, 13)",
-    "DecimalType(11, 14)": "NUMBER(11, 14)",
-    "DecimalType(11, 15)": "NUMBER(11, 15)",
-    "DecimalType(11, 16)": "NUMBER(11, 16)",
-    "DecimalType(11, 17)": "NUMBER(11, 17)",
-    "DecimalType(11, 18)": "NUMBER(11, 18)",
-    "DecimalType(11, 19)": "NUMBER(11, 19)",
-    "DecimalType(11, 20)": "NUMBER(11, 20)",
-    "DecimalType(11, 21)": "NUMBER(11, 21)",
-    "DecimalType(11, 22)": "NUMBER(11, 22)",
-    "DecimalType(11, 23)": "NUMBER(11, 23)",
-    "DecimalType(11, 24)": "NUMBER(11, 24)",
-    "DecimalType(11, 25)": "NUMBER(11, 25)",
-    "DecimalType(11, 26)": "NUMBER(11, 26)",
-    "DecimalType(11, 27)": "NUMBER(11, 27)",
-    "DecimalType(11, 28)": "NUMBER(11, 28)",
-    "DecimalType(11, 29)": "NUMBER(11, 29)",
-    "DecimalType(11, 30)": "NUMBER(11, 30)",
-    "DecimalType(11, 31)": "NUMBER(11, 31)",
-    "DecimalType(11, 32)": "NUMBER(11, 32)",
-    "DecimalType(11, 33)": "NUMBER(11, 33)",
-    "DecimalType(11, 34)": "NUMBER(11, 34)",
-    "DecimalType(11, 35)": "NUMBER(11, 35)",
-    "DecimalType(11, 36)": "NUMBER(11, 36)",
-    "DecimalType(11, 37)": "NUMBER(11, 37)",
-    "DecimalType(11, 38)": "NUMBER(11, 38)",
-    "DecimalType(12, 0)": "NUMBER(12, 0)",
-    "DecimalType(12, 1)": "NUMBER(12, 1)",
-    "DecimalType(12, 2)": "NUMBER(12, 2)",
-    "DecimalType(12, 3)": "NUMBER(12, 3)",
-    "DecimalType(12, 4)": "NUMBER(12, 4)",
-    "DecimalType(12, 5)": "NUMBER(12, 5)",
-    "DecimalType(12, 6)": "NUMBER(12, 6)",
-    "DecimalType(12, 7)": "NUMBER(12, 7)",
-    "DecimalType(12, 8)": "NUMBER(12, 8)",
-    "DecimalType(12, 9)": "NUMBER(12, 9)",
-    "DecimalType(12, 10)": "NUMBER(12, 10)",
-    "DecimalType(12, 11)": "NUMBER(12, 11)",
-    "DecimalType(12, 12)": "NUMBER(12, 12)",
-    "DecimalType(12, 13)": "NUMBER(12, 13)",
-    "DecimalType(12, 14)": "NUMBER(12, 14)",
-    "DecimalType(12, 15)": "NUMBER(12, 15)",
-    "DecimalType(12, 16)": "NUMBER(12, 16)",
-    "DecimalType(12, 17)": "NUMBER(12, 17)",
-    "DecimalType(12, 18)": "NUMBER(12, 18)",
-    "DecimalType(12, 19)": "NUMBER(12, 19)",
-    "DecimalType(12, 20)": "NUMBER(12, 20)",
-    "DecimalType(12, 21)": "NUMBER(12, 21)",
-    "DecimalType(12, 22)": "NUMBER(12, 22)",
-    "DecimalType(12, 23)": "NUMBER(12, 23)",
-    "DecimalType(12, 24)": "NUMBER(12, 24)",
-    "DecimalType(12, 25)": "NUMBER(12, 25)",
-    "DecimalType(12, 26)": "NUMBER(12, 26)",
-    "DecimalType(12, 27)": "NUMBER(12, 27)",
-    "DecimalType(12, 28)": "NUMBER(12, 28)",
-    "DecimalType(12, 29)": "NUMBER(12, 29)",
-    "DecimalType(12, 30)": "NUMBER(12, 30)",
-    "DecimalType(12, 31)": "NUMBER(12, 31)",
-    "DecimalType(12, 32)": "NUMBER(12, 32)",
-    "DecimalType(12, 33)": "NUMBER(12, 33)",
-    "DecimalType(12, 34)": "NUMBER(12, 34)",
-    "DecimalType(12, 35)": "NUMBER(12, 35)",
-    "DecimalType(12, 36)": "NUMBER(12, 36)",
-    "DecimalType(12, 37)": "NUMBER(12, 37)",
-    "DecimalType(12, 38)": "NUMBER(12, 38)",
-    "DecimalType(13, 0)": "NUMBER(13, 0)",
-    "DecimalType(13, 1)": "NUMBER(13, 1)",
-    "DecimalType(13, 2)": "NUMBER(13, 2)",
-    "DecimalType(13, 3)": "NUMBER(13, 3)",
-    "DecimalType(13, 4)": "NUMBER(13, 4)",
-    "DecimalType(13, 5)": "NUMBER(13, 5)",
-    "DecimalType(13, 6)": "NUMBER(13, 6)",
-    "DecimalType(13, 7)": "NUMBER(13, 7)",
-    "DecimalType(13, 8)": "NUMBER(13, 8)",
-    "DecimalType(13, 9)": "NUMBER(13, 9)",
-    "DecimalType(13, 10)": "NUMBER(13, 10)",
-    "DecimalType(13, 11)": "NUMBER(13, 11)",
-    "DecimalType(13, 12)": "NUMBER(13, 12)",
-    "DecimalType(13, 13)": "NUMBER(13, 13)",
-    "DecimalType(13, 14)": "NUMBER(13, 14)",
-    "DecimalType(13, 15)": "NUMBER(13, 15)",
-    "DecimalType(13, 16)": "NUMBER(13, 16)",
-    "DecimalType(13, 17)": "NUMBER(13, 17)",
-    "DecimalType(13, 18)": "NUMBER(13, 18)",
-    "DecimalType(13, 19)": "NUMBER(13, 19)",
-    "DecimalType(13, 20)": "NUMBER(13, 20)",
-    "DecimalType(13, 21)": "NUMBER(13, 21)",
-    "DecimalType(13, 22)": "NUMBER(13, 22)",
-    "DecimalType(13, 23)": "NUMBER(13, 23)",
-    "DecimalType(13, 24)": "NUMBER(13, 24)",
-    "DecimalType(13, 25)": "NUMBER(13, 25)",
-    "DecimalType(13, 26)": "NUMBER(13, 26)",
-    "DecimalType(13, 27)": "NUMBER(13, 27)",
-    "DecimalType(13, 28)": "NUMBER(13, 28)",
-    "DecimalType(13, 29)": "NUMBER(13, 29)",
-    "DecimalType(13, 30)": "NUMBER(13, 30)",
-    "DecimalType(13, 31)": "NUMBER(13, 31)",
-    "DecimalType(13, 32)": "NUMBER(13, 32)",
-    "DecimalType(13, 33)": "NUMBER(13, 33)",
-    "DecimalType(13, 34)": "NUMBER(13, 34)",
-    "DecimalType(13, 35)": "NUMBER(13, 35)",
-    "DecimalType(13, 36)": "NUMBER(13, 36)",
-    "DecimalType(13, 37)": "NUMBER(13, 37)",
-    "DecimalType(13, 38)": "NUMBER(13, 38)",
-    "DecimalType(14, 0)": "NUMBER(14, 0)",
-    "DecimalType(14, 1)": "NUMBER(14, 1)",
-    "DecimalType(14, 2)": "NUMBER(14, 2)",
-    "DecimalType(14, 3)": "NUMBER(14, 3)",
-    "DecimalType(14, 4)": "NUMBER(14, 4)",
-    "DecimalType(14, 5)": "NUMBER(14, 5)",
-    "DecimalType(14, 6)": "NUMBER(14, 6)",
-    "DecimalType(14, 7)": "NUMBER(14, 7)",
-    "DecimalType(14, 8)": "NUMBER(14, 8)",
-    "DecimalType(14, 9)": "NUMBER(14, 9)",
-    "DecimalType(14, 10)": "NUMBER(14, 10)",
-    "DecimalType(14, 11)": "NUMBER(14, 11)",
-    "DecimalType(14, 12)": "NUMBER(14, 12)",
-    "DecimalType(14, 13)": "NUMBER(14, 13)",
-    "DecimalType(14, 14)": "NUMBER(14, 14)",
-    "DecimalType(14, 15)": "NUMBER(14, 15)",
-    "DecimalType(14, 16)": "NUMBER(14, 16)",
-    "DecimalType(14, 17)": "NUMBER(14, 17)",
-    "DecimalType(14, 18)": "NUMBER(14, 18)",
-    "DecimalType(14, 19)": "NUMBER(14, 19)",
-    "DecimalType(14, 20)": "NUMBER(14, 20)",
-    "DecimalType(14, 21)": "NUMBER(14, 21)",
-    "DecimalType(14, 22)": "NUMBER(14, 22)",
-    "DecimalType(14, 23)": "NUMBER(14, 23)",
-    "DecimalType(14, 24)": "NUMBER(14, 24)",
-    "DecimalType(14, 25)": "NUMBER(14, 25)",
-    "DecimalType(14, 26)": "NUMBER(14, 26)",
-    "DecimalType(14, 27)": "NUMBER(14, 27)",
-    "DecimalType(14, 28)": "NUMBER(14, 28)",
-    "DecimalType(14, 29)": "NUMBER(14, 29)",
-    "DecimalType(14, 30)": "NUMBER(14, 30)",
-    "DecimalType(14, 31)": "NUMBER(14, 31)",
-    "DecimalType(14, 32)": "NUMBER(14, 32)",
-    "DecimalType(14, 33)": "NUMBER(14, 33)",
-    "DecimalType(14, 34)": "NUMBER(14, 34)",
-    "DecimalType(14, 35)": "NUMBER(14, 35)",
-    "DecimalType(14, 36)": "NUMBER(14, 36)",
-    "DecimalType(14, 37)": "NUMBER(14, 37)",
-    "DecimalType(14, 38)": "NUMBER(14, 38)",
-    "DecimalType(15, 0)": "NUMBER(15, 0)",
-    "DecimalType(15, 1)": "NUMBER(15, 1)",
-    "DecimalType(15, 2)": "NUMBER(15, 2)",
-    "DecimalType(15, 3)": "NUMBER(15, 3)",
-    "DecimalType(15, 4)": "NUMBER(15, 4)",
-    "DecimalType(15, 5)": "NUMBER(15, 5)",
-    "DecimalType(15, 6)": "NUMBER(15, 6)",
-    "DecimalType(15, 7)": "NUMBER(15, 7)",
-    "DecimalType(15, 8)": "NUMBER(15, 8)",
-    "DecimalType(15, 9)": "NUMBER(15, 9)",
-    "DecimalType(15, 10)": "NUMBER(15, 10)",
-    "DecimalType(15, 11)": "NUMBER(15, 11)",
-    "DecimalType(15, 12)": "NUMBER(15, 12)",
-    "DecimalType(15, 13)": "NUMBER(15, 13)",
-    "DecimalType(15, 14)": "NUMBER(15, 14)",
-    "DecimalType(15, 15)": "NUMBER(15, 15)",
-    "DecimalType(15, 16)": "NUMBER(15, 16)",
-    "DecimalType(15, 17)": "NUMBER(15, 17)",
-    "DecimalType(15, 18)": "NUMBER(15, 18)",
-    "DecimalType(15, 19)": "NUMBER(15, 19)",
-    "DecimalType(15, 20)": "NUMBER(15, 20)",
-    "DecimalType(15, 21)": "NUMBER(15, 21)",
-    "DecimalType(15, 22)": "NUMBER(15, 22)",
-    "DecimalType(15, 23)": "NUMBER(15, 23)",
-    "DecimalType(15, 24)": "NUMBER(15, 24)",
-    "DecimalType(15, 25)": "NUMBER(15, 25)",
-    "DecimalType(15, 26)": "NUMBER(15, 26)",
-    "DecimalType(15, 27)": "NUMBER(15, 27)",
-    "DecimalType(15, 28)": "NUMBER(15, 28)",
-    "DecimalType(15, 29)": "NUMBER(15, 29)",
-    "DecimalType(15, 30)": "NUMBER(15, 30)",
-    "DecimalType(15, 31)": "NUMBER(15, 31)",
-    "DecimalType(15, 32)": "NUMBER(15, 32)",
-    "DecimalType(15, 33)": "NUMBER(15, 33)",
-    "DecimalType(15, 34)": "NUMBER(15, 34)",
-    "DecimalType(15, 35)": "NUMBER(15, 35)",
-    "DecimalType(15, 36)": "NUMBER(15, 36)",
-    "DecimalType(15, 37)": "NUMBER(15, 37)",
-    "DecimalType(15, 38)": "NUMBER(15, 38)",
-    "DecimalType(16, 0)": "NUMBER(16, 0)",
-    "DecimalType(16, 1)": "NUMBER(16, 1)",
-    "DecimalType(16, 2)": "NUMBER(16, 2)",
-    "DecimalType(16, 3)": "NUMBER(16, 3)",
-    "DecimalType(16, 4)": "NUMBER(16, 4)",
-    "DecimalType(16, 5)": "NUMBER(16, 5)",
-    "DecimalType(16, 6)": "NUMBER(16, 6)",
-    "DecimalType(16, 7)": "NUMBER(16, 7)",
-    "DecimalType(16, 8)": "NUMBER(16, 8)",
-    "DecimalType(16, 9)": "NUMBER(16, 9)",
-    "DecimalType(16, 10)": "NUMBER(16, 10)",
-    "DecimalType(16, 11)": "NUMBER(16, 11)",
-    "DecimalType(16, 12)": "NUMBER(16, 12)",
-    "DecimalType(16, 13)": "NUMBER(16, 13)",
-    "DecimalType(16, 14)": "NUMBER(16, 14)",
-    "DecimalType(16, 15)": "NUMBER(16, 15)",
-    "DecimalType(16, 16)": "NUMBER(16, 16)",
-    "DecimalType(16, 17)": "NUMBER(16, 17)",
-    "DecimalType(16, 18)": "NUMBER(16, 18)",
-    "DecimalType(16, 19)": "NUMBER(16, 19)",
-    "DecimalType(16, 20)": "NUMBER(16, 20)",
-    "DecimalType(16, 21)": "NUMBER(16, 21)",
-    "DecimalType(16, 22)": "NUMBER(16, 22)",
-    "DecimalType(16, 23)": "NUMBER(16, 23)",
-    "DecimalType(16, 24)": "NUMBER(16, 24)",
-    "DecimalType(16, 25)": "NUMBER(16, 25)",
-    "DecimalType(16, 26)": "NUMBER(16, 26)",
-    "DecimalType(16, 27)": "NUMBER(16, 27)",
-    "DecimalType(16, 28)": "NUMBER(16, 28)",
-    "DecimalType(16, 29)": "NUMBER(16, 29)",
-    "DecimalType(16, 30)": "NUMBER(16, 30)",
-    "DecimalType(16, 31)": "NUMBER(16, 31)",
-    "DecimalType(16, 32)": "NUMBER(16, 32)",
-    "DecimalType(16, 33)": "NUMBER(16, 33)",
-    "DecimalType(16, 34)": "NUMBER(16, 34)",
-    "DecimalType(16, 35)": "NUMBER(16, 35)",
-    "DecimalType(16, 36)": "NUMBER(16, 36)",
-    "DecimalType(16, 37)": "NUMBER(16, 37)",
-    "DecimalType(16, 38)": "NUMBER(16, 38)",
-    "DecimalType(17, 0)": "NUMBER(17, 0)",
-    "DecimalType(17, 1)": "NUMBER(17, 1)",
-    "DecimalType(17, 2)": "NUMBER(17, 2)",
-    "DecimalType(17, 3)": "NUMBER(17, 3)",
-    "DecimalType(17, 4)": "NUMBER(17, 4)",
-    "DecimalType(17, 5)": "NUMBER(17, 5)",
-    "DecimalType(17, 6)": "NUMBER(17, 6)",
-    "DecimalType(17, 7)": "NUMBER(17, 7)",
-    "DecimalType(17, 8)": "NUMBER(17, 8)",
-    "DecimalType(17, 9)": "NUMBER(17, 9)",
-    "DecimalType(17, 10)": "NUMBER(17, 10)",
-    "DecimalType(17, 11)": "NUMBER(17, 11)",
-    "DecimalType(17, 12)": "NUMBER(17, 12)",
-    "DecimalType(17, 13)": "NUMBER(17, 13)",
-    "DecimalType(17, 14)": "NUMBER(17, 14)",
-    "DecimalType(17, 15)": "NUMBER(17, 15)",
-    "DecimalType(17, 16)": "NUMBER(17, 16)",
-    "DecimalType(17, 17)": "NUMBER(17, 17)",
-    "DecimalType(17, 18)": "NUMBER(17, 18)",
-    "DecimalType(17, 19)": "NUMBER(17, 19)",
-    "DecimalType(17, 20)": "NUMBER(17, 20)",
-    "DecimalType(17, 21)": "NUMBER(17, 21)",
-    "DecimalType(17, 22)": "NUMBER(17, 22)",
-    "DecimalType(17, 23)": "NUMBER(17, 23)",
-    "DecimalType(17, 24)": "NUMBER(17, 24)",
-    "DecimalType(17, 25)": "NUMBER(17, 25)",
-    "DecimalType(17, 26)": "NUMBER(17, 26)",
-    "DecimalType(17, 27)": "NUMBER(17, 27)",
-    "DecimalType(17, 28)": "NUMBER(17, 28)",
-    "DecimalType(17, 29)": "NUMBER(17, 29)",
-    "DecimalType(17, 30)": "NUMBER(17, 30)",
-    "DecimalType(17, 31)": "NUMBER(17, 31)",
-    "DecimalType(17, 32)": "NUMBER(17, 32)",
-    "DecimalType(17, 33)": "NUMBER(17, 33)",
-    "DecimalType(17, 34)": "NUMBER(17, 34)",
-    "DecimalType(17, 35)": "NUMBER(17, 35)",
-    "DecimalType(17, 36)": "NUMBER(17, 36)",
-    "DecimalType(17, 37)": "NUMBER(17, 37)",
-    "DecimalType(17, 38)": "NUMBER(17, 38)",
-    "DecimalType(18, 0)": "NUMBER(18, 0)",
-    "DecimalType(18, 1)": "NUMBER(18, 1)",
-    "DecimalType(18, 2)": "NUMBER(18, 2)",
-    "DecimalType(18, 3)": "NUMBER(18, 3)",
-    "DecimalType(18, 4)": "NUMBER(18, 4)",
-    "DecimalType(18, 5)": "NUMBER(18, 5)",
-    "DecimalType(18, 6)": "NUMBER(18, 6)",
-    "DecimalType(18, 7)": "NUMBER(18, 7)",
-    "DecimalType(18, 8)": "NUMBER(18, 8)",
-    "DecimalType(18, 9)": "NUMBER(18, 9)",
-    "DecimalType(18, 10)": "NUMBER(18, 10)",
-    "DecimalType(18, 11)": "NUMBER(18, 11)",
-    "DecimalType(18, 12)": "NUMBER(18, 12)",
-    "DecimalType(18, 13)": "NUMBER(18, 13)",
-    "DecimalType(18, 14)": "NUMBER(18, 14)",
-    "DecimalType(18, 15)": "NUMBER(18, 15)",
-    "DecimalType(18, 16)": "NUMBER(18, 16)",
-    "DecimalType(18, 17)": "NUMBER(18, 17)",
-    "DecimalType(18, 18)": "NUMBER(18, 18)",
-    "DecimalType(18, 19)": "NUMBER(18, 19)",
-    "DecimalType(18, 20)": "NUMBER(18, 20)",
-    "DecimalType(18, 21)": "NUMBER(18, 21)",
-    "DecimalType(18, 22)": "NUMBER(18, 22)",
-    "DecimalType(18, 23)": "NUMBER(18, 23)",
-    "DecimalType(18, 24)": "NUMBER(18, 24)",
-    "DecimalType(18, 25)": "NUMBER(18, 25)",
-    "DecimalType(18, 26)": "NUMBER(18, 26)",
-    "DecimalType(18, 27)": "NUMBER(18, 27)",
-    "DecimalType(18, 28)": "NUMBER(18, 28)",
-    "DecimalType(18, 29)": "NUMBER(18, 29)",
-    "DecimalType(18, 30)": "NUMBER(18, 30)",
-    "DecimalType(18, 31)": "NUMBER(18, 31)",
-    "DecimalType(18, 32)": "NUMBER(18, 32)",
-    "DecimalType(18, 33)": "NUMBER(18, 33)",
-    "DecimalType(18, 34)": "NUMBER(18, 34)",
-    "DecimalType(18, 35)": "NUMBER(18, 35)",
-    "DecimalType(18, 36)": "NUMBER(18, 36)",
-    "DecimalType(18, 37)": "NUMBER(18, 37)",
-    "DecimalType(18, 38)": "NUMBER(18, 38)",
-    "DecimalType(19, 0)": "NUMBER(19, 0)",
-    "DecimalType(19, 1)": "NUMBER(19, 1)",
-    "DecimalType(19, 2)": "NUMBER(19, 2)",
-    "DecimalType(19, 3)": "NUMBER(19, 3)",
-    "DecimalType(19, 4)": "NUMBER(19, 4)",
-    "DecimalType(19, 5)": "NUMBER(19, 5)",
-    "DecimalType(19, 6)": "NUMBER(19, 6)",
-    "DecimalType(19, 7)": "NUMBER(19, 7)",
-    "DecimalType(19, 8)": "NUMBER(19, 8)",
-    "DecimalType(19, 9)": "NUMBER(19, 9)",
-    "DecimalType(19, 10)": "NUMBER(19, 10)",
-    "DecimalType(19, 11)": "NUMBER(19, 11)",
-    "DecimalType(19, 12)": "NUMBER(19, 12)",
-    "DecimalType(19, 13)": "NUMBER(19, 13)",
-    "DecimalType(19, 14)": "NUMBER(19, 14)",
-    "DecimalType(19, 15)": "NUMBER(19, 15)",
-    "DecimalType(19, 16)": "NUMBER(19, 16)",
-    "DecimalType(19, 17)": "NUMBER(19, 17)",
-    "DecimalType(19, 18)": "NUMBER(19, 18)",
-    "DecimalType(19, 19)": "NUMBER(19, 19)",
-    "DecimalType(19, 20)": "NUMBER(19, 20)",
-    "DecimalType(19, 21)": "NUMBER(19, 21)",
-    "DecimalType(19, 22)": "NUMBER(19, 22)",
-    "DecimalType(19, 23)": "NUMBER(19, 23)",
-    "DecimalType(19, 24)": "NUMBER(19, 24)",
-    "DecimalType(19, 25)": "NUMBER(19, 25)",
-    "DecimalType(19, 26)": "NUMBER(19, 26)",
-    "DecimalType(19, 27)": "NUMBER(19, 27)",
-    "DecimalType(19, 28)": "NUMBER(19, 28)",
-    "DecimalType(19, 29)": "NUMBER(19, 29)",
-    "DecimalType(19, 30)": "NUMBER(19, 30)",
-    "DecimalType(19, 31)": "NUMBER(19, 31)",
-    "DecimalType(19, 32)": "NUMBER(19, 32)",
-    "DecimalType(19, 33)": "NUMBER(19, 33)",
-    "DecimalType(19, 34)": "NUMBER(19, 34)",
-    "DecimalType(19, 35)": "NUMBER(19, 35)",
-    "DecimalType(19, 36)": "NUMBER(19, 36)",
-    "DecimalType(19, 37)": "NUMBER(19, 37)",
-    "DecimalType(19, 38)": "NUMBER(19, 38)",
-    "DecimalType(20, 0)": "NUMBER(20, 0)",
-    "DecimalType(20, 1)": "NUMBER(20, 1)",
-    "DecimalType(20, 2)": "NUMBER(20, 2)",
-    "DecimalType(20, 3)": "NUMBER(20, 3)",
-    "DecimalType(20, 4)": "NUMBER(20, 4)",
-    "DecimalType(20, 5)": "NUMBER(20, 5)",
-    "DecimalType(20, 6)": "NUMBER(20, 6)",
-    "DecimalType(20, 7)": "NUMBER(20, 7)",
-    "DecimalType(20, 8)": "NUMBER(20, 8)",
-    "DecimalType(20, 9)": "NUMBER(20, 9)",
-    "DecimalType(20, 10)": "NUMBER(20, 10)",
-    "DecimalType(20, 11)": "NUMBER(20, 11)",
-    "DecimalType(20, 12)": "NUMBER(20, 12)",
-    "DecimalType(20, 13)": "NUMBER(20, 13)",
-    "DecimalType(20, 14)": "NUMBER(20, 14)",
-    "DecimalType(20, 15)": "NUMBER(20, 15)",
-    "DecimalType(20, 16)": "NUMBER(20, 16)",
-    "DecimalType(20, 17)": "NUMBER(20, 17)",
-    "DecimalType(20, 18)": "NUMBER(20, 18)",
-    "DecimalType(20, 19)": "NUMBER(20, 19)",
-    "DecimalType(20, 20)": "NUMBER(20, 20)",
-    "DecimalType(20, 21)": "NUMBER(20, 21)",
-    "DecimalType(20, 22)": "NUMBER(20, 22)",
-    "DecimalType(20, 23)": "NUMBER(20, 23)",
-    "DecimalType(20, 24)": "NUMBER(20, 24)",
-    "DecimalType(20, 25)": "NUMBER(20, 25)",
-    "DecimalType(20, 26)": "NUMBER(20, 26)",
-    "DecimalType(20, 27)": "NUMBER(20, 27)",
-    "DecimalType(20, 28)": "NUMBER(20, 28)",
-    "DecimalType(20, 29)": "NUMBER(20, 29)",
-    "DecimalType(20, 30)": "NUMBER(20, 30)",
-    "DecimalType(20, 31)": "NUMBER(20, 31)",
-    "DecimalType(20, 32)": "NUMBER(20, 32)",
-    "DecimalType(20, 33)": "NUMBER(20, 33)",
-    "DecimalType(20, 34)": "NUMBER(20, 34)",
-    "DecimalType(20, 35)": "NUMBER(20, 35)",
-    "DecimalType(20, 36)": "NUMBER(20, 36)",
-    "DecimalType(20, 37)": "NUMBER(20, 37)",
-    "DecimalType(20, 38)": "NUMBER(20, 38)",
-    "DecimalType(21, 0)": "NUMBER(21, 0)",
-    "DecimalType(21, 1)": "NUMBER(21, 1)",
-    "DecimalType(21, 2)": "NUMBER(21, 2)",
-    "DecimalType(21, 3)": "NUMBER(21, 3)",
-    "DecimalType(21, 4)": "NUMBER(21, 4)",
-    "DecimalType(21, 5)": "NUMBER(21, 5)",
-    "DecimalType(21, 6)": "NUMBER(21, 6)",
-    "DecimalType(21, 7)": "NUMBER(21, 7)",
-    "DecimalType(21, 8)": "NUMBER(21, 8)",
-    "DecimalType(21, 9)": "NUMBER(21, 9)",
-    "DecimalType(21, 10)": "NUMBER(21, 10)",
-    "DecimalType(21, 11)": "NUMBER(21, 11)",
-    "DecimalType(21, 12)": "NUMBER(21, 12)",
-    "DecimalType(21, 13)": "NUMBER(21, 13)",
-    "DecimalType(21, 14)": "NUMBER(21, 14)",
-    "DecimalType(21, 15)": "NUMBER(21, 15)",
-    "DecimalType(21, 16)": "NUMBER(21, 16)",
-    "DecimalType(21, 17)": "NUMBER(21, 17)",
-    "DecimalType(21, 18)": "NUMBER(21, 18)",
-    "DecimalType(21, 19)": "NUMBER(21, 19)",
-    "DecimalType(21, 20)": "NUMBER(21, 20)",
-    "DecimalType(21, 21)": "NUMBER(21, 21)",
-    "DecimalType(21, 22)": "NUMBER(21, 22)",
-    "DecimalType(21, 23)": "NUMBER(21, 23)",
-    "DecimalType(21, 24)": "NUMBER(21, 24)",
-    "DecimalType(21, 25)": "NUMBER(21, 25)",
-    "DecimalType(21, 26)": "NUMBER(21, 26)",
-    "DecimalType(21, 27)": "NUMBER(21, 27)",
-    "DecimalType(21, 28)": "NUMBER(21, 28)",
-    "DecimalType(21, 29)": "NUMBER(21, 29)",
-    "DecimalType(21, 30)": "NUMBER(21, 30)",
-    "DecimalType(21, 31)": "NUMBER(21, 31)",
-    "DecimalType(21, 32)": "NUMBER(21, 32)",
-    "DecimalType(21, 33)": "NUMBER(21, 33)",
-    "DecimalType(21, 34)": "NUMBER(21, 34)",
-    "DecimalType(21, 35)": "NUMBER(21, 35)",
-    "DecimalType(21, 36)": "NUMBER(21, 36)",
-    "DecimalType(21, 37)": "NUMBER(21, 37)",
-    "DecimalType(21, 38)": "NUMBER(21, 38)",
-    "DecimalType(22, 0)": "NUMBER(22, 0)",
-    "DecimalType(22, 1)": "NUMBER(22, 1)",
-    "DecimalType(22, 2)": "NUMBER(22, 2)",
-    "DecimalType(22, 3)": "NUMBER(22, 3)",
-    "DecimalType(22, 4)": "NUMBER(22, 4)",
-    "DecimalType(22, 5)": "NUMBER(22, 5)",
-    "DecimalType(22, 6)": "NUMBER(22, 6)",
-    "DecimalType(22, 7)": "NUMBER(22, 7)",
-    "DecimalType(22, 8)": "NUMBER(22, 8)",
-    "DecimalType(22, 9)": "NUMBER(22, 9)",
-    "DecimalType(22, 10)": "NUMBER(22, 10)",
-    "DecimalType(22, 11)": "NUMBER(22, 11)",
-    "DecimalType(22, 12)": "NUMBER(22, 12)",
-    "DecimalType(22, 13)": "NUMBER(22, 13)",
-    "DecimalType(22, 14)": "NUMBER(22, 14)",
-    "DecimalType(22, 15)": "NUMBER(22, 15)",
-    "DecimalType(22, 16)": "NUMBER(22, 16)",
-    "DecimalType(22, 17)": "NUMBER(22, 17)",
-    "DecimalType(22, 18)": "NUMBER(22, 18)",
-    "DecimalType(22, 19)": "NUMBER(22, 19)",
-    "DecimalType(22, 20)": "NUMBER(22, 20)",
-    "DecimalType(22, 21)": "NUMBER(22, 21)",
-    "DecimalType(22, 22)": "NUMBER(22, 22)",
-    "DecimalType(22, 23)": "NUMBER(22, 23)",
-    "DecimalType(22, 24)": "NUMBER(22, 24)",
-    "DecimalType(22, 25)": "NUMBER(22, 25)",
-    "DecimalType(22, 26)": "NUMBER(22, 26)",
-    "DecimalType(22, 27)": "NUMBER(22, 27)",
-    "DecimalType(22, 28)": "NUMBER(22, 28)",
-    "DecimalType(22, 29)": "NUMBER(22, 29)",
-    "DecimalType(22, 30)": "NUMBER(22, 30)",
-    "DecimalType(22, 31)": "NUMBER(22, 31)",
-    "DecimalType(22, 32)": "NUMBER(22, 32)",
-    "DecimalType(22, 33)": "NUMBER(22, 33)",
-    "DecimalType(22, 34)": "NUMBER(22, 34)",
-    "DecimalType(22, 35)": "NUMBER(22, 35)",
-    "DecimalType(22, 36)": "NUMBER(22, 36)",
-    "DecimalType(22, 37)": "NUMBER(22, 37)",
-    "DecimalType(22, 38)": "NUMBER(22, 38)",
-    "DecimalType(23, 0)": "NUMBER(23, 0)",
-    "DecimalType(23, 1)": "NUMBER(23, 1)",
-    "DecimalType(23, 2)": "NUMBER(23, 2)",
-    "DecimalType(23, 3)": "NUMBER(23, 3)",
-    "DecimalType(23, 4)": "NUMBER(23, 4)",
-    "DecimalType(23, 5)": "NUMBER(23, 5)",
-    "DecimalType(23, 6)": "NUMBER(23, 6)",
-    "DecimalType(23, 7)": "NUMBER(23, 7)",
-    "DecimalType(23, 8)": "NUMBER(23, 8)",
-    "DecimalType(23, 9)": "NUMBER(23, 9)",
-    "DecimalType(23, 10)": "NUMBER(23, 10)",
-    "DecimalType(23, 11)": "NUMBER(23, 11)",
-    "DecimalType(23, 12)": "NUMBER(23, 12)",
-    "DecimalType(23, 13)": "NUMBER(23, 13)",
-    "DecimalType(23, 14)": "NUMBER(23, 14)",
-    "DecimalType(23, 15)": "NUMBER(23, 15)",
-    "DecimalType(23, 16)": "NUMBER(23, 16)",
-    "DecimalType(23, 17)": "NUMBER(23, 17)",
-    "DecimalType(23, 18)": "NUMBER(23, 18)",
-    "DecimalType(23, 19)": "NUMBER(23, 19)",
-    "DecimalType(23, 20)": "NUMBER(23, 20)",
-    "DecimalType(23, 21)": "NUMBER(23, 21)",
-    "DecimalType(23, 22)": "NUMBER(23, 22)",
-    "DecimalType(23, 23)": "NUMBER(23, 23)",
-    "DecimalType(23, 24)": "NUMBER(23, 24)",
-    "DecimalType(23, 25)": "NUMBER(23, 25)",
-    "DecimalType(23, 26)": "NUMBER(23, 26)",
-    "DecimalType(23, 27)": "NUMBER(23, 27)",
-    "DecimalType(23, 28)": "NUMBER(23, 28)",
-    "DecimalType(23, 29)": "NUMBER(23, 29)",
-    "DecimalType(23, 30)": "NUMBER(23, 30)",
-    "DecimalType(23, 31)": "NUMBER(23, 31)",
-    "DecimalType(23, 32)": "NUMBER(23, 32)",
-    "DecimalType(23, 33)": "NUMBER(23, 33)",
-    "DecimalType(23, 34)": "NUMBER(23, 34)",
-    "DecimalType(23, 35)": "NUMBER(23, 35)",
-    "DecimalType(23, 36)": "NUMBER(23, 36)",
-    "DecimalType(23, 37)": "NUMBER(23, 37)",
-    "DecimalType(23, 38)": "NUMBER(23, 38)",
-    "DecimalType(24, 0)": "NUMBER(24, 0)",
-    "DecimalType(24, 1)": "NUMBER(24, 1)",
-    "DecimalType(24, 2)": "NUMBER(24, 2)",
-    "DecimalType(24, 3)": "NUMBER(24, 3)",
-    "DecimalType(24, 4)": "NUMBER(24, 4)",
-    "DecimalType(24, 5)": "NUMBER(24, 5)",
-    "DecimalType(24, 6)": "NUMBER(24, 6)",
-    "DecimalType(24, 7)": "NUMBER(24, 7)",
-    "DecimalType(24, 8)": "NUMBER(24, 8)",
-    "DecimalType(24, 9)": "NUMBER(24, 9)",
-    "DecimalType(24, 10)": "NUMBER(24, 10)",
-    "DecimalType(24, 11)": "NUMBER(24, 11)",
-    "DecimalType(24, 12)": "NUMBER(24, 12)",
-    "DecimalType(24, 13)": "NUMBER(24, 13)",
-    "DecimalType(24, 14)": "NUMBER(24, 14)",
-    "DecimalType(24, 15)": "NUMBER(24, 15)",
-    "DecimalType(24, 16)": "NUMBER(24, 16)",
-    "DecimalType(24, 17)": "NUMBER(24, 17)",
-    "DecimalType(24, 18)": "NUMBER(24, 18)",
-    "DecimalType(24, 19)": "NUMBER(24, 19)",
-    "DecimalType(24, 20)": "NUMBER(24, 20)",
-    "DecimalType(24, 21)": "NUMBER(24, 21)",
-    "DecimalType(24, 22)": "NUMBER(24, 22)",
-    "DecimalType(24, 23)": "NUMBER(24, 23)",
-    "DecimalType(24, 24)": "NUMBER(24, 24)",
-    "DecimalType(24, 25)": "NUMBER(24, 25)",
-    "DecimalType(24, 26)": "NUMBER(24, 26)",
-    "DecimalType(24, 27)": "NUMBER(24, 27)",
-    "DecimalType(24, 28)": "NUMBER(24, 28)",
-    "DecimalType(24, 29)": "NUMBER(24, 29)",
-    "DecimalType(24, 30)": "NUMBER(24, 30)",
-    "DecimalType(24, 31)": "NUMBER(24, 31)",
-    "DecimalType(24, 32)": "NUMBER(24, 32)",
-    "DecimalType(24, 33)": "NUMBER(24, 33)",
-    "DecimalType(24, 34)": "NUMBER(24, 34)",
-    "DecimalType(24, 35)": "NUMBER(24, 35)",
-    "DecimalType(24, 36)": "NUMBER(24, 36)",
-    "DecimalType(24, 37)": "NUMBER(24, 37)",
-    "DecimalType(24, 38)": "NUMBER(24, 38)",
-    "DecimalType(25, 0)": "NUMBER(25, 0)",
-    "DecimalType(25, 1)": "NUMBER(25, 1)",
-    "DecimalType(25, 2)": "NUMBER(25, 2)",
-    "DecimalType(25, 3)": "NUMBER(25, 3)",
-    "DecimalType(25, 4)": "NUMBER(25, 4)",
-    "DecimalType(25, 5)": "NUMBER(25, 5)",
-    "DecimalType(25, 6)": "NUMBER(25, 6)",
-    "DecimalType(25, 7)": "NUMBER(25, 7)",
-    "DecimalType(25, 8)": "NUMBER(25, 8)",
-    "DecimalType(25, 9)": "NUMBER(25, 9)",
-    "DecimalType(25, 10)": "NUMBER(25, 10)",
-    "DecimalType(25, 11)": "NUMBER(25, 11)",
-    "DecimalType(25, 12)": "NUMBER(25, 12)",
-    "DecimalType(25, 13)": "NUMBER(25, 13)",
-    "DecimalType(25, 14)": "NUMBER(25, 14)",
-    "DecimalType(25, 15)": "NUMBER(25, 15)",
-    "DecimalType(25, 16)": "NUMBER(25, 16)",
-    "DecimalType(25, 17)": "NUMBER(25, 17)",
-    "DecimalType(25, 18)": "NUMBER(25, 18)",
-    "DecimalType(25, 19)": "NUMBER(25, 19)",
-    "DecimalType(25, 20)": "NUMBER(25, 20)",
-    "DecimalType(25, 21)": "NUMBER(25, 21)",
-    "DecimalType(25, 22)": "NUMBER(25, 22)",
-    "DecimalType(25, 23)": "NUMBER(25, 23)",
-    "DecimalType(25, 24)": "NUMBER(25, 24)",
-    "DecimalType(25, 25)": "NUMBER(25, 25)",
-    "DecimalType(25, 26)": "NUMBER(25, 26)",
-    "DecimalType(25, 27)": "NUMBER(25, 27)",
-    "DecimalType(25, 28)": "NUMBER(25, 28)",
-    "DecimalType(25, 29)": "NUMBER(25, 29)",
-    "DecimalType(25, 30)": "NUMBER(25, 30)",
-    "DecimalType(25, 31)": "NUMBER(25, 31)",
-    "DecimalType(25, 32)": "NUMBER(25, 32)",
-    "DecimalType(25, 33)": "NUMBER(25, 33)",
-    "DecimalType(25, 34)": "NUMBER(25, 34)",
-    "DecimalType(25, 35)": "NUMBER(25, 35)",
-    "DecimalType(25, 36)": "NUMBER(25, 36)",
-    "DecimalType(25, 37)": "NUMBER(25, 37)",
-    "DecimalType(25, 38)": "NUMBER(25, 38)",
-    "DecimalType(26, 0)": "NUMBER(26, 0)",
-    "DecimalType(26, 1)": "NUMBER(26, 1)",
-    "DecimalType(26, 2)": "NUMBER(26, 2)",
-    "DecimalType(26, 3)": "NUMBER(26, 3)",
-    "DecimalType(26, 4)": "NUMBER(26, 4)",
-    "DecimalType(26, 5)": "NUMBER(26, 5)",
-    "DecimalType(26, 6)": "NUMBER(26, 6)",
-    "DecimalType(26, 7)": "NUMBER(26, 7)",
-    "DecimalType(26, 8)": "NUMBER(26, 8)",
-    "DecimalType(26, 9)": "NUMBER(26, 9)",
-    "DecimalType(26, 10)": "NUMBER(26, 10)",
-    "DecimalType(26, 11)": "NUMBER(26, 11)",
-    "DecimalType(26, 12)": "NUMBER(26, 12)",
-    "DecimalType(26, 13)": "NUMBER(26, 13)",
-    "DecimalType(26, 14)": "NUMBER(26, 14)",
-    "DecimalType(26, 15)": "NUMBER(26, 15)",
-    "DecimalType(26, 16)": "NUMBER(26, 16)",
-    "DecimalType(26, 17)": "NUMBER(26, 17)",
-    "DecimalType(26, 18)": "NUMBER(26, 18)",
-    "DecimalType(26, 19)": "NUMBER(26, 19)",
-    "DecimalType(26, 20)": "NUMBER(26, 20)",
-    "DecimalType(26, 21)": "NUMBER(26, 21)",
-    "DecimalType(26, 22)": "NUMBER(26, 22)",
-    "DecimalType(26, 23)": "NUMBER(26, 23)",
-    "DecimalType(26, 24)": "NUMBER(26, 24)",
-    "DecimalType(26, 25)": "NUMBER(26, 25)",
-    "DecimalType(26, 26)": "NUMBER(26, 26)",
-    "DecimalType(26, 27)": "NUMBER(26, 27)",
-    "DecimalType(26, 28)": "NUMBER(26, 28)",
-    "DecimalType(26, 29)": "NUMBER(26, 29)",
-    "DecimalType(26, 30)": "NUMBER(26, 30)",
-    "DecimalType(26, 31)": "NUMBER(26, 31)",
-    "DecimalType(26, 32)": "NUMBER(26, 32)",
-    "DecimalType(26, 33)": "NUMBER(26, 33)",
-    "DecimalType(26, 34)": "NUMBER(26, 34)",
-    "DecimalType(26, 35)": "NUMBER(26, 35)",
-    "DecimalType(26, 36)": "NUMBER(26, 36)",
-    "DecimalType(26, 37)": "NUMBER(26, 37)",
-    "DecimalType(26, 38)": "NUMBER(26, 38)",
-    "DecimalType(27, 0)": "NUMBER(27, 0)",
-    "DecimalType(27, 1)": "NUMBER(27, 1)",
-    "DecimalType(27, 2)": "NUMBER(27, 2)",
-    "DecimalType(27, 3)": "NUMBER(27, 3)",
-    "DecimalType(27, 4)": "NUMBER(27, 4)",
-    "DecimalType(27, 5)": "NUMBER(27, 5)",
-    "DecimalType(27, 6)": "NUMBER(27, 6)",
-    "DecimalType(27, 7)": "NUMBER(27, 7)",
-    "DecimalType(27, 8)": "NUMBER(27, 8)",
-    "DecimalType(27, 9)": "NUMBER(27, 9)",
-    "DecimalType(27, 10)": "NUMBER(27, 10)",
-    "DecimalType(27, 11)": "NUMBER(27, 11)",
-    "DecimalType(27, 12)": "NUMBER(27, 12)",
-    "DecimalType(27, 13)": "NUMBER(27, 13)",
-    "DecimalType(27, 14)": "NUMBER(27, 14)",
-    "DecimalType(27, 15)": "NUMBER(27, 15)",
-    "DecimalType(27, 16)": "NUMBER(27, 16)",
-    "DecimalType(27, 17)": "NUMBER(27, 17)",
-    "DecimalType(27, 18)": "NUMBER(27, 18)",
-    "DecimalType(27, 19)": "NUMBER(27, 19)",
-    "DecimalType(27, 20)": "NUMBER(27, 20)",
-    "DecimalType(27, 21)": "NUMBER(27, 21)",
-    "DecimalType(27, 22)": "NUMBER(27, 22)",
-    "DecimalType(27, 23)": "NUMBER(27, 23)",
-    "DecimalType(27, 24)": "NUMBER(27, 24)",
-    "DecimalType(27, 25)": "NUMBER(27, 25)",
-    "DecimalType(27, 26)": "NUMBER(27, 26)",
-    "DecimalType(27, 27)": "NUMBER(27, 27)",
-    "DecimalType(27, 28)": "NUMBER(27, 28)",
-    "DecimalType(27, 29)": "NUMBER(27, 29)",
-    "DecimalType(27, 30)": "NUMBER(27, 30)",
-    "DecimalType(27, 31)": "NUMBER(27, 31)",
-    "DecimalType(27, 32)": "NUMBER(27, 32)",
-    "DecimalType(27, 33)": "NUMBER(27, 33)",
-    "DecimalType(27, 34)": "NUMBER(27, 34)",
-    "DecimalType(27, 35)": "NUMBER(27, 35)",
-    "DecimalType(27, 36)": "NUMBER(27, 36)",
-    "DecimalType(27, 37)": "NUMBER(27, 37)",
-    "DecimalType(27, 38)": "NUMBER(27, 38)",
-    "DecimalType(28, 0)": "NUMBER(28, 0)",
-    "DecimalType(28, 1)": "NUMBER(28, 1)",
-    "DecimalType(28, 2)": "NUMBER(28, 2)",
-    "DecimalType(28, 3)": "NUMBER(28, 3)",
-    "DecimalType(28, 4)": "NUMBER(28, 4)",
-    "DecimalType(28, 5)": "NUMBER(28, 5)",
-    "DecimalType(28, 6)": "NUMBER(28, 6)",
-    "DecimalType(28, 7)": "NUMBER(28, 7)",
-    "DecimalType(28, 8)": "NUMBER(28, 8)",
-    "DecimalType(28, 9)": "NUMBER(28, 9)",
-    "DecimalType(28, 10)": "NUMBER(28, 10)",
-    "DecimalType(28, 11)": "NUMBER(28, 11)",
-    "DecimalType(28, 12)": "NUMBER(28, 12)",
-    "DecimalType(28, 13)": "NUMBER(28, 13)",
-    "DecimalType(28, 14)": "NUMBER(28, 14)",
-    "DecimalType(28, 15)": "NUMBER(28, 15)",
-    "DecimalType(28, 16)": "NUMBER(28, 16)",
-    "DecimalType(28, 17)": "NUMBER(28, 17)",
-    "DecimalType(28, 18)": "NUMBER(28, 18)",
-    "DecimalType(28, 19)": "NUMBER(28, 19)",
-    "DecimalType(28, 20)": "NUMBER(28, 20)",
-    "DecimalType(28, 21)": "NUMBER(28, 21)",
-    "DecimalType(28, 22)": "NUMBER(28, 22)",
-    "DecimalType(28, 23)": "NUMBER(28, 23)",
-    "DecimalType(28, 24)": "NUMBER(28, 24)",
-    "DecimalType(28, 25)": "NUMBER(28, 25)",
-    "DecimalType(28, 26)": "NUMBER(28, 26)",
-    "DecimalType(28, 27)": "NUMBER(28, 27)",
-    "DecimalType(28, 28)": "NUMBER(28, 28)",
-    "DecimalType(28, 29)": "NUMBER(28, 29)",
-    "DecimalType(28, 30)": "NUMBER(28, 30)",
-    "DecimalType(28, 31)": "NUMBER(28, 31)",
-    "DecimalType(28, 32)": "NUMBER(28, 32)",
-    "DecimalType(28, 33)": "NUMBER(28, 33)",
-    "DecimalType(28, 34)": "NUMBER(28, 34)",
-    "DecimalType(28, 35)": "NUMBER(28, 35)",
-    "DecimalType(28, 36)": "NUMBER(28, 36)",
-    "DecimalType(28, 37)": "NUMBER(28, 37)",
-    "DecimalType(28, 38)": "NUMBER(28, 38)",
-    "DecimalType(29, 0)": "NUMBER(29, 0)",
-    "DecimalType(29, 1)": "NUMBER(29, 1)",
-    "DecimalType(29, 2)": "NUMBER(29, 2)",
-    "DecimalType(29, 3)": "NUMBER(29, 3)",
-    "DecimalType(29, 4)": "NUMBER(29, 4)",
-    "DecimalType(29, 5)": "NUMBER(29, 5)",
-    "DecimalType(29, 6)": "NUMBER(29, 6)",
-    "DecimalType(29, 7)": "NUMBER(29, 7)",
-    "DecimalType(29, 8)": "NUMBER(29, 8)",
-    "DecimalType(29, 9)": "NUMBER(29, 9)",
-    "DecimalType(29, 10)": "NUMBER(29, 10)",
-    "DecimalType(29, 11)": "NUMBER(29, 11)",
-    "DecimalType(29, 12)": "NUMBER(29, 12)",
-    "DecimalType(29, 13)": "NUMBER(29, 13)",
-    "DecimalType(29, 14)": "NUMBER(29, 14)",
-    "DecimalType(29, 15)": "NUMBER(29, 15)",
-    "DecimalType(29, 16)": "NUMBER(29, 16)",
-    "DecimalType(29, 17)": "NUMBER(29, 17)",
-    "DecimalType(29, 18)": "NUMBER(29, 18)",
-    "DecimalType(29, 19)": "NUMBER(29, 19)",
-    "DecimalType(29, 20)": "NUMBER(29, 20)",
-    "DecimalType(29, 21)": "NUMBER(29, 21)",
-    "DecimalType(29, 22)": "NUMBER(29, 22)",
-    "DecimalType(29, 23)": "NUMBER(29, 23)",
-    "DecimalType(29, 24)": "NUMBER(29, 24)",
-    "DecimalType(29, 25)": "NUMBER(29, 25)",
-    "DecimalType(29, 26)": "NUMBER(29, 26)",
-    "DecimalType(29, 27)": "NUMBER(29, 27)",
-    "DecimalType(29, 28)": "NUMBER(29, 28)",
-    "DecimalType(29, 29)": "NUMBER(29, 29)",
-    "DecimalType(29, 30)": "NUMBER(29, 30)",
-    "DecimalType(29, 31)": "NUMBER(29, 31)",
-    "DecimalType(29, 32)": "NUMBER(29, 32)",
-    "DecimalType(29, 33)": "NUMBER(29, 33)",
-    "DecimalType(29, 34)": "NUMBER(29, 34)",
-    "DecimalType(29, 35)": "NUMBER(29, 35)",
-    "DecimalType(29, 36)": "NUMBER(29, 36)",
-    "DecimalType(29, 37)": "NUMBER(29, 37)",
-    "DecimalType(29, 38)": "NUMBER(29, 38)",
-    "DecimalType(30, 0)": "NUMBER(30, 0)",
-    "DecimalType(30, 1)": "NUMBER(30, 1)",
-    "DecimalType(30, 2)": "NUMBER(30, 2)",
-    "DecimalType(30, 3)": "NUMBER(30, 3)",
-    "DecimalType(30, 4)": "NUMBER(30, 4)",
-    "DecimalType(30, 5)": "NUMBER(30, 5)",
-    "DecimalType(30, 6)": "NUMBER(30, 6)",
-    "DecimalType(30, 7)": "NUMBER(30, 7)",
-    "DecimalType(30, 8)": "NUMBER(30, 8)",
-    "DecimalType(30, 9)": "NUMBER(30, 9)",
-    "DecimalType(30, 10)": "NUMBER(30, 10)",
-    "DecimalType(30, 11)": "NUMBER(30, 11)",
-    "DecimalType(30, 12)": "NUMBER(30, 12)",
-    "DecimalType(30, 13)": "NUMBER(30, 13)",
-    "DecimalType(30, 14)": "NUMBER(30, 14)",
-    "DecimalType(30, 15)": "NUMBER(30, 15)",
-    "DecimalType(30, 16)": "NUMBER(30, 16)",
-    "DecimalType(30, 17)": "NUMBER(30, 17)",
-    "DecimalType(30, 18)": "NUMBER(30, 18)",
-    "DecimalType(30, 19)": "NUMBER(30, 19)",
-    "DecimalType(30, 20)": "NUMBER(30, 20)",
-    "DecimalType(30, 21)": "NUMBER(30, 21)",
-    "DecimalType(30, 22)": "NUMBER(30, 22)",
-    "DecimalType(30, 23)": "NUMBER(30, 23)",
-    "DecimalType(30, 24)": "NUMBER(30, 24)",
-    "DecimalType(30, 25)": "NUMBER(30, 25)",
-    "DecimalType(30, 26)": "NUMBER(30, 26)",
-    "DecimalType(30, 27)": "NUMBER(30, 27)",
-    "DecimalType(30, 28)": "NUMBER(30, 28)",
-    "DecimalType(30, 29)": "NUMBER(30, 29)",
-    "DecimalType(30, 30)": "NUMBER(30, 30)",
-    "DecimalType(30, 31)": "NUMBER(30, 31)",
-    "DecimalType(30, 32)": "NUMBER(30, 32)",
-    "DecimalType(30, 33)": "NUMBER(30, 33)",
-    "DecimalType(30, 34)": "NUMBER(30, 34)",
-    "DecimalType(30, 35)": "NUMBER(30, 35)",
-    "DecimalType(30, 36)": "NUMBER(30, 36)",
-    "DecimalType(30, 37)": "NUMBER(30, 37)",
-    "DecimalType(30, 38)": "NUMBER(30, 38)",
-    "DecimalType(31, 0)": "NUMBER(31, 0)",
-    "DecimalType(31, 1)": "NUMBER(31, 1)",
-    "DecimalType(31, 2)": "NUMBER(31, 2)",
-    "DecimalType(31, 3)": "NUMBER(31, 3)",
-    "DecimalType(31, 4)": "NUMBER(31, 4)",
-    "DecimalType(31, 5)": "NUMBER(31, 5)",
-    "DecimalType(31, 6)": "NUMBER(31, 6)",
-    "DecimalType(31, 7)": "NUMBER(31, 7)",
-    "DecimalType(31, 8)": "NUMBER(31, 8)",
-    "DecimalType(31, 9)": "NUMBER(31, 9)",
-    "DecimalType(31, 10)": "NUMBER(31, 10)",
-    "DecimalType(31, 11)": "NUMBER(31, 11)",
-    "DecimalType(31, 12)": "NUMBER(31, 12)",
-    "DecimalType(31, 13)": "NUMBER(31, 13)",
-    "DecimalType(31, 14)": "NUMBER(31, 14)",
-    "DecimalType(31, 15)": "NUMBER(31, 15)",
-    "DecimalType(31, 16)": "NUMBER(31, 16)",
-    "DecimalType(31, 17)": "NUMBER(31, 17)",
-    "DecimalType(31, 18)": "NUMBER(31, 18)",
-    "DecimalType(31, 19)": "NUMBER(31, 19)",
-    "DecimalType(31, 20)": "NUMBER(31, 20)",
-    "DecimalType(31, 21)": "NUMBER(31, 21)",
-    "DecimalType(31, 22)": "NUMBER(31, 22)",
-    "DecimalType(31, 23)": "NUMBER(31, 23)",
-    "DecimalType(31, 24)": "NUMBER(31, 24)",
-    "DecimalType(31, 25)": "NUMBER(31, 25)",
-    "DecimalType(31, 26)": "NUMBER(31, 26)",
-    "DecimalType(31, 27)": "NUMBER(31, 27)",
-    "DecimalType(31, 28)": "NUMBER(31, 28)",
-    "DecimalType(31, 29)": "NUMBER(31, 29)",
-    "DecimalType(31, 30)": "NUMBER(31, 30)",
-    "DecimalType(31, 31)": "NUMBER(31, 31)",
-    "DecimalType(31, 32)": "NUMBER(31, 32)",
-    "DecimalType(31, 33)": "NUMBER(31, 33)",
-    "DecimalType(31, 34)": "NUMBER(31, 34)",
-    "DecimalType(31, 35)": "NUMBER(31, 35)",
-    "DecimalType(31, 36)": "NUMBER(31, 36)",
-    "DecimalType(31, 37)": "NUMBER(31, 37)",
-    "DecimalType(31, 38)": "NUMBER(31, 38)",
-    "DecimalType(32, 0)": "NUMBER(32, 0)",
-    "DecimalType(32, 1)": "NUMBER(32, 1)",
-    "DecimalType(32, 2)": "NUMBER(32, 2)",
-    "DecimalType(32, 3)": "NUMBER(32, 3)",
-    "DecimalType(32, 4)": "NUMBER(32, 4)",
-    "DecimalType(32, 5)": "NUMBER(32, 5)",
-    "DecimalType(32, 6)": "NUMBER(32, 6)",
-    "DecimalType(32, 7)": "NUMBER(32, 7)",
-    "DecimalType(32, 8)": "NUMBER(32, 8)",
-    "DecimalType(32, 9)": "NUMBER(32, 9)",
-    "DecimalType(32, 10)": "NUMBER(32, 10)",
-    "DecimalType(32, 11)": "NUMBER(32, 11)",
-    "DecimalType(32, 12)": "NUMBER(32, 12)",
-    "DecimalType(32, 13)": "NUMBER(32, 13)",
-    "DecimalType(32, 14)": "NUMBER(32, 14)",
-    "DecimalType(32, 15)": "NUMBER(32, 15)",
-    "DecimalType(32, 16)": "NUMBER(32, 16)",
-    "DecimalType(32, 17)": "NUMBER(32, 17)",
-    "DecimalType(32, 18)": "NUMBER(32, 18)",
-    "DecimalType(32, 19)": "NUMBER(32, 19)",
-    "DecimalType(32, 20)": "NUMBER(32, 20)",
-    "DecimalType(32, 21)": "NUMBER(32, 21)",
-    "DecimalType(32, 22)": "NUMBER(32, 22)",
-    "DecimalType(32, 23)": "NUMBER(32, 23)",
-    "DecimalType(32, 24)": "NUMBER(32, 24)",
-    "DecimalType(32, 25)": "NUMBER(32, 25)",
-    "DecimalType(32, 26)": "NUMBER(32, 26)",
-    "DecimalType(32, 27)": "NUMBER(32, 27)",
-    "DecimalType(32, 28)": "NUMBER(32, 28)",
-    "DecimalType(32, 29)": "NUMBER(32, 29)",
-    "DecimalType(32, 30)": "NUMBER(32, 30)",
-    "DecimalType(32, 31)": "NUMBER(32, 31)",
-    "DecimalType(32, 32)": "NUMBER(32, 32)",
-    "DecimalType(32, 33)": "NUMBER(32, 33)",
-    "DecimalType(32, 34)": "NUMBER(32, 34)",
-    "DecimalType(32, 35)": "NUMBER(32, 35)",
-    "DecimalType(32, 36)": "NUMBER(32, 36)",
-    "DecimalType(32, 37)": "NUMBER(32, 37)",
-    "DecimalType(32, 38)": "NUMBER(32, 38)",
-    "DecimalType(33, 0)": "NUMBER(33, 0)",
-    "DecimalType(33, 1)": "NUMBER(33, 1)",
-    "DecimalType(33, 2)": "NUMBER(33, 2)",
-    "DecimalType(33, 3)": "NUMBER(33, 3)",
-    "DecimalType(33, 4)": "NUMBER(33, 4)",
-    "DecimalType(33, 5)": "NUMBER(33, 5)",
-    "DecimalType(33, 6)": "NUMBER(33, 6)",
-    "DecimalType(33, 7)": "NUMBER(33, 7)",
-    "DecimalType(33, 8)": "NUMBER(33, 8)",
-    "DecimalType(33, 9)": "NUMBER(33, 9)",
-    "DecimalType(33, 10)": "NUMBER(33, 10)",
-    "DecimalType(33, 11)": "NUMBER(33, 11)",
-    "DecimalType(33, 12)": "NUMBER(33, 12)",
-    "DecimalType(33, 13)": "NUMBER(33, 13)",
-    "DecimalType(33, 14)": "NUMBER(33, 14)",
-    "DecimalType(33, 15)": "NUMBER(33, 15)",
-    "DecimalType(33, 16)": "NUMBER(33, 16)",
-    "DecimalType(33, 17)": "NUMBER(33, 17)",
-    "DecimalType(33, 18)": "NUMBER(33, 18)",
-    "DecimalType(33, 19)": "NUMBER(33, 19)",
-    "DecimalType(33, 20)": "NUMBER(33, 20)",
-    "DecimalType(33, 21)": "NUMBER(33, 21)",
-    "DecimalType(33, 22)": "NUMBER(33, 22)",
-    "DecimalType(33, 23)": "NUMBER(33, 23)",
-    "DecimalType(33, 24)": "NUMBER(33, 24)",
-    "DecimalType(33, 25)": "NUMBER(33, 25)",
-    "DecimalType(33, 26)": "NUMBER(33, 26)",
-    "DecimalType(33, 27)": "NUMBER(33, 27)",
-    "DecimalType(33, 28)": "NUMBER(33, 28)",
-    "DecimalType(33, 29)": "NUMBER(33, 29)",
-    "DecimalType(33, 30)": "NUMBER(33, 30)",
-    "DecimalType(33, 31)": "NUMBER(33, 31)",
-    "DecimalType(33, 32)": "NUMBER(33, 32)",
-    "DecimalType(33, 33)": "NUMBER(33, 33)",
-    "DecimalType(33, 34)": "NUMBER(33, 34)",
-    "DecimalType(33, 35)": "NUMBER(33, 35)",
-    "DecimalType(33, 36)": "NUMBER(33, 36)",
-    "DecimalType(33, 37)": "NUMBER(33, 37)",
-    "DecimalType(33, 38)": "NUMBER(33, 38)",
-    "DecimalType(34, 0)": "NUMBER(34, 0)",
-    "DecimalType(34, 1)": "NUMBER(34, 1)",
-    "DecimalType(34, 2)": "NUMBER(34, 2)",
-    "DecimalType(34, 3)": "NUMBER(34, 3)",
-    "DecimalType(34, 4)": "NUMBER(34, 4)",
-    "DecimalType(34, 5)": "NUMBER(34, 5)",
-    "DecimalType(34, 6)": "NUMBER(34, 6)",
-    "DecimalType(34, 7)": "NUMBER(34, 7)",
-    "DecimalType(34, 8)": "NUMBER(34, 8)",
-    "DecimalType(34, 9)": "NUMBER(34, 9)",
-    "DecimalType(34, 10)": "NUMBER(34, 10)",
-    "DecimalType(34, 11)": "NUMBER(34, 11)",
-    "DecimalType(34, 12)": "NUMBER(34, 12)",
-    "DecimalType(34, 13)": "NUMBER(34, 13)",
-    "DecimalType(34, 14)": "NUMBER(34, 14)",
-    "DecimalType(34, 15)": "NUMBER(34, 15)",
-    "DecimalType(34, 16)": "NUMBER(34, 16)",
-    "DecimalType(34, 17)": "NUMBER(34, 17)",
-    "DecimalType(34, 18)": "NUMBER(34, 18)",
-    "DecimalType(34, 19)": "NUMBER(34, 19)",
-    "DecimalType(34, 20)": "NUMBER(34, 20)",
-    "DecimalType(34, 21)": "NUMBER(34, 21)",
-    "DecimalType(34, 22)": "NUMBER(34, 22)",
-    "DecimalType(34, 23)": "NUMBER(34, 23)",
-    "DecimalType(34, 24)": "NUMBER(34, 24)",
-    "DecimalType(34, 25)": "NUMBER(34, 25)",
-    "DecimalType(34, 26)": "NUMBER(34, 26)",
-    "DecimalType(34, 27)": "NUMBER(34, 27)",
-    "DecimalType(34, 28)": "NUMBER(34, 28)",
-    "DecimalType(34, 29)": "NUMBER(34, 29)",
-    "DecimalType(34, 30)": "NUMBER(34, 30)",
-    "DecimalType(34, 31)": "NUMBER(34, 31)",
-    "DecimalType(34, 32)": "NUMBER(34, 32)",
-    "DecimalType(34, 33)": "NUMBER(34, 33)",
-    "DecimalType(34, 34)": "NUMBER(34, 34)",
-    "DecimalType(34, 35)": "NUMBER(34, 35)",
-    "DecimalType(34, 36)": "NUMBER(34, 36)",
-    "DecimalType(34, 37)": "NUMBER(34, 37)",
-    "DecimalType(34, 38)": "NUMBER(34, 38)",
-    "DecimalType(35, 0)": "NUMBER(35, 0)",
-    "DecimalType(35, 1)": "NUMBER(35, 1)",
-    "DecimalType(35, 2)": "NUMBER(35, 2)",
-    "DecimalType(35, 3)": "NUMBER(35, 3)",
-    "DecimalType(35, 4)": "NUMBER(35, 4)",
-    "DecimalType(35, 5)": "NUMBER(35, 5)",
-    "DecimalType(35, 6)": "NUMBER(35, 6)",
-    "DecimalType(35, 7)": "NUMBER(35, 7)",
-    "DecimalType(35, 8)": "NUMBER(35, 8)",
-    "DecimalType(35, 9)": "NUMBER(35, 9)",
-    "DecimalType(35, 10)": "NUMBER(35, 10)",
-    "DecimalType(35, 11)": "NUMBER(35, 11)",
-    "DecimalType(35, 12)": "NUMBER(35, 12)",
-    "DecimalType(35, 13)": "NUMBER(35, 13)",
-    "DecimalType(35, 14)": "NUMBER(35, 14)",
-    "DecimalType(35, 15)": "NUMBER(35, 15)",
-    "DecimalType(35, 16)": "NUMBER(35, 16)",
-    "DecimalType(35, 17)": "NUMBER(35, 17)",
-    "DecimalType(35, 18)": "NUMBER(35, 18)",
-    "DecimalType(35, 19)": "NUMBER(35, 19)",
-    "DecimalType(35, 20)": "NUMBER(35, 20)",
-    "DecimalType(35, 21)": "NUMBER(35, 21)",
-    "DecimalType(35, 22)": "NUMBER(35, 22)",
-    "DecimalType(35, 23)": "NUMBER(35, 23)",
-    "DecimalType(35, 24)": "NUMBER(35, 24)",
-    "DecimalType(35, 25)": "NUMBER(35, 25)",
-    "DecimalType(35, 26)": "NUMBER(35, 26)",
-    "DecimalType(35, 27)": "NUMBER(35, 27)",
-    "DecimalType(35, 28)": "NUMBER(35, 28)",
-    "DecimalType(35, 29)": "NUMBER(35, 29)",
-    "DecimalType(35, 30)": "NUMBER(35, 30)",
-    "DecimalType(35, 31)": "NUMBER(35, 31)",
-    "DecimalType(35, 32)": "NUMBER(35, 32)",
-    "DecimalType(35, 33)": "NUMBER(35, 33)",
-    "DecimalType(35, 34)": "NUMBER(35, 34)",
-    "DecimalType(35, 35)": "NUMBER(35, 35)",
-    "DecimalType(35, 36)": "NUMBER(35, 36)",
-    "DecimalType(35, 37)": "NUMBER(35, 37)",
-    "DecimalType(35, 38)": "NUMBER(35, 38)",
-    "DecimalType(36, 0)": "NUMBER(36, 0)",
-    "DecimalType(36, 1)": "NUMBER(36, 1)",
-    "DecimalType(36, 2)": "NUMBER(36, 2)",
-    "DecimalType(36, 3)": "NUMBER(36, 3)",
-    "DecimalType(36, 4)": "NUMBER(36, 4)",
-    "DecimalType(36, 5)": "NUMBER(36, 5)",
-    "DecimalType(36, 6)": "NUMBER(36, 6)",
-    "DecimalType(36, 7)": "NUMBER(36, 7)",
-    "DecimalType(36, 8)": "NUMBER(36, 8)",
-    "DecimalType(36, 9)": "NUMBER(36, 9)",
-    "DecimalType(36, 10)": "NUMBER(36, 10)",
-    "DecimalType(36, 11)": "NUMBER(36, 11)",
-    "DecimalType(36, 12)": "NUMBER(36, 12)",
-    "DecimalType(36, 13)": "NUMBER(36, 13)",
-    "DecimalType(36, 14)": "NUMBER(36, 14)",
-    "DecimalType(36, 15)": "NUMBER(36, 15)",
-    "DecimalType(36, 16)": "NUMBER(36, 16)",
-    "DecimalType(36, 17)": "NUMBER(36, 17)",
-    "DecimalType(36, 18)": "NUMBER(36, 18)",
-    "DecimalType(36, 19)": "NUMBER(36, 19)",
-    "DecimalType(36, 20)": "NUMBER(36, 20)",
-    "DecimalType(36, 21)": "NUMBER(36, 21)",
-    "DecimalType(36, 22)": "NUMBER(36, 22)",
-    "DecimalType(36, 23)": "NUMBER(36, 23)",
-    "DecimalType(36, 24)": "NUMBER(36, 24)",
-    "DecimalType(36, 25)": "NUMBER(36, 25)",
-    "DecimalType(36, 26)": "NUMBER(36, 26)",
-    "DecimalType(36, 27)": "NUMBER(36, 27)",
-    "DecimalType(36, 28)": "NUMBER(36, 28)",
-    "DecimalType(36, 29)": "NUMBER(36, 29)",
-    "DecimalType(36, 30)": "NUMBER(36, 30)",
-    "DecimalType(36, 31)": "NUMBER(36, 31)",
-    "DecimalType(36, 32)": "NUMBER(36, 32)",
-    "DecimalType(36, 33)": "NUMBER(36, 33)",
-    "DecimalType(36, 34)": "NUMBER(36, 34)",
-    "DecimalType(36, 35)": "NUMBER(36, 35)",
-    "DecimalType(36, 36)": "NUMBER(36, 36)",
-    "DecimalType(36, 37)": "NUMBER(36, 37)",
-    "DecimalType(36, 38)": "NUMBER(36, 38)",
-    "DecimalType(37, 0)": "NUMBER(37, 0)",
-    "DecimalType(37, 1)": "NUMBER(37, 1)",
-    "DecimalType(37, 2)": "NUMBER(37, 2)",
-    "DecimalType(37, 3)": "NUMBER(37, 3)",
-    "DecimalType(37, 4)": "NUMBER(37, 4)",
-    "DecimalType(37, 5)": "NUMBER(37, 5)",
-    "DecimalType(37, 6)": "NUMBER(37, 6)",
-    "DecimalType(37, 7)": "NUMBER(37, 7)",
-    "DecimalType(37, 8)": "NUMBER(37, 8)",
-    "DecimalType(37, 9)": "NUMBER(37, 9)",
-    "DecimalType(37, 10)": "NUMBER(37, 10)",
-    "DecimalType(37, 11)": "NUMBER(37, 11)",
-    "DecimalType(37, 12)": "NUMBER(37, 12)",
-    "DecimalType(37, 13)": "NUMBER(37, 13)",
-    "DecimalType(37, 14)": "NUMBER(37, 14)",
-    "DecimalType(37, 15)": "NUMBER(37, 15)",
-    "DecimalType(37, 16)": "NUMBER(37, 16)",
-    "DecimalType(37, 17)": "NUMBER(37, 17)",
-    "DecimalType(37, 18)": "NUMBER(37, 18)",
-    "DecimalType(37, 19)": "NUMBER(37, 19)",
-    "DecimalType(37, 20)": "NUMBER(37, 20)",
-    "DecimalType(37, 21)": "NUMBER(37, 21)",
-    "DecimalType(37, 22)": "NUMBER(37, 22)",
-    "DecimalType(37, 23)": "NUMBER(37, 23)",
-    "DecimalType(37, 24)": "NUMBER(37, 24)",
-    "DecimalType(37, 25)": "NUMBER(37, 25)",
-    "DecimalType(37, 26)": "NUMBER(37, 26)",
-    "DecimalType(37, 27)": "NUMBER(37, 27)",
-    "DecimalType(37, 28)": "NUMBER(37, 28)",
-    "DecimalType(37, 29)": "NUMBER(37, 29)",
-    "DecimalType(37, 30)": "NUMBER(37, 30)",
-    "DecimalType(37, 31)": "NUMBER(37, 31)",
-    "DecimalType(37, 32)": "NUMBER(37, 32)",
-    "DecimalType(37, 33)": "NUMBER(37, 33)",
-    "DecimalType(37, 34)": "NUMBER(37, 34)",
-    "DecimalType(37, 35)": "NUMBER(37, 35)",
-    "DecimalType(37, 36)": "NUMBER(37, 36)",
-    "DecimalType(37, 37)": "NUMBER(37, 37)",
-    "DecimalType(37, 38)": "NUMBER(37, 38)",
-    "DecimalType(38, 0)": "NUMBER(38, 0)",
-    "DecimalType(38, 1)": "NUMBER(38, 1)",
-    "DecimalType(38, 2)": "NUMBER(38, 2)",
-    "DecimalType(38, 3)": "NUMBER(38, 3)",
-    "DecimalType(38, 4)": "NUMBER(38, 4)",
-    "DecimalType(38, 5)": "NUMBER(38, 5)",
-    "DecimalType(38, 6)": "NUMBER(38, 6)",
-    "DecimalType(38, 7)": "NUMBER(38, 7)",
-    "DecimalType(38, 8)": "NUMBER(38, 8)",
-    "DecimalType(38, 9)": "NUMBER(38, 9)",
-    "DecimalType(38, 10)": "NUMBER(38, 10)",
-    "DecimalType(38, 11)": "NUMBER(38, 11)",
-    "DecimalType(38, 12)": "NUMBER(38, 12)",
-    "DecimalType(38, 13)": "NUMBER(38, 13)",
-    "DecimalType(38, 14)": "NUMBER(38, 14)",
-    "DecimalType(38, 15)": "NUMBER(38, 15)",
-    "DecimalType(38, 16)": "NUMBER(38, 16)",
-    "DecimalType(38, 17)": "NUMBER(38, 17)",
-    "DecimalType(38, 18)": "NUMBER(38, 18)",
-    "DecimalType(38, 19)": "NUMBER(38, 19)",
-    "DecimalType(38, 20)": "NUMBER(38, 20)",
-    "DecimalType(38, 21)": "NUMBER(38, 21)",
-    "DecimalType(38, 22)": "NUMBER(38, 22)",
-    "DecimalType(38, 23)": "NUMBER(38, 23)",
-    "DecimalType(38, 24)": "NUMBER(38, 24)",
-    "DecimalType(38, 25)": "NUMBER(38, 25)",
-    "DecimalType(38, 26)": "NUMBER(38, 26)",
-    "DecimalType(38, 27)": "NUMBER(38, 27)",
-    "DecimalType(38, 28)": "NUMBER(38, 28)",
-    "DecimalType(38, 29)": "NUMBER(38, 29)",
-    "DecimalType(38, 30)": "NUMBER(38, 30)",
-    "DecimalType(38, 31)": "NUMBER(38, 31)",
-    "DecimalType(38, 32)": "NUMBER(38, 32)",
-    "DecimalType(38, 33)": "NUMBER(38, 33)",
-    "DecimalType(38, 34)": "NUMBER(38, 34)",
-    "DecimalType(38, 35)": "NUMBER(38, 35)",
-    "DecimalType(38, 36)": "NUMBER(38, 36)",
-    "DecimalType(38, 37)": "NUMBER(38, 37)",
-    "DecimalType(38, 38)": "NUMBER(38, 38)",
-    "FloatType()": "NUMBER(P=19, S=4)",
-    "DoubleType()": "NUMBER(P=19, S=4)",
-    "StringType()": "CLOB",
-    "TimestampType()": "TIMESTAMP(6)",
-    "BinaryType()": "BLOB",
-    "BooleanType()": "BOOLEAN",
-}
