+++ tests/input_ddl/mssql_mysql/simple_create.txt
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS test.Persons (
+    PersonID int,
+    LastName varchar,
+    Money DECIMAL(5),
+    time TIME,
+    xml XML,
+);
