+++ tests/output_ddl/mssql_mysql/simple_create.txt
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS test.Persons (
+    PersonID INT, 
+    LastName LONGTEXT, 
+    Money DECIMAL(5, 0), 
+    time TIMESTAMP(6), 
+    xml LONGTEXT, 
+);
