+++ tests/output_ddl/mssql_oracle/simple_create.txt
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS test.Persons (
+    PersonID NUMBER(19, 4), 
+    LastName CLOB, 
+    Money NUMBER(5, 0), 
+    time TIMESTAMP(6), 
+    xml CLOB, 
+);
