+++ tests/test_oracle_mssql.py
@@ -3,10 +3,10 @@ import pytest
 from ddl_spark_converter.main_converter import DatabaseConverter
 
 SOURCE_DB = "oracle"
-TARGET_DB = "mysql"
+TARGET_DB = "mssql"
 
-test_ddls = open("tests/input_ddl/oracle_mysql/simple_create.txt", "r")
-expected_ddls = open("tests/output_ddl/oracle_mysql/simple_create.txt", "r")
+test_ddls = open("tests/input_ddl/oracle_mssql/simple_create.txt", "r")
+expected_ddls = open("tests/output_ddl/oracle_mssql/simple_create.txt", "r")
 
 test_ddls = [test_ddl for test_ddl in test_ddls.read().split("\n\n")]
 expected_ddls = [test_ddl for test_ddl in expected_ddls.read().split("\n\n")]
